fix:  migrate to macos_ui 2.0.0 (fixes #96) part 1
