+++ packages/rapid_cli/lib/src/commands/activate.dart
@@ -159,24 +159,6 @@ mixin _ActivateMixin on _Rapid {
             () => _flutterConfigEnablePlatformTask(platform: platform),
           ],
         );
-
-        // TODO: Required due to https://github.com/jtdLab/rapid/issues/96
-        // Sets the `osx` version inside the `Podfile` of the `macos` app
-        // to `10.15.7.7`. This is required because rapid projects use a
-        // [macos_ui](https://pub.dev/packages/macos_ui) version from the dev channel.
-        final podFile = rootPackage.nativeDirectory.podFile;
-        if (podFile.existsSync()) {
-          replaceInFile(
-            podFile,
-            'platform :osx, \'10.14\'',
-            'platform :osx, \'10.15.7.7\'',
-          );
-        } else {
-          // Required because on ci (maybe other systems too) the Podfile is not generated by flutter tooling.
-          podFile
-            ..createSync(recursive: true)
-            ..writeAsStringSync(_podFileContent);
-        }
       },
     );
   }
@@ -391,49 +373,3 @@ class PlatformAlreadyActivatedException extends RapidException {
   PlatformAlreadyActivatedException._(Platform platform)
       : super('The platform ${platform.prettyName} is already activated.');
 }
-
-const _podFileContent = '''
-platform :osx, '10.15.7.7'
-
-# CocoaPods analytics sends network stats synchronously affecting flutter build latency.
-ENV['COCOAPODS_DISABLE_STATS'] = 'true'
-
-project 'Runner', {
-  'Debug' => :debug,
-  'Profile' => :release,
-  'Release' => :release,
-}
-
-def flutter_root
-  generated_xcode_build_settings_path = File.expand_path(File.join('..', 'Flutter', 'ephemeral', 'Flutter-Generated.xcconfig'), __FILE__)
-  unless File.exist?(generated_xcode_build_settings_path)
-    raise "#{generated_xcode_build_settings_path} must exist. If you're running pod install manually, make sure \\"flutter pub get\\" is executed first"
-  end
-
-  File.foreach(generated_xcode_build_settings_path) do |line|
-    matches = line.match(/FLUTTER_ROOT\\=(.*)/)
-    return matches[1].strip if matches
-  end
-  raise "FLUTTER_ROOT not found in #{generated_xcode_build_settings_path}. Try deleting Flutter-Generated.xcconfig, then run \\"flutter pub get\\""
-end
-
-require File.expand_path(File.join('packages', 'flutter_tools', 'bin', 'podhelper'), flutter_root)
-
-flutter_macos_podfile_setup
-
-target 'Runner' do
-  use_frameworks!
-  use_modular_headers!
-
-  flutter_install_all_macos_pods File.dirname(File.realpath(__FILE__))
-  target 'RunnerTests' do
-    inherit! :search_paths
-  end
-end
-
-post_install do |installer|
-  installer.pods_project.targets.each do |target|
-    flutter_additional_macos_build_settings(target)
-  end
-end
-''';
