+++ packages/rapid_cli/lib/src/project/app_module/platform_native_directory.dart
@@ -80,8 +80,6 @@ class MacosNativeDirectory extends PlatformNativeDirectory {
       },
     );
   }
-
-  File get podFile => File(p.join(path, 'Podfile'));
 }
 
 // TODO better name
