+++ packages/rapid_cli/lib/src/project/bundles/macos_native_directory_bundle.dart
@@ -30,9 +30,15 @@ final macosNativeDirectoryBundle = MasonBundle.fromJson(<String, dynamic>{
       "type": "text"
     },
     {
+      "path": "Podfile",
+      "data":
+          "cGxhdGZvcm0gOm9zeCwgJzEwLjE0LjYnCgojIENvY29hUG9kcyBhbmFseXRpY3Mgc2VuZHMgbmV0d29yayBzdGF0cyBzeW5jaHJvbm91c2x5IGFmZmVjdGluZyBmbHV0dGVyIGJ1aWxkIGxhdGVuY3kuCkVOVlsnQ09DT0FQT0RTX0RJU0FCTEVfU1RBVFMnXSA9ICd0cnVlJwoKcHJvamVjdCAnUnVubmVyJywgewogICdEZWJ1ZycgPT4gOmRlYnVnLAogICdQcm9maWxlJyA9PiA6cmVsZWFzZSwKICAnUmVsZWFzZScgPT4gOnJlbGVhc2UsCn0KCmRlZiBmbHV0dGVyX3Jvb3QKICBnZW5lcmF0ZWRfeGNvZGVfYnVpbGRfc2V0dGluZ3NfcGF0aCA9IEZpbGUuZXhwYW5kX3BhdGgoRmlsZS5qb2luKCcuLicsICdGbHV0dGVyJywgJ2VwaGVtZXJhbCcsICdGbHV0dGVyLUdlbmVyYXRlZC54Y2NvbmZpZycpLCBfX0ZJTEVfXykKICB1bmxlc3MgRmlsZS5leGlzdD8oZ2VuZXJhdGVkX3hjb2RlX2J1aWxkX3NldHRpbmdzX3BhdGgpCiAgICByYWlzZSAiI3tnZW5lcmF0ZWRfeGNvZGVfYnVpbGRfc2V0dGluZ3NfcGF0aH0gbXVzdCBleGlzdC4gSWYgeW91J3JlIHJ1bm5pbmcgcG9kIGluc3RhbGwgbWFudWFsbHksIG1ha2Ugc3VyZSBcImZsdXR0ZXIgcHViIGdldFwiIGlzIGV4ZWN1dGVkIGZpcnN0IgogIGVuZAoKICBGaWxlLmZvcmVhY2goZ2VuZXJhdGVkX3hjb2RlX2J1aWxkX3NldHRpbmdzX3BhdGgpIGRvIHxsaW5lfAogICAgbWF0Y2hlcyA9IGxpbmUubWF0Y2goL0ZMVVRURVJfUk9PVFw9KC4qKS8pCiAgICByZXR1cm4gbWF0Y2hlc1sxXS5zdHJpcCBpZiBtYXRjaGVzCiAgZW5kCiAgcmFpc2UgIkZMVVRURVJfUk9PVCBub3QgZm91bmQgaW4gI3tnZW5lcmF0ZWRfeGNvZGVfYnVpbGRfc2V0dGluZ3NfcGF0aH0uIFRyeSBkZWxldGluZyBGbHV0dGVyLUdlbmVyYXRlZC54Y2NvbmZpZywgdGhlbiBydW4gXCJmbHV0dGVyIHB1YiBnZXRcIiIKZW5kCgpyZXF1aXJlIEZpbGUuZXhwYW5kX3BhdGgoRmlsZS5qb2luKCdwYWNrYWdlcycsICdmbHV0dGVyX3Rvb2xzJywgJ2JpbicsICdwb2RoZWxwZXInKSwgZmx1dHRlcl9yb290KQoKZmx1dHRlcl9tYWNvc19wb2RmaWxlX3NldHVwCgp0YXJnZXQgJ1J1bm5lcicgZG8KICB1c2VfZnJhbWV3b3JrcyEKICB1c2VfbW9kdWxhcl9oZWFkZXJzIQoKICBmbHV0dGVyX2luc3RhbGxfYWxsX21hY29zX3BvZHMgRmlsZS5kaXJuYW1lKEZpbGUucmVhbHBhdGgoX19GSUxFX18pKQogIHRhcmdldCAnUnVubmVyVGVzdHMnIGRvCiAgICBpbmhlcml0ISA6c2VhcmNoX3BhdGhzCiAgZW5kCmVuZAoKcG9zdF9pbnN0YWxsIGRvIHxpbnN0YWxsZXJ8CiAgaW5zdGFsbGVyLnBvZHNfcHJvamVjdC50YXJnZXRzLmVhY2ggZG8gfHRhcmdldHwKICAgIGZsdXR0ZXJfYWRkaXRpb25hbF9tYWNvc19idWlsZF9zZXR0aW5ncyh0YXJnZXQpCiAgZW5kCmVuZAo=",
+      "type": "text"
+    },
+    {
       "path": "Runner.xcodeproj/project.pbxproj",
       "data":
-          "Ly8gISQqVVRGOCokIQp7CglhcmNoaXZlVmVyc2lvbiA9IDE7CgljbGFzc2VzID0gewoJfTsKCW9iamVjdFZlcnNpb24gPSA1NDsKCW9iamVjdHMgPSB7CgovKiBCZWdpbiBQQlhBZ2dyZWdhdGVUYXJnZXQgc2VjdGlvbiAqLwoJCTMzQ0MxMTFBMjA0NEM2QkEwMDAzQzA0NSAvKiBGbHV0dGVyIEFzc2VtYmxlICovID0gewoJCQlpc2EgPSBQQlhBZ2dyZWdhdGVUYXJnZXQ7CgkJCWJ1aWxkQ29uZmlndXJhdGlvbkxpc3QgPSAzM0NDMTExQjIwNDRDNkJBMDAwM0MwNDUgLyogQnVpbGQgY29uZmlndXJhdGlvbiBsaXN0IGZvciBQQlhBZ2dyZWdhdGVUYXJnZXQgIkZsdXR0ZXIgQXNzZW1ibGUiICovOwoJCQlidWlsZFBoYXNlcyA9ICgKCQkJCTMzQ0MxMTFFMjA0NEM2QkYwMDAzQzA0NSAvKiBTaGVsbFNjcmlwdCAqLywKCQkJKTsKCQkJZGVwZW5kZW5jaWVzID0gKAoJCQkpOwoJCQluYW1lID0gIkZsdXR0ZXIgQXNzZW1ibGUiOwoJCQlwcm9kdWN0TmFtZSA9IEZMWDsKCQl9OwovKiBFbmQgUEJYQWdncmVnYXRlVGFyZ2V0IHNlY3Rpb24gKi8KCi8qIEJlZ2luIFBCWEJ1aWxkRmlsZSBzZWN0aW9uICovCgkJMzMxQzgwRDgyOTRDRjcxMDAwMjYzQkU1IC8qIFJ1bm5lclRlc3RzLnN3aWZ0IGluIFNvdXJjZXMgKi8gPSB7aXNhID0gUEJYQnVpbGRGaWxlOyBmaWxlUmVmID0gMzMxQzgwRDcyOTRDRjcxMDAwMjYzQkU1IC8qIFJ1bm5lclRlc3RzLnN3aWZ0ICovOyB9OwoJCTMzNUJCRDFCMjJBOUExNUUwMEU5MDcxRCAvKiBHZW5lcmF0ZWRQbHVnaW5SZWdpc3RyYW50LnN3aWZ0IGluIFNvdXJjZXMgKi8gPSB7aXNhID0gUEJYQnVpbGRGaWxlOyBmaWxlUmVmID0gMzM1QkJEMUEyMkE5QTE1RTAwRTkwNzFEIC8qIEdlbmVyYXRlZFBsdWdpblJlZ2lzdHJhbnQuc3dpZnQgKi87IH07CgkJMzNDQzEwRjEyMDQ0QTNDNjAwMDNDMDQ1IC8qIEFwcERlbGVnYXRlLnN3aWZ0IGluIFNvdXJjZXMgKi8gPSB7aXNhID0gUEJYQnVpbGRGaWxlOyBmaWxlUmVmID0gMzNDQzEwRjAyMDQ0QTNDNjAwMDNDMDQ1IC8qIEFwcERlbGVnYXRlLnN3aWZ0ICovOyB9OwoJCTMzQ0MxMEYzMjA0NEEzQzYwMDAzQzA0NSAvKiBBc3NldHMueGNhc3NldHMgaW4gUmVzb3VyY2VzICovID0ge2lzYSA9IFBCWEJ1aWxkRmlsZTsgZmlsZVJlZiA9IDMzQ0MxMEYyMjA0NEEzQzYwMDAzQzA0NSAvKiBBc3NldHMueGNhc3NldHMgKi87IH07CgkJMzNDQzEwRjYyMDQ0QTNDNjAwMDNDMDQ1IC8qIE1haW5NZW51LnhpYiBpbiBSZXNvdXJjZXMgKi8gPSB7aXNhID0gUEJYQnVpbGRGaWxlOyBmaWxlUmVmID0gMzNDQzEwRjQyMDQ0QTNDNjAwMDNDMDQ1IC8qIE1haW5NZW51LnhpYiAqLzsgfTsKCQkzM0NDMTExMzIwNDRCRkEwMDAwM0MwNDUgLyogTWFpbkZsdXR0ZXJXaW5kb3cuc3dpZnQgaW4gU291cmNlcyAqLyA9IHtpc2EgPSBQQlhCdWlsZEZpbGU7IGZpbGVSZWYgPSAzM0NDMTExMjIwNDRCRkEwMDAwM0MwNDUgLyogTWFpbkZsdXR0ZXJXaW5kb3cuc3dpZnQgKi87IH07Ci8qIEVuZCBQQlhCdWlsZEZpbGUgc2VjdGlvbiAqLwoKLyogQmVnaW4gUEJYQ29udGFpbmVySXRlbVByb3h5IHNlY3Rpb24gKi8KCQkzMzFDODBEOTI5NENGNzEwMDAyNjNCRTUgLyogUEJYQ29udGFpbmVySXRlbVByb3h5ICovID0gewoJCQlpc2EgPSBQQlhDb250YWluZXJJdGVtUHJveHk7CgkJCWNvbnRhaW5lclBvcnRhbCA9IDMzQ0MxMEU1MjA0NEEzQzYwMDAzQzA0NSAvKiBQcm9qZWN0IG9iamVjdCAqLzsKCQkJcHJveHlUeXBlID0gMTsKCQkJcmVtb3RlR2xvYmFsSURTdHJpbmcgPSAzM0NDMTBFQzIwNDRBM0M2MDAwM0MwNDU7CgkJCXJlbW90ZUluZm8gPSBSdW5uZXI7CgkJfTsKCQkzM0NDMTExRjIwNDRDNzlGMDAwM0MwNDUgLyogUEJYQ29udGFpbmVySXRlbVByb3h5ICovID0gewoJCQlpc2EgPSBQQlhDb250YWluZXJJdGVtUHJveHk7CgkJCWNvbnRhaW5lclBvcnRhbCA9IDMzQ0MxMEU1MjA0NEEzQzYwMDAzQzA0NSAvKiBQcm9qZWN0IG9iamVjdCAqLzsKCQkJcHJveHlUeXBlID0gMTsKCQkJcmVtb3RlR2xvYmFsSURTdHJpbmcgPSAzM0NDMTExQTIwNDRDNkJBMDAwM0MwNDU7CgkJCXJlbW90ZUluZm8gPSBGTFg7CgkJfTsKLyogRW5kIFBCWENvbnRhaW5lckl0ZW1Qcm94eSBzZWN0aW9uICovCgovKiBCZWdpbiBQQlhDb3B5RmlsZXNCdWlsZFBoYXNlIHNlY3Rpb24gKi8KCQkzM0NDMTEwRTIwNDRBODg0MDAwM0MwNDUgLyogQnVuZGxlIEZyYW1ld29yayAqLyA9IHsKCQkJaXNhID0gUEJYQ29weUZpbGVzQnVpbGRQaGFzZTsKCQkJYnVpbGRBY3Rpb25NYXNrID0gMjE0NzQ4MzY0NzsKCQkJZHN0UGF0aCA9ICIiOwoJCQlkc3RTdWJmb2xkZXJTcGVjID0gMTA7CgkJCWZpbGVzID0gKAoJCQkpOwoJCQluYW1lID0gIkJ1bmRsZSBGcmFtZXdvcmsiOwoJCQlydW5Pbmx5Rm9yRGVwbG95bWVudFBvc3Rwcm9jZXNzaW5nID0gMDsKCQl9OwovKiBFbmQgUEJYQ29weUZpbGVzQnVpbGRQaGFzZSBzZWN0aW9uICovCgovKiBCZWdpbiBQQlhGaWxlUmVmZXJlbmNlIHNlY3Rpb24gKi8KCQkzMzFDODBENTI5NENGNzEwMDAyNjNCRTUgLyogUnVubmVyVGVzdHMueGN0ZXN0ICovID0ge2lzYSA9IFBCWEZpbGVSZWZlcmVuY2U7IGV4cGxpY2l0RmlsZVR5cGUgPSB3cmFwcGVyLmNmYnVuZGxlOyBpbmNsdWRlSW5JbmRleCA9IDA7IHBhdGggPSBSdW5uZXJUZXN0cy54Y3Rlc3Q7IHNvdXJjZVRyZWUgPSBCVUlMVF9QUk9EVUNUU19ESVI7IH07CgkJMzMxQzgwRDcyOTRDRjcxMDAwMjYzQkU1IC8qIFJ1bm5lclRlc3RzLnN3aWZ0ICovID0ge2lzYSA9IFBCWEZpbGVSZWZlcmVuY2U7IGxhc3RLbm93bkZpbGVUeXBlID0gc291cmNlY29kZS5zd2lmdDsgcGF0aCA9IFJ1bm5lclRlc3RzLnN3aWZ0OyBzb3VyY2VUcmVlID0gIjxncm91cD4iOyB9OwoJCTMzMzAwMEVEMjJEM0RFNUQwMDU1NDE2MiAvKiBXYXJuaW5ncy54Y2NvbmZpZyAqLyA9IHtpc2EgPSBQQlhGaWxlUmVmZXJlbmNlOyBsYXN0S25vd25GaWxlVHlwZSA9IHRleHQueGNjb25maWc7IHBhdGggPSBXYXJuaW5ncy54Y2NvbmZpZzsgc291cmNlVHJlZSA9ICI8Z3JvdXA+IjsgfTsKCQkzMzVCQkQxQTIyQTlBMTVFMDBFOTA3MUQgLyogR2VuZXJhdGVkUGx1Z2luUmVnaXN0cmFudC5zd2lmdCAqLyA9IHtpc2EgPSBQQlhGaWxlUmVmZXJlbmNlOyBmaWxlRW5jb2RpbmcgPSA0OyBsYXN0S25vd25GaWxlVHlwZSA9IHNvdXJjZWNvZGUuc3dpZnQ7IHBhdGggPSBHZW5lcmF0ZWRQbHVnaW5SZWdpc3RyYW50LnN3aWZ0OyBzb3VyY2VUcmVlID0gIjxncm91cD4iOyB9OwoJCTMzQ0MxMEVEMjA0NEEzQzYwMDAzQzA0NSAvKiB7e3Byb2plY3RfbmFtZX19LmFwcCAqLyA9IHtpc2EgPSBQQlhGaWxlUmVmZXJlbmNlOyBleHBsaWNpdEZpbGVUeXBlID0gd3JhcHBlci5hcHBsaWNhdGlvbjsgaW5jbHVkZUluSW5kZXggPSAwOyBwYXRoID0gInt7cHJvamVjdF9uYW1lfX0uYXBwIjsgc291cmNlVHJlZSA9IEJVSUxUX1BST0RVQ1RTX0RJUjsgfTsKCQkzM0NDMTBGMDIwNDRBM0M2MDAwM0MwNDUgLyogQXBwRGVsZWdhdGUuc3dpZnQgKi8gPSB7aXNhID0gUEJYRmlsZVJlZmVyZW5jZTsgbGFzdEtub3duRmlsZVR5cGUgPSBzb3VyY2Vjb2RlLnN3aWZ0OyBwYXRoID0gQXBwRGVsZWdhdGUuc3dpZnQ7IHNvdXJjZVRyZWUgPSAiPGdyb3VwPiI7IH07CgkJMzNDQzEwRjIyMDQ0QTNDNjAwMDNDMDQ1IC8qIEFzc2V0cy54Y2Fzc2V0cyAqLyA9IHtpc2EgPSBQQlhGaWxlUmVmZXJlbmNlOyBsYXN0S25vd25GaWxlVHlwZSA9IGZvbGRlci5hc3NldGNhdGFsb2c7IG5hbWUgPSBBc3NldHMueGNhc3NldHM7IHBhdGggPSBSdW5uZXIvQXNzZXRzLnhjYXNzZXRzOyBzb3VyY2VUcmVlID0gIjxncm91cD4iOyB9OwoJCTMzQ0MxMEY1MjA0NEEzQzYwMDAzQzA0NSAvKiBCYXNlICovID0ge2lzYSA9IFBCWEZpbGVSZWZlcmVuY2U7IGxhc3RLbm93bkZpbGVUeXBlID0gZmlsZS54aWI7IG5hbWUgPSBCYXNlOyBwYXRoID0gQmFzZS5scHJvai9NYWluTWVudS54aWI7IHNvdXJjZVRyZWUgPSAiPGdyb3VwPiI7IH07CgkJMzNDQzEwRjcyMDQ0QTNDNjAwMDNDMDQ1IC8qIEluZm8ucGxpc3QgKi8gPSB7aXNhID0gUEJYRmlsZVJlZmVyZW5jZTsgbGFzdEtub3duRmlsZVR5cGUgPSB0ZXh0LnBsaXN0LnhtbDsgbmFtZSA9IEluZm8ucGxpc3Q7IHBhdGggPSBSdW5uZXIvSW5mby5wbGlzdDsgc291cmNlVHJlZSA9ICI8Z3JvdXA+IjsgfTsKCQkzM0NDMTExMjIwNDRCRkEwMDAwM0MwNDUgLyogTWFpbkZsdXR0ZXJXaW5kb3cuc3dpZnQgKi8gPSB7aXNhID0gUEJYRmlsZVJlZmVyZW5jZTsgbGFzdEtub3duRmlsZVR5cGUgPSBzb3VyY2Vjb2RlLnN3aWZ0OyBwYXRoID0gTWFpbkZsdXR0ZXJXaW5kb3cuc3dpZnQ7IHNvdXJjZVRyZWUgPSAiPGdyb3VwPiI7IH07CgkJMzNDRUI0NzIyMkEwNTc3MTAwNEYyQUMwIC8qIEZsdXR0ZXItRGVidWcueGNjb25maWcgKi8gPSB7aXNhID0gUEJYRmlsZVJlZmVyZW5jZTsgbGFzdEtub3duRmlsZVR5cGUgPSB0ZXh0LnhjY29uZmlnOyBwYXRoID0gIkZsdXR0ZXItRGVidWcueGNjb25maWciOyBzb3VyY2VUcmVlID0gIjxncm91cD4iOyB9OwoJCTMzQ0VCNDc0MjJBMDU3NzEwMDRGMkFDMCAvKiBGbHV0dGVyLVJlbGVhc2UueGNjb25maWcgKi8gPSB7aXNhID0gUEJYRmlsZVJlZmVyZW5jZTsgbGFzdEtub3duRmlsZVR5cGUgPSB0ZXh0LnhjY29uZmlnOyBwYXRoID0gIkZsdXR0ZXItUmVsZWFzZS54Y2NvbmZpZyI7IHNvdXJjZVRyZWUgPSAiPGdyb3VwPiI7IH07CgkJMzNDRUI0NzcyMkEwNTc4QTAwNEYyQUMwIC8qIEZsdXR0ZXItR2VuZXJhdGVkLnhjY29uZmlnICovID0ge2lzYSA9IFBCWEZpbGVSZWZlcmVuY2U7IGxhc3RLbm93bkZpbGVUeXBlID0gdGV4dC54Y2NvbmZpZzsgbmFtZSA9ICJGbHV0dGVyLUdlbmVyYXRlZC54Y2NvbmZpZyI7IHBhdGggPSAiZXBoZW1lcmFsL0ZsdXR0ZXItR2VuZXJhdGVkLnhjY29uZmlnIjsgc291cmNlVHJlZSA9ICI8Z3JvdXA+IjsgfTsKCQkzM0U1MTkxMzIzMTc0N0Y0MDAyNkVFNEQgLyogRGVidWdQcm9maWxlLmVudGl0bGVtZW50cyAqLyA9IHtpc2EgPSBQQlhGaWxlUmVmZXJlbmNlOyBsYXN0S25vd25GaWxlVHlwZSA9IHRleHQucGxpc3QuZW50aXRsZW1lbnRzOyBwYXRoID0gRGVidWdQcm9maWxlLmVudGl0bGVtZW50czsgc291cmNlVHJlZSA9ICI8Z3JvdXA+IjsgfTsKCQkzM0U1MTkxNDIzMTc0OTM4MDAyNkVFNEQgLyogUmVsZWFzZS5lbnRpdGxlbWVudHMgKi8gPSB7aXNhID0gUEJYRmlsZVJlZmVyZW5jZTsgZmlsZUVuY29kaW5nID0gNDsgbGFzdEtub3duRmlsZVR5cGUgPSB0ZXh0LnBsaXN0LmVudGl0bGVtZW50czsgcGF0aCA9IFJlbGVhc2UuZW50aXRsZW1lbnRzOyBzb3VyY2VUcmVlID0gIjxncm91cD4iOyB9OwoJCTMzRTUxOTRGMjMyODI4ODYwMDI2RUU0RCAvKiBBcHBJbmZvLnhjY29uZmlnICovID0ge2lzYSA9IFBCWEZpbGVSZWZlcmVuY2U7IGxhc3RLbm93bkZpbGVUeXBlID0gdGV4dC54Y2NvbmZpZzsgcGF0aCA9IEFwcEluZm8ueGNjb25maWc7IHNvdXJjZVRyZWUgPSAiPGdyb3VwPiI7IH07CgkJN0FGQTNDOEUxRDM1MzYwQzAwODMwODJFIC8qIFJlbGVhc2UueGNjb25maWcgKi8gPSB7aXNhID0gUEJYRmlsZVJlZmVyZW5jZTsgbGFzdEtub3duRmlsZVR5cGUgPSB0ZXh0LnhjY29uZmlnOyBwYXRoID0gUmVsZWFzZS54Y2NvbmZpZzsgc291cmNlVHJlZSA9ICI8Z3JvdXA+IjsgfTsKCQk5NzQwRUVCMjFDRjkwMTk1MDA0Mzg0RkMgLyogRGVidWcueGNjb25maWcgKi8gPSB7aXNhID0gUEJYRmlsZVJlZmVyZW5jZTsgZmlsZUVuY29kaW5nID0gNDsgbGFzdEtub3duRmlsZVR5cGUgPSB0ZXh0LnhjY29uZmlnOyBwYXRoID0gRGVidWcueGNjb25maWc7IHNvdXJjZVRyZWUgPSAiPGdyb3VwPiI7IH07Ci8qIEVuZCBQQlhGaWxlUmVmZXJlbmNlIHNlY3Rpb24gKi8KCi8qIEJlZ2luIFBCWEZyYW1ld29ya3NCdWlsZFBoYXNlIHNlY3Rpb24gKi8KCQkzMzFDODBEMjI5NENGNzBGMDAyNjNCRTUgLyogRnJhbWV3b3JrcyAqLyA9IHsKCQkJaXNhID0gUEJYRnJhbWV3b3Jrc0J1aWxkUGhhc2U7CgkJCWJ1aWxkQWN0aW9uTWFzayA9IDIxNDc0ODM2NDc7CgkJCWZpbGVzID0gKAoJCQkpOwoJCQlydW5Pbmx5Rm9yRGVwbG95bWVudFBvc3Rwcm9jZXNzaW5nID0gMDsKCQl9OwoJCTMzQ0MxMEVBMjA0NEEzQzYwMDAzQzA0NSAvKiBGcmFtZXdvcmtzICovID0gewoJCQlpc2EgPSBQQlhGcmFtZXdvcmtzQnVpbGRQaGFzZTsKCQkJYnVpbGRBY3Rpb25NYXNrID0gMjE0NzQ4MzY0NzsKCQkJZmlsZXMgPSAoCgkJCSk7CgkJCXJ1bk9ubHlGb3JEZXBsb3ltZW50UG9zdHByb2Nlc3NpbmcgPSAwOwoJCX07Ci8qIEVuZCBQQlhGcmFtZXdvcmtzQnVpbGRQaGFzZSBzZWN0aW9uICovCgovKiBCZWdpbiBQQlhHcm91cCBzZWN0aW9uICovCgkJMzMxQzgwRDYyOTRDRjcxMDAwMjYzQkU1IC8qIFJ1bm5lclRlc3RzICovID0gewoJCQlpc2EgPSBQQlhHcm91cDsKCQkJY2hpbGRyZW4gPSAoCgkJCQkzMzFDODBENzI5NENGNzEwMDAyNjNCRTUgLyogUnVubmVyVGVzdHMuc3dpZnQgKi8sCgkJCSk7CgkJCXBhdGggPSBSdW5uZXJUZXN0czsKCQkJc291cmNlVHJlZSA9ICI8Z3JvdXA+IjsKCQl9OwoJCTMzQkE4ODZBMjI2RTc4QUYwMDMzMjlENSAvKiBDb25maWdzICovID0gewoJCQlpc2EgPSBQQlhHcm91cDsKCQkJY2hpbGRyZW4gPSAoCgkJCQkzM0U1MTk0RjIzMjgyODg2MDAyNkVFNEQgLyogQXBwSW5mby54Y2NvbmZpZyAqLywKCQkJCTk3NDBFRUIyMUNGOTAxOTUwMDQzODRGQyAvKiBEZWJ1Zy54Y2NvbmZpZyAqLywKCQkJCTdBRkEzQzhFMUQzNTM2MEMwMDgzMDgyRSAvKiBSZWxlYXNlLnhjY29uZmlnICovLAoJCQkJMzMzMDAwRUQyMkQzREU1RDAwNTU0MTYyIC8qIFdhcm5pbmdzLnhjY29uZmlnICovLAoJCQkpOwoJCQlwYXRoID0gQ29uZmlnczsKCQkJc291cmNlVHJlZSA9ICI8Z3JvdXA+IjsKCQl9OwoJCTMzQ0MxMEU0MjA0NEEzQzYwMDAzQzA0NSA9IHsKCQkJaXNhID0gUEJYR3JvdXA7CgkJCWNoaWxkcmVuID0gKAoJCQkJMzNGQUI2NzEyMzI4MzY3NDAwNjVBQzFFIC8qIFJ1bm5lciAqLywKCQkJCTMzQ0VCNDcxMjJBMDU3NzEwMDRGMkFDMCAvKiBGbHV0dGVyICovLAoJCQkJMzMxQzgwRDYyOTRDRjcxMDAwMjYzQkU1IC8qIFJ1bm5lclRlc3RzICovLAoJCQkJMzNDQzEwRUUyMDQ0QTNDNjAwMDNDMDQ1IC8qIFByb2R1Y3RzICovLAoJCQkJRDczOTEyRUMyMkYzN0YzRDAwMEQxM0EwIC8qIEZyYW1ld29ya3MgKi8sCgkJCSk7CgkJCXNvdXJjZVRyZWUgPSAiPGdyb3VwPiI7CgkJfTsKCQkzM0NDMTBFRTIwNDRBM0M2MDAwM0MwNDUgLyogUHJvZHVjdHMgKi8gPSB7CgkJCWlzYSA9IFBCWEdyb3VwOwoJCQljaGlsZHJlbiA9ICgKCQkJCTMzQ0MxMEVEMjA0NEEzQzYwMDAzQzA0NSAvKiB7e3Byb2plY3RfbmFtZX19LmFwcCAqLywKCQkJCTMzMUM4MEQ1Mjk0Q0Y3MTAwMDI2M0JFNSAvKiBSdW5uZXJUZXN0cy54Y3Rlc3QgKi8sCgkJCSk7CgkJCW5hbWUgPSBQcm9kdWN0czsKCQkJc291cmNlVHJlZSA9ICI8Z3JvdXA+IjsKCQl9OwoJCTMzQ0MxMTI0MjA0NEQ2NkUwMDAzQzA0NSAvKiBSZXNvdXJjZXMgKi8gPSB7CgkJCWlzYSA9IFBCWEdyb3VwOwoJCQljaGlsZHJlbiA9ICgKCQkJCTMzQ0MxMEYyMjA0NEEzQzYwMDAzQzA0NSAvKiBBc3NldHMueGNhc3NldHMgKi8sCgkJCQkzM0NDMTBGNDIwNDRBM0M2MDAwM0MwNDUgLyogTWFpbk1lbnUueGliICovLAoJCQkJMzNDQzEwRjcyMDQ0QTNDNjAwMDNDMDQ1IC8qIEluZm8ucGxpc3QgKi8sCgkJCSk7CgkJCW5hbWUgPSBSZXNvdXJjZXM7CgkJCXBhdGggPSAuLjsKCQkJc291cmNlVHJlZSA9ICI8Z3JvdXA+IjsKCQl9OwoJCTMzQ0VCNDcxMjJBMDU3NzEwMDRGMkFDMCAvKiBGbHV0dGVyICovID0gewoJCQlpc2EgPSBQQlhHcm91cDsKCQkJY2hpbGRyZW4gPSAoCgkJCQkzMzVCQkQxQTIyQTlBMTVFMDBFOTA3MUQgLyogR2VuZXJhdGVkUGx1Z2luUmVnaXN0cmFudC5zd2lmdCAqLywKCQkJCTMzQ0VCNDcyMjJBMDU3NzEwMDRGMkFDMCAvKiBGbHV0dGVyLURlYnVnLnhjY29uZmlnICovLAoJCQkJMzNDRUI0NzQyMkEwNTc3MTAwNEYyQUMwIC8qIEZsdXR0ZXItUmVsZWFzZS54Y2NvbmZpZyAqLywKCQkJCTMzQ0VCNDc3MjJBMDU3OEEwMDRGMkFDMCAvKiBGbHV0dGVyLUdlbmVyYXRlZC54Y2NvbmZpZyAqLywKCQkJKTsKCQkJcGF0aCA9IEZsdXR0ZXI7CgkJCXNvdXJjZVRyZWUgPSAiPGdyb3VwPiI7CgkJfTsKCQkzM0ZBQjY3MTIzMjgzNjc0MDA2NUFDMUUgLyogUnVubmVyICovID0gewoJCQlpc2EgPSBQQlhHcm91cDsKCQkJY2hpbGRyZW4gPSAoCgkJCQkzM0NDMTBGMDIwNDRBM0M2MDAwM0MwNDUgLyogQXBwRGVsZWdhdGUuc3dpZnQgKi8sCgkJCQkzM0NDMTExMjIwNDRCRkEwMDAwM0MwNDUgLyogTWFpbkZsdXR0ZXJXaW5kb3cuc3dpZnQgKi8sCgkJCQkzM0U1MTkxMzIzMTc0N0Y0MDAyNkVFNEQgLyogRGVidWdQcm9maWxlLmVudGl0bGVtZW50cyAqLywKCQkJCTMzRTUxOTE0MjMxNzQ5MzgwMDI2RUU0RCAvKiBSZWxlYXNlLmVudGl0bGVtZW50cyAqLywKCQkJCTMzQ0MxMTI0MjA0NEQ2NkUwMDAzQzA0NSAvKiBSZXNvdXJjZXMgKi8sCgkJCQkzM0JBODg2QTIyNkU3OEFGMDAzMzI5RDUgLyogQ29uZmlncyAqLywKCQkJKTsKCQkJcGF0aCA9IFJ1bm5lcjsKCQkJc291cmNlVHJlZSA9ICI8Z3JvdXA+IjsKCQl9OwoJCUQ3MzkxMkVDMjJGMzdGM0QwMDBEMTNBMCAvKiBGcmFtZXdvcmtzICovID0gewoJCQlpc2EgPSBQQlhHcm91cDsKCQkJY2hpbGRyZW4gPSAoCgkJCSk7CgkJCW5hbWUgPSBGcmFtZXdvcmtzOwoJCQlzb3VyY2VUcmVlID0gIjxncm91cD4iOwoJCX07Ci8qIEVuZCBQQlhHcm91cCBzZWN0aW9uICovCgovKiBCZWdpbiBQQlhOYXRpdmVUYXJnZXQgc2VjdGlvbiAqLwoJCTMzMUM4MEQ0Mjk0Q0Y3MEYwMDI2M0JFNSAvKiBSdW5uZXJUZXN0cyAqLyA9IHsKCQkJaXNhID0gUEJYTmF0aXZlVGFyZ2V0OwoJCQlidWlsZENvbmZpZ3VyYXRpb25MaXN0ID0gMzMxQzgwREUyOTRDRjcxMDAwMjYzQkU1IC8qIEJ1aWxkIGNvbmZpZ3VyYXRpb24gbGlzdCBmb3IgUEJYTmF0aXZlVGFyZ2V0ICJSdW5uZXJUZXN0cyIgKi87CgkJCWJ1aWxkUGhhc2VzID0gKAoJCQkJMzMxQzgwRDEyOTRDRjcwRjAwMjYzQkU1IC8qIFNvdXJjZXMgKi8sCgkJCQkzMzFDODBEMjI5NENGNzBGMDAyNjNCRTUgLyogRnJhbWV3b3JrcyAqLywKCQkJCTMzMUM4MEQzMjk0Q0Y3MEYwMDI2M0JFNSAvKiBSZXNvdXJjZXMgKi8sCgkJCSk7CgkJCWJ1aWxkUnVsZXMgPSAoCgkJCSk7CgkJCWRlcGVuZGVuY2llcyA9ICgKCQkJCTMzMUM4MERBMjk0Q0Y3MTAwMDI2M0JFNSAvKiBQQlhUYXJnZXREZXBlbmRlbmN5ICovLAoJCQkpOwoJCQluYW1lID0gUnVubmVyVGVzdHM7CgkJCXByb2R1Y3ROYW1lID0gUnVubmVyVGVzdHM7CgkJCXByb2R1Y3RSZWZlcmVuY2UgPSAzMzFDODBENTI5NENGNzEwMDAyNjNCRTUgLyogUnVubmVyVGVzdHMueGN0ZXN0ICovOwoJCQlwcm9kdWN0VHlwZSA9ICJjb20uYXBwbGUucHJvZHVjdC10eXBlLmJ1bmRsZS51bml0LXRlc3QiOwoJCX07CgkJMzNDQzEwRUMyMDQ0QTNDNjAwMDNDMDQ1IC8qIFJ1bm5lciAqLyA9IHsKCQkJaXNhID0gUEJYTmF0aXZlVGFyZ2V0OwoJCQlidWlsZENvbmZpZ3VyYXRpb25MaXN0ID0gMzNDQzEwRkIyMDQ0QTNDNjAwMDNDMDQ1IC8qIEJ1aWxkIGNvbmZpZ3VyYXRpb24gbGlzdCBmb3IgUEJYTmF0aXZlVGFyZ2V0ICJSdW5uZXIiICovOwoJCQlidWlsZFBoYXNlcyA9ICgKCQkJCTMzQ0MxMEU5MjA0NEEzQzYwMDAzQzA0NSAvKiBTb3VyY2VzICovLAoJCQkJMzNDQzEwRUEyMDQ0QTNDNjAwMDNDMDQ1IC8qIEZyYW1ld29ya3MgKi8sCgkJCQkzM0NDMTBFQjIwNDRBM0M2MDAwM0MwNDUgLyogUmVzb3VyY2VzICovLAoJCQkJMzNDQzExMEUyMDQ0QTg4NDAwMDNDMDQ1IC8qIEJ1bmRsZSBGcmFtZXdvcmsgKi8sCgkJCQkzMzk5RDQ5MDIyOEIyNENGMDA5QTc5QzcgLyogU2hlbGxTY3JpcHQgKi8sCgkJCSk7CgkJCWJ1aWxkUnVsZXMgPSAoCgkJCSk7CgkJCWRlcGVuZGVuY2llcyA9ICgKCQkJCTMzQ0MxMTIwMjA0NEM3OUYwMDAzQzA0NSAvKiBQQlhUYXJnZXREZXBlbmRlbmN5ICovLAoJCQkpOwoJCQluYW1lID0gUnVubmVyOwoJCQlwcm9kdWN0TmFtZSA9IFJ1bm5lcjsKCQkJcHJvZHVjdFJlZmVyZW5jZSA9IDMzQ0MxMEVEMjA0NEEzQzYwMDAzQzA0NSAvKiB7e3Byb2plY3RfbmFtZX19LmFwcCAqLzsKCQkJcHJvZHVjdFR5cGUgPSAiY29tLmFwcGxlLnByb2R1Y3QtdHlwZS5hcHBsaWNhdGlvbiI7CgkJfTsKLyogRW5kIFBCWE5hdGl2ZVRhcmdldCBzZWN0aW9uICovCgovKiBCZWdpbiBQQlhQcm9qZWN0IHNlY3Rpb24gKi8KCQkzM0NDMTBFNTIwNDRBM0M2MDAwM0MwNDUgLyogUHJvamVjdCBvYmplY3QgKi8gPSB7CgkJCWlzYSA9IFBCWFByb2plY3Q7CgkJCWF0dHJpYnV0ZXMgPSB7CgkJCQlMYXN0U3dpZnRVcGRhdGVDaGVjayA9IDA5MjA7CgkJCQlMYXN0VXBncmFkZUNoZWNrID0gMTMwMDsKCQkJCU9SR0FOSVpBVElPTk5BTUUgPSAiIjsKCQkJCVRhcmdldEF0dHJpYnV0ZXMgPSB7CgkJCQkJMzMxQzgwRDQyOTRDRjcwRjAwMjYzQkU1ID0gewoJCQkJCQlDcmVhdGVkT25Ub29sc1ZlcnNpb24gPSAxNC4wOwoJCQkJCQlUZXN0VGFyZ2V0SUQgPSAzM0NDMTBFQzIwNDRBM0M2MDAwM0MwNDU7CgkJCQkJfTsKCQkJCQkzM0NDMTBFQzIwNDRBM0M2MDAwM0MwNDUgPSB7CgkJCQkJCUNyZWF0ZWRPblRvb2xzVmVyc2lvbiA9IDkuMjsKCQkJCQkJTGFzdFN3aWZ0TWlncmF0aW9uID0gMTEwMDsKCQkJCQkJUHJvdmlzaW9uaW5nU3R5bGUgPSBBdXRvbWF0aWM7CgkJCQkJCVN5c3RlbUNhcGFiaWxpdGllcyA9IHsKCQkJCQkJCWNvbS5hcHBsZS5TYW5kYm94ID0gewoJCQkJCQkJCWVuYWJsZWQgPSAxOwoJCQkJCQkJfTsKCQkJCQkJfTsKCQkJCQl9OwoJCQkJCTMzQ0MxMTFBMjA0NEM2QkEwMDAzQzA0NSA9IHsKCQkJCQkJQ3JlYXRlZE9uVG9vbHNWZXJzaW9uID0gOS4yOwoJCQkJCQlQcm92aXNpb25pbmdTdHlsZSA9IE1hbnVhbDsKCQkJCQl9OwoJCQkJfTsKCQkJfTsKCQkJYnVpbGRDb25maWd1cmF0aW9uTGlzdCA9IDMzQ0MxMEU4MjA0NEEzQzYwMDAzQzA0NSAvKiBCdWlsZCBjb25maWd1cmF0aW9uIGxpc3QgZm9yIFBCWFByb2plY3QgIlJ1bm5lciIgKi87CgkJCWNvbXBhdGliaWxpdHlWZXJzaW9uID0gIlhjb2RlIDkuMyI7CgkJCWRldmVsb3BtZW50UmVnaW9uID0gZW47CgkJCWhhc1NjYW5uZWRGb3JFbmNvZGluZ3MgPSAwOwoJCQlrbm93blJlZ2lvbnMgPSAoCgkJCQllbiwKCQkJCUJhc2UsCgkJCSk7CgkJCW1haW5Hcm91cCA9IDMzQ0MxMEU0MjA0NEEzQzYwMDAzQzA0NTsKCQkJcHJvZHVjdFJlZkdyb3VwID0gMzNDQzEwRUUyMDQ0QTNDNjAwMDNDMDQ1IC8qIFByb2R1Y3RzICovOwoJCQlwcm9qZWN0RGlyUGF0aCA9ICIiOwoJCQlwcm9qZWN0Um9vdCA9ICIiOwoJCQl0YXJnZXRzID0gKAoJCQkJMzNDQzEwRUMyMDQ0QTNDNjAwMDNDMDQ1IC8qIFJ1bm5lciAqLywKCQkJCTMzMUM4MEQ0Mjk0Q0Y3MEYwMDI2M0JFNSAvKiBSdW5uZXJUZXN0cyAqLywKCQkJCTMzQ0MxMTFBMjA0NEM2QkEwMDAzQzA0NSAvKiBGbHV0dGVyIEFzc2VtYmxlICovLAoJCQkpOwoJCX07Ci8qIEVuZCBQQlhQcm9qZWN0IHNlY3Rpb24gKi8KCi8qIEJlZ2luIFBCWFJlc291cmNlc0J1aWxkUGhhc2Ugc2VjdGlvbiAqLwoJCTMzMUM4MEQzMjk0Q0Y3MEYwMDI2M0JFNSAvKiBSZXNvdXJjZXMgKi8gPSB7CgkJCWlzYSA9IFBCWFJlc291cmNlc0J1aWxkUGhhc2U7CgkJCWJ1aWxkQWN0aW9uTWFzayA9IDIxNDc0ODM2NDc7CgkJCWZpbGVzID0gKAoJCQkpOwoJCQlydW5Pbmx5Rm9yRGVwbG95bWVudFBvc3Rwcm9jZXNzaW5nID0gMDsKCQl9OwoJCTMzQ0MxMEVCMjA0NEEzQzYwMDAzQzA0NSAvKiBSZXNvdXJjZXMgKi8gPSB7CgkJCWlzYSA9IFBCWFJlc291cmNlc0J1aWxkUGhhc2U7CgkJCWJ1aWxkQWN0aW9uTWFzayA9IDIxNDc0ODM2NDc7CgkJCWZpbGVzID0gKAoJCQkJMzNDQzEwRjMyMDQ0QTNDNjAwMDNDMDQ1IC8qIEFzc2V0cy54Y2Fzc2V0cyBpbiBSZXNvdXJjZXMgKi8sCgkJCQkzM0NDMTBGNjIwNDRBM0M2MDAwM0MwNDUgLyogTWFpbk1lbnUueGliIGluIFJlc291cmNlcyAqLywKCQkJKTsKCQkJcnVuT25seUZvckRlcGxveW1lbnRQb3N0cHJvY2Vzc2luZyA9IDA7CgkJfTsKLyogRW5kIFBCWFJlc291cmNlc0J1aWxkUGhhc2Ugc2VjdGlvbiAqLwoKLyogQmVnaW4gUEJYU2hlbGxTY3JpcHRCdWlsZFBoYXNlIHNlY3Rpb24gKi8KCQkzMzk5RDQ5MDIyOEIyNENGMDA5QTc5QzcgLyogU2hlbGxTY3JpcHQgKi8gPSB7CgkJCWlzYSA9IFBCWFNoZWxsU2NyaXB0QnVpbGRQaGFzZTsKCQkJYWx3YXlzT3V0T2ZEYXRlID0gMTsKCQkJYnVpbGRBY3Rpb25NYXNrID0gMjE0NzQ4MzY0NzsKCQkJZmlsZXMgPSAoCgkJCSk7CgkJCWlucHV0RmlsZUxpc3RQYXRocyA9ICgKCQkJKTsKCQkJaW5wdXRQYXRocyA9ICgKCQkJKTsKCQkJb3V0cHV0RmlsZUxpc3RQYXRocyA9ICgKCQkJKTsKCQkJb3V0cHV0UGF0aHMgPSAoCgkJCSk7CgkJCXJ1bk9ubHlGb3JEZXBsb3ltZW50UG9zdHByb2Nlc3NpbmcgPSAwOwoJCQlzaGVsbFBhdGggPSAvYmluL3NoOwoJCQlzaGVsbFNjcmlwdCA9ICJlY2hvIFwiJFBST0RVQ1RfTkFNRS5hcHBcIiA+IFwiJFBST0pFQ1RfRElSXCIvRmx1dHRlci9lcGhlbWVyYWwvLmFwcF9maWxlbmFtZSAmJiBcIiRGTFVUVEVSX1JPT1RcIi9wYWNrYWdlcy9mbHV0dGVyX3Rvb2xzL2Jpbi9tYWNvc19hc3NlbWJsZS5zaCBlbWJlZFxuIjsKCQl9OwoJCTMzQ0MxMTFFMjA0NEM2QkYwMDAzQzA0NSAvKiBTaGVsbFNjcmlwdCAqLyA9IHsKCQkJaXNhID0gUEJYU2hlbGxTY3JpcHRCdWlsZFBoYXNlOwoJCQlidWlsZEFjdGlvbk1hc2sgPSAyMTQ3NDgzNjQ3OwoJCQlmaWxlcyA9ICgKCQkJKTsKCQkJaW5wdXRGaWxlTGlzdFBhdGhzID0gKAoJCQkJRmx1dHRlci9lcGhlbWVyYWwvRmx1dHRlcklucHV0cy54Y2ZpbGVsaXN0LAoJCQkpOwoJCQlpbnB1dFBhdGhzID0gKAoJCQkJRmx1dHRlci9lcGhlbWVyYWwvdHJpcHdpcmUsCgkJCSk7CgkJCW91dHB1dEZpbGVMaXN0UGF0aHMgPSAoCgkJCQlGbHV0dGVyL2VwaGVtZXJhbC9GbHV0dGVyT3V0cHV0cy54Y2ZpbGVsaXN0LAoJCQkpOwoJCQlvdXRwdXRQYXRocyA9ICgKCQkJKTsKCQkJcnVuT25seUZvckRlcGxveW1lbnRQb3N0cHJvY2Vzc2luZyA9IDA7CgkJCXNoZWxsUGF0aCA9IC9iaW4vc2g7CgkJCXNoZWxsU2NyaXB0ID0gIlwiJEZMVVRURVJfUk9PVFwiL3BhY2thZ2VzL2ZsdXR0ZXJfdG9vbHMvYmluL21hY29zX2Fzc2VtYmxlLnNoICYmIHRvdWNoIEZsdXR0ZXIvZXBoZW1lcmFsL3RyaXB3aXJlIjsKCQl9OwovKiBFbmQgUEJYU2hlbGxTY3JpcHRCdWlsZFBoYXNlIHNlY3Rpb24gKi8KCi8qIEJlZ2luIFBCWFNvdXJjZXNCdWlsZFBoYXNlIHNlY3Rpb24gKi8KCQkzMzFDODBEMTI5NENGNzBGMDAyNjNCRTUgLyogU291cmNlcyAqLyA9IHsKCQkJaXNhID0gUEJYU291cmNlc0J1aWxkUGhhc2U7CgkJCWJ1aWxkQWN0aW9uTWFzayA9IDIxNDc0ODM2NDc7CgkJCWZpbGVzID0gKAoJCQkJMzMxQzgwRDgyOTRDRjcxMDAwMjYzQkU1IC8qIFJ1bm5lclRlc3RzLnN3aWZ0IGluIFNvdXJjZXMgKi8sCgkJCSk7CgkJCXJ1bk9ubHlGb3JEZXBsb3ltZW50UG9zdHByb2Nlc3NpbmcgPSAwOwoJCX07CgkJMzNDQzEwRTkyMDQ0QTNDNjAwMDNDMDQ1IC8qIFNvdXJjZXMgKi8gPSB7CgkJCWlzYSA9IFBCWFNvdXJjZXNCdWlsZFBoYXNlOwoJCQlidWlsZEFjdGlvbk1hc2sgPSAyMTQ3NDgzNjQ3OwoJCQlmaWxlcyA9ICgKCQkJCTMzQ0MxMTEzMjA0NEJGQTAwMDAzQzA0NSAvKiBNYWluRmx1dHRlcldpbmRvdy5zd2lmdCBpbiBTb3VyY2VzICovLAoJCQkJMzNDQzEwRjEyMDQ0QTNDNjAwMDNDMDQ1IC8qIEFwcERlbGVnYXRlLnN3aWZ0IGluIFNvdXJjZXMgKi8sCgkJCQkzMzVCQkQxQjIyQTlBMTVFMDBFOTA3MUQgLyogR2VuZXJhdGVkUGx1Z2luUmVnaXN0cmFudC5zd2lmdCBpbiBTb3VyY2VzICovLAoJCQkpOwoJCQlydW5Pbmx5Rm9yRGVwbG95bWVudFBvc3Rwcm9jZXNzaW5nID0gMDsKCQl9OwovKiBFbmQgUEJYU291cmNlc0J1aWxkUGhhc2Ugc2VjdGlvbiAqLwoKLyogQmVnaW4gUEJYVGFyZ2V0RGVwZW5kZW5jeSBzZWN0aW9uICovCgkJMzMxQzgwREEyOTRDRjcxMDAwMjYzQkU1IC8qIFBCWFRhcmdldERlcGVuZGVuY3kgKi8gPSB7CgkJCWlzYSA9IFBCWFRhcmdldERlcGVuZGVuY3k7CgkJCXRhcmdldCA9IDMzQ0MxMEVDMjA0NEEzQzYwMDAzQzA0NSAvKiBSdW5uZXIgKi87CgkJCXRhcmdldFByb3h5ID0gMzMxQzgwRDkyOTRDRjcxMDAwMjYzQkU1IC8qIFBCWENvbnRhaW5lckl0ZW1Qcm94eSAqLzsKCQl9OwoJCTMzQ0MxMTIwMjA0NEM3OUYwMDAzQzA0NSAvKiBQQlhUYXJnZXREZXBlbmRlbmN5ICovID0gewoJCQlpc2EgPSBQQlhUYXJnZXREZXBlbmRlbmN5OwoJCQl0YXJnZXQgPSAzM0NDMTExQTIwNDRDNkJBMDAwM0MwNDUgLyogRmx1dHRlciBBc3NlbWJsZSAqLzsKCQkJdGFyZ2V0UHJveHkgPSAzM0NDMTExRjIwNDRDNzlGMDAwM0MwNDUgLyogUEJYQ29udGFpbmVySXRlbVByb3h5ICovOwoJCX07Ci8qIEVuZCBQQlhUYXJnZXREZXBlbmRlbmN5IHNlY3Rpb24gKi8KCi8qIEJlZ2luIFBCWFZhcmlhbnRHcm91cCBzZWN0aW9uICovCgkJMzNDQzEwRjQyMDQ0QTNDNjAwMDNDMDQ1IC8qIE1haW5NZW51LnhpYiAqLyA9IHsKCQkJaXNhID0gUEJYVmFyaWFudEdyb3VwOwoJCQljaGlsZHJlbiA9ICgKCQkJCTMzQ0MxMEY1MjA0NEEzQzYwMDAzQzA0NSAvKiBCYXNlICovLAoJCQkpOwoJCQluYW1lID0gTWFpbk1lbnUueGliOwoJCQlwYXRoID0gUnVubmVyOwoJCQlzb3VyY2VUcmVlID0gIjxncm91cD4iOwoJCX07Ci8qIEVuZCBQQlhWYXJpYW50R3JvdXAgc2VjdGlvbiAqLwoKLyogQmVnaW4gWENCdWlsZENvbmZpZ3VyYXRpb24gc2VjdGlvbiAqLwoJCTMzMUM4MERCMjk0Q0Y3MTAwMDI2M0JFNSAvKiBEZWJ1ZyAqLyA9IHsKCQkJaXNhID0gWENCdWlsZENvbmZpZ3VyYXRpb247CgkJCWJ1aWxkU2V0dGluZ3MgPSB7CgkJCQlCVU5ETEVfTE9BREVSID0gIiQoVEVTVF9IT1NUKSI7CgkJCQlDVVJSRU5UX1BST0pFQ1RfVkVSU0lPTiA9IDE7CgkJCQlHRU5FUkFURV9JTkZPUExJU1RfRklMRSA9IFlFUzsKCQkJCU1BUktFVElOR19WRVJTSU9OID0gMS4wOwoJCQkJUFJPRFVDVF9CVU5ETEVfSURFTlRJRklFUiA9IHt7b3JnX25hbWV9fS57e3Byb2plY3RfbmFtZS5jYW1lbENhc2UoKX19LlJ1bm5lclRlc3RzOwoJCQkJUFJPRFVDVF9OQU1FID0gIiQoVEFSR0VUX05BTUUpIjsKCQkJCVNXSUZUX1ZFUlNJT04gPSA1LjA7CgkJCQlURVNUX0hPU1QgPSAiJChCVUlMVF9QUk9EVUNUU19ESVIpL3t7cHJvamVjdF9uYW1lfX0uYXBwLyQoQlVORExFX0VYRUNVVEFCTEVfRk9MREVSX1BBVEgpL3t7cHJvamVjdF9uYW1lfX0iOwoJCQl9OwoJCQluYW1lID0gRGVidWc7CgkJfTsKCQkzMzFDODBEQzI5NENGNzEwMDAyNjNCRTUgLyogUmVsZWFzZSAqLyA9IHsKCQkJaXNhID0gWENCdWlsZENvbmZpZ3VyYXRpb247CgkJCWJ1aWxkU2V0dGluZ3MgPSB7CgkJCQlCVU5ETEVfTE9BREVSID0gIiQoVEVTVF9IT1NUKSI7CgkJCQlDVVJSRU5UX1BST0pFQ1RfVkVSU0lPTiA9IDE7CgkJCQlHRU5FUkFURV9JTkZPUExJU1RfRklMRSA9IFlFUzsKCQkJCU1BUktFVElOR19WRVJTSU9OID0gMS4wOwoJCQkJUFJPRFVDVF9CVU5ETEVfSURFTlRJRklFUiA9IHt7b3JnX25hbWV9fS57e3Byb2plY3RfbmFtZS5jYW1lbENhc2UoKX19LlJ1bm5lclRlc3RzOwoJCQkJUFJPRFVDVF9OQU1FID0gIiQoVEFSR0VUX05BTUUpIjsKCQkJCVNXSUZUX1ZFUlNJT04gPSA1LjA7CgkJCQlURVNUX0hPU1QgPSAiJChCVUlMVF9QUk9EVUNUU19ESVIpL3t7cHJvamVjdF9uYW1lfX0uYXBwLyQoQlVORExFX0VYRUNVVEFCTEVfRk9MREVSX1BBVEgpL3t7cHJvamVjdF9uYW1lfX0iOwoJCQl9OwoJCQluYW1lID0gUmVsZWFzZTsKCQl9OwoJCTMzMUM4MEREMjk0Q0Y3MTAwMDI2M0JFNSAvKiBQcm9maWxlICovID0gewoJCQlpc2EgPSBYQ0J1aWxkQ29uZmlndXJhdGlvbjsKCQkJYnVpbGRTZXR0aW5ncyA9IHsKCQkJCUJVTkRMRV9MT0FERVIgPSAiJChURVNUX0hPU1QpIjsKCQkJCUNVUlJFTlRfUFJPSkVDVF9WRVJTSU9OID0gMTsKCQkJCUdFTkVSQVRFX0lORk9QTElTVF9GSUxFID0gWUVTOwoJCQkJTUFSS0VUSU5HX1ZFUlNJT04gPSAxLjA7CgkJCQlQUk9EVUNUX0JVTkRMRV9JREVOVElGSUVSID0ge3tvcmdfbmFtZX19Lnt7cHJvamVjdF9uYW1lLmNhbWVsQ2FzZSgpfX0uUnVubmVyVGVzdHM7CgkJCQlQUk9EVUNUX05BTUUgPSAiJChUQVJHRVRfTkFNRSkiOwoJCQkJU1dJRlRfVkVSU0lPTiA9IDUuMDsKCQkJCVRFU1RfSE9TVCA9ICIkKEJVSUxUX1BST0RVQ1RTX0RJUikve3twcm9qZWN0X25hbWV9fS5hcHAvJChCVU5ETEVfRVhFQ1VUQUJMRV9GT0xERVJfUEFUSCkve3twcm9qZWN0X25hbWV9fSI7CgkJCX07CgkJCW5hbWUgPSBQcm9maWxlOwoJCX07CgkJMzM4RDBDRTkyMzE0NThCRDAwRkE1Rjc1IC8qIFByb2ZpbGUgKi8gPSB7CgkJCWlzYSA9IFhDQnVpbGRDb25maWd1cmF0aW9uOwoJCQliYXNlQ29uZmlndXJhdGlvblJlZmVyZW5jZSA9IDdBRkEzQzhFMUQzNTM2MEMwMDgzMDgyRSAvKiBSZWxlYXNlLnhjY29uZmlnICovOwoJCQlidWlsZFNldHRpbmdzID0gewoJCQkJQUxXQVlTX1NFQVJDSF9VU0VSX1BBVEhTID0gTk87CgkJCQlDTEFOR19BTkFMWVpFUl9OT05OVUxMID0gWUVTOwoJCQkJQ0xBTkdfQU5BTFlaRVJfTlVNQkVSX09CSkVDVF9DT05WRVJTSU9OID0gWUVTX0FHR1JFU1NJVkU7CgkJCQlDTEFOR19DWFhfTEFOR1VBR0VfU1RBTkRBUkQgPSAiZ251KysxNCI7CgkJCQlDTEFOR19DWFhfTElCUkFSWSA9ICJsaWJjKysiOwoJCQkJQ0xBTkdfRU5BQkxFX01PRFVMRVMgPSBZRVM7CgkJCQlDTEFOR19FTkFCTEVfT0JKQ19BUkMgPSBZRVM7CgkJCQlDTEFOR19XQVJOX0JMT0NLX0NBUFRVUkVfQVVUT1JFTEVBU0lORyA9IFlFUzsKCQkJCUNMQU5HX1dBUk5fQk9PTF9DT05WRVJTSU9OID0gWUVTOwoJCQkJQ0xBTkdfV0FSTl9DT05TVEFOVF9DT05WRVJTSU9OID0gWUVTOwoJCQkJQ0xBTkdfV0FSTl9ERVBSRUNBVEVEX09CSkNfSU1QTEVNRU5UQVRJT05TID0gWUVTOwoJCQkJQ0xBTkdfV0FSTl9ESVJFQ1RfT0JKQ19JU0FfVVNBR0UgPSBZRVNfRVJST1I7CgkJCQlDTEFOR19XQVJOX0RPQ1VNRU5UQVRJT05fQ09NTUVOVFMgPSBZRVM7CgkJCQlDTEFOR19XQVJOX0VNUFRZX0JPRFkgPSBZRVM7CgkJCQlDTEFOR19XQVJOX0VOVU1fQ09OVkVSU0lPTiA9IFlFUzsKCQkJCUNMQU5HX1dBUk5fSU5GSU5JVEVfUkVDVVJTSU9OID0gWUVTOwoJCQkJQ0xBTkdfV0FSTl9JTlRfQ09OVkVSU0lPTiA9IFlFUzsKCQkJCUNMQU5HX1dBUk5fTk9OX0xJVEVSQUxfTlVMTF9DT05WRVJTSU9OID0gWUVTOwoJCQkJQ0xBTkdfV0FSTl9PQkpDX0xJVEVSQUxfQ09OVkVSU0lPTiA9IFlFUzsKCQkJCUNMQU5HX1dBUk5fT0JKQ19ST09UX0NMQVNTID0gWUVTX0VSUk9SOwoJCQkJQ0xBTkdfV0FSTl9SQU5HRV9MT09QX0FOQUxZU0lTID0gWUVTOwoJCQkJQ0xBTkdfV0FSTl9TVVNQSUNJT1VTX01PVkUgPSBZRVM7CgkJCQlDT0RFX1NJR05fSURFTlRJVFkgPSAiLSI7CgkJCQlDT1BZX1BIQVNFX1NUUklQID0gTk87CgkJCQlERUJVR19JTkZPUk1BVElPTl9GT1JNQVQgPSAiZHdhcmYtd2l0aC1kc3ltIjsKCQkJCUVOQUJMRV9OU19BU1NFUlRJT05TID0gTk87CgkJCQlFTkFCTEVfU1RSSUNUX09CSkNfTVNHU0VORCA9IFlFUzsKCQkJCUdDQ19DX0xBTkdVQUdFX1NUQU5EQVJEID0gZ251MTE7CgkJCQlHQ0NfTk9fQ09NTU9OX0JMT0NLUyA9IFlFUzsKCQkJCUdDQ19XQVJOXzY0X1RPXzMyX0JJVF9DT05WRVJTSU9OID0gWUVTOwoJCQkJR0NDX1dBUk5fQUJPVVRfUkVUVVJOX1RZUEUgPSBZRVNfRVJST1I7CgkJCQlHQ0NfV0FSTl9VTklOSVRJQUxJWkVEX0FVVE9TID0gWUVTX0FHR1JFU1NJVkU7CgkJCQlHQ0NfV0FSTl9VTlVTRURfRlVOQ1RJT04gPSBZRVM7CgkJCQlHQ0NfV0FSTl9VTlVTRURfVkFSSUFCTEUgPSBZRVM7CgkJCQlNQUNPU1hfREVQTE9ZTUVOVF9UQVJHRVQgPSAxMC4xNS43OwoJCQkJTVRMX0VOQUJMRV9ERUJVR19JTkZPID0gTk87CgkJCQlTREtST09UID0gbWFjb3N4OwoJCQkJU1dJRlRfQ09NUElMQVRJT05fTU9ERSA9IHdob2xlbW9kdWxlOwoJCQkJU1dJRlRfT1BUSU1JWkFUSU9OX0xFVkVMID0gIi1PIjsKCQkJfTsKCQkJbmFtZSA9IFByb2ZpbGU7CgkJfTsKCQkzMzhEMENFQTIzMTQ1OEJEMDBGQTVGNzUgLyogUHJvZmlsZSAqLyA9IHsKCQkJaXNhID0gWENCdWlsZENvbmZpZ3VyYXRpb247CgkJCWJhc2VDb25maWd1cmF0aW9uUmVmZXJlbmNlID0gMzNFNTE5NEYyMzI4Mjg4NjAwMjZFRTREIC8qIEFwcEluZm8ueGNjb25maWcgKi87CgkJCWJ1aWxkU2V0dGluZ3MgPSB7CgkJCQlBU1NFVENBVEFMT0dfQ09NUElMRVJfQVBQSUNPTl9OQU1FID0gQXBwSWNvbjsKCQkJCUNMQU5HX0VOQUJMRV9NT0RVTEVTID0gWUVTOwoJCQkJQ09ERV9TSUdOX0VOVElUTEVNRU5UUyA9IFJ1bm5lci9EZWJ1Z1Byb2ZpbGUuZW50aXRsZW1lbnRzOwoJCQkJQ09ERV9TSUdOX1NUWUxFID0gQXV0b21hdGljOwoJCQkJQ09NQklORV9ISURQSV9JTUFHRVMgPSBZRVM7CgkJCQlJTkZPUExJU1RfRklMRSA9IFJ1bm5lci9JbmZvLnBsaXN0OwoJCQkJTERfUlVOUEFUSF9TRUFSQ0hfUEFUSFMgPSAoCgkJCQkJIiQoaW5oZXJpdGVkKSIsCgkJCQkJIkBleGVjdXRhYmxlX3BhdGgvLi4vRnJhbWV3b3JrcyIsCgkJCQkpOwoJCQkJUFJPVklTSU9OSU5HX1BST0ZJTEVfU1BFQ0lGSUVSID0gIiI7CgkJCQlTV0lGVF9WRVJTSU9OID0gNS4wOwoJCQl9OwoJCQluYW1lID0gUHJvZmlsZTsKCQl9OwoJCTMzOEQwQ0VCMjMxNDU4QkQwMEZBNUY3NSAvKiBQcm9maWxlICovID0gewoJCQlpc2EgPSBYQ0J1aWxkQ29uZmlndXJhdGlvbjsKCQkJYnVpbGRTZXR0aW5ncyA9IHsKCQkJCUNPREVfU0lHTl9TVFlMRSA9IE1hbnVhbDsKCQkJCVBST0RVQ1RfTkFNRSA9ICIkKFRBUkdFVF9OQU1FKSI7CgkJCX07CgkJCW5hbWUgPSBQcm9maWxlOwoJCX07CgkJMzNDQzEwRjkyMDQ0QTNDNjAwMDNDMDQ1IC8qIERlYnVnICovID0gewoJCQlpc2EgPSBYQ0J1aWxkQ29uZmlndXJhdGlvbjsKCQkJYmFzZUNvbmZpZ3VyYXRpb25SZWZlcmVuY2UgPSA5NzQwRUVCMjFDRjkwMTk1MDA0Mzg0RkMgLyogRGVidWcueGNjb25maWcgKi87CgkJCWJ1aWxkU2V0dGluZ3MgPSB7CgkJCQlBTFdBWVNfU0VBUkNIX1VTRVJfUEFUSFMgPSBOTzsKCQkJCUNMQU5HX0FOQUxZWkVSX05PTk5VTEwgPSBZRVM7CgkJCQlDTEFOR19BTkFMWVpFUl9OVU1CRVJfT0JKRUNUX0NPTlZFUlNJT04gPSBZRVNfQUdHUkVTU0lWRTsKCQkJCUNMQU5HX0NYWF9MQU5HVUFHRV9TVEFOREFSRCA9ICJnbnUrKzE0IjsKCQkJCUNMQU5HX0NYWF9MSUJSQVJZID0gImxpYmMrKyI7CgkJCQlDTEFOR19FTkFCTEVfTU9EVUxFUyA9IFlFUzsKCQkJCUNMQU5HX0VOQUJMRV9PQkpDX0FSQyA9IFlFUzsKCQkJCUNMQU5HX1dBUk5fQkxPQ0tfQ0FQVFVSRV9BVVRPUkVMRUFTSU5HID0gWUVTOwoJCQkJQ0xBTkdfV0FSTl9CT09MX0NPTlZFUlNJT04gPSBZRVM7CgkJCQlDTEFOR19XQVJOX0NPTlNUQU5UX0NPTlZFUlNJT04gPSBZRVM7CgkJCQlDTEFOR19XQVJOX0RFUFJFQ0FURURfT0JKQ19JTVBMRU1FTlRBVElPTlMgPSBZRVM7CgkJCQlDTEFOR19XQVJOX0RJUkVDVF9PQkpDX0lTQV9VU0FHRSA9IFlFU19FUlJPUjsKCQkJCUNMQU5HX1dBUk5fRE9DVU1FTlRBVElPTl9DT01NRU5UUyA9IFlFUzsKCQkJCUNMQU5HX1dBUk5fRU1QVFlfQk9EWSA9IFlFUzsKCQkJCUNMQU5HX1dBUk5fRU5VTV9DT05WRVJTSU9OID0gWUVTOwoJCQkJQ0xBTkdfV0FSTl9JTkZJTklURV9SRUNVUlNJT04gPSBZRVM7CgkJCQlDTEFOR19XQVJOX0lOVF9DT05WRVJTSU9OID0gWUVTOwoJCQkJQ0xBTkdfV0FSTl9OT05fTElURVJBTF9OVUxMX0NPTlZFUlNJT04gPSBZRVM7CgkJCQlDTEFOR19XQVJOX09CSkNfTElURVJBTF9DT05WRVJTSU9OID0gWUVTOwoJCQkJQ0xBTkdfV0FSTl9PQkpDX1JPT1RfQ0xBU1MgPSBZRVNfRVJST1I7CgkJCQlDTEFOR19XQVJOX1JBTkdFX0xPT1BfQU5BTFlTSVMgPSBZRVM7CgkJCQlDTEFOR19XQVJOX1NVU1BJQ0lPVVNfTU9WRSA9IFlFUzsKCQkJCUNPREVfU0lHTl9JREVOVElUWSA9ICItIjsKCQkJCUNPUFlfUEhBU0VfU1RSSVAgPSBOTzsKCQkJCURFQlVHX0lORk9STUFUSU9OX0ZPUk1BVCA9IGR3YXJmOwoJCQkJRU5BQkxFX1NUUklDVF9PQkpDX01TR1NFTkQgPSBZRVM7CgkJCQlFTkFCTEVfVEVTVEFCSUxJVFkgPSBZRVM7CgkJCQlHQ0NfQ19MQU5HVUFHRV9TVEFOREFSRCA9IGdudTExOwoJCQkJR0NDX0RZTkFNSUNfTk9fUElDID0gTk87CgkJCQlHQ0NfTk9fQ09NTU9OX0JMT0NLUyA9IFlFUzsKCQkJCUdDQ19PUFRJTUlaQVRJT05fTEVWRUwgPSAwOwoJCQkJR0NDX1BSRVBST0NFU1NPUl9ERUZJTklUSU9OUyA9ICgKCQkJCQkiREVCVUc9MSIsCgkJCQkJIiQoaW5oZXJpdGVkKSIsCgkJCQkpOwoJCQkJR0NDX1dBUk5fNjRfVE9fMzJfQklUX0NPTlZFUlNJT04gPSBZRVM7CgkJCQlHQ0NfV0FSTl9BQk9VVF9SRVRVUk5fVFlQRSA9IFlFU19FUlJPUjsKCQkJCUdDQ19XQVJOX1VOSU5JVElBTElaRURfQVVUT1MgPSBZRVNfQUdHUkVTU0lWRTsKCQkJCUdDQ19XQVJOX1VOVVNFRF9GVU5DVElPTiA9IFlFUzsKCQkJCUdDQ19XQVJOX1VOVVNFRF9WQVJJQUJMRSA9IFlFUzsKCQkJCU1BQ09TWF9ERVBMT1lNRU5UX1RBUkdFVCA9IDEwLjE1Ljc7CgkJCQlNVExfRU5BQkxFX0RFQlVHX0lORk8gPSBZRVM7CgkJCQlPTkxZX0FDVElWRV9BUkNIID0gWUVTOwoJCQkJU0RLUk9PVCA9IG1hY29zeDsKCQkJCVNXSUZUX0FDVElWRV9DT01QSUxBVElPTl9DT05ESVRJT05TID0gREVCVUc7CgkJCQlTV0lGVF9PUFRJTUlaQVRJT05fTEVWRUwgPSAiLU9ub25lIjsKCQkJfTsKCQkJbmFtZSA9IERlYnVnOwoJCX07CgkJMzNDQzEwRkEyMDQ0QTNDNjAwMDNDMDQ1IC8qIFJlbGVhc2UgKi8gPSB7CgkJCWlzYSA9IFhDQnVpbGRDb25maWd1cmF0aW9uOwoJCQliYXNlQ29uZmlndXJhdGlvblJlZmVyZW5jZSA9IDdBRkEzQzhFMUQzNTM2MEMwMDgzMDgyRSAvKiBSZWxlYXNlLnhjY29uZmlnICovOwoJCQlidWlsZFNldHRpbmdzID0gewoJCQkJQUxXQVlTX1NFQVJDSF9VU0VSX1BBVEhTID0gTk87CgkJCQlDTEFOR19BTkFMWVpFUl9OT05OVUxMID0gWUVTOwoJCQkJQ0xBTkdfQU5BTFlaRVJfTlVNQkVSX09CSkVDVF9DT05WRVJTSU9OID0gWUVTX0FHR1JFU1NJVkU7CgkJCQlDTEFOR19DWFhfTEFOR1VBR0VfU1RBTkRBUkQgPSAiZ251KysxNCI7CgkJCQlDTEFOR19DWFhfTElCUkFSWSA9ICJsaWJjKysiOwoJCQkJQ0xBTkdfRU5BQkxFX01PRFVMRVMgPSBZRVM7CgkJCQlDTEFOR19FTkFCTEVfT0JKQ19BUkMgPSBZRVM7CgkJCQlDTEFOR19XQVJOX0JMT0NLX0NBUFRVUkVfQVVUT1JFTEVBU0lORyA9IFlFUzsKCQkJCUNMQU5HX1dBUk5fQk9PTF9DT05WRVJTSU9OID0gWUVTOwoJCQkJQ0xBTkdfV0FSTl9DT05TVEFOVF9DT05WRVJTSU9OID0gWUVTOwoJCQkJQ0xBTkdfV0FSTl9ERVBSRUNBVEVEX09CSkNfSU1QTEVNRU5UQVRJT05TID0gWUVTOwoJCQkJQ0xBTkdfV0FSTl9ESVJFQ1RfT0JKQ19JU0FfVVNBR0UgPSBZRVNfRVJST1I7CgkJCQlDTEFOR19XQVJOX0RPQ1VNRU5UQVRJT05fQ09NTUVOVFMgPSBZRVM7CgkJCQlDTEFOR19XQVJOX0VNUFRZX0JPRFkgPSBZRVM7CgkJCQlDTEFOR19XQVJOX0VOVU1fQ09OVkVSU0lPTiA9IFlFUzsKCQkJCUNMQU5HX1dBUk5fSU5GSU5JVEVfUkVDVVJTSU9OID0gWUVTOwoJCQkJQ0xBTkdfV0FSTl9JTlRfQ09OVkVSU0lPTiA9IFlFUzsKCQkJCUNMQU5HX1dBUk5fTk9OX0xJVEVSQUxfTlVMTF9DT05WRVJTSU9OID0gWUVTOwoJCQkJQ0xBTkdfV0FSTl9PQkpDX0xJVEVSQUxfQ09OVkVSU0lPTiA9IFlFUzsKCQkJCUNMQU5HX1dBUk5fT0JKQ19ST09UX0NMQVNTID0gWUVTX0VSUk9SOwoJCQkJQ0xBTkdfV0FSTl9SQU5HRV9MT09QX0FOQUxZU0lTID0gWUVTOwoJCQkJQ0xBTkdfV0FSTl9TVVNQSUNJT1VTX01PVkUgPSBZRVM7CgkJCQlDT0RFX1NJR05fSURFTlRJVFkgPSAiLSI7CgkJCQlDT1BZX1BIQVNFX1NUUklQID0gTk87CgkJCQlERUJVR19JTkZPUk1BVElPTl9GT1JNQVQgPSAiZHdhcmYtd2l0aC1kc3ltIjsKCQkJCUVOQUJMRV9OU19BU1NFUlRJT05TID0gTk87CgkJCQlFTkFCTEVfU1RSSUNUX09CSkNfTVNHU0VORCA9IFlFUzsKCQkJCUdDQ19DX0xBTkdVQUdFX1NUQU5EQVJEID0gZ251MTE7CgkJCQlHQ0NfTk9fQ09NTU9OX0JMT0NLUyA9IFlFUzsKCQkJCUdDQ19XQVJOXzY0X1RPXzMyX0JJVF9DT05WRVJTSU9OID0gWUVTOwoJCQkJR0NDX1dBUk5fQUJPVVRfUkVUVVJOX1RZUEUgPSBZRVNfRVJST1I7CgkJCQlHQ0NfV0FSTl9VTklOSVRJQUxJWkVEX0FVVE9TID0gWUVTX0FHR1JFU1NJVkU7CgkJCQlHQ0NfV0FSTl9VTlVTRURfRlVOQ1RJT04gPSBZRVM7CgkJCQlHQ0NfV0FSTl9VTlVTRURfVkFSSUFCTEUgPSBZRVM7CgkJCQlNQUNPU1hfREVQTE9ZTUVOVF9UQVJHRVQgPSAxMC4xNS43OwoJCQkJTVRMX0VOQUJMRV9ERUJVR19JTkZPID0gTk87CgkJCQlTREtST09UID0gbWFjb3N4OwoJCQkJU1dJRlRfQ09NUElMQVRJT05fTU9ERSA9IHdob2xlbW9kdWxlOwoJCQkJU1dJRlRfT1BUSU1JWkFUSU9OX0xFVkVMID0gIi1PIjsKCQkJfTsKCQkJbmFtZSA9IFJlbGVhc2U7CgkJfTsKCQkzM0NDMTBGQzIwNDRBM0M2MDAwM0MwNDUgLyogRGVidWcgKi8gPSB7CgkJCWlzYSA9IFhDQnVpbGRDb25maWd1cmF0aW9uOwoJCQliYXNlQ29uZmlndXJhdGlvblJlZmVyZW5jZSA9IDMzRTUxOTRGMjMyODI4ODYwMDI2RUU0RCAvKiBBcHBJbmZvLnhjY29uZmlnICovOwoJCQlidWlsZFNldHRpbmdzID0gewoJCQkJQVNTRVRDQVRBTE9HX0NPTVBJTEVSX0FQUElDT05fTkFNRSA9IEFwcEljb247CgkJCQlDTEFOR19FTkFCTEVfTU9EVUxFUyA9IFlFUzsKCQkJCUNPREVfU0lHTl9FTlRJVExFTUVOVFMgPSBSdW5uZXIvRGVidWdQcm9maWxlLmVudGl0bGVtZW50czsKCQkJCUNPREVfU0lHTl9TVFlMRSA9IEF1dG9tYXRpYzsKCQkJCUNPTUJJTkVfSElEUElfSU1BR0VTID0gWUVTOwoJCQkJSU5GT1BMSVNUX0ZJTEUgPSBSdW5uZXIvSW5mby5wbGlzdDsKCQkJCUxEX1JVTlBBVEhfU0VBUkNIX1BBVEhTID0gKAoJCQkJCSIkKGluaGVyaXRlZCkiLAoJCQkJCSJAZXhlY3V0YWJsZV9wYXRoLy4uL0ZyYW1ld29ya3MiLAoJCQkJKTsKCQkJCVBST1ZJU0lPTklOR19QUk9GSUxFX1NQRUNJRklFUiA9ICIiOwoJCQkJU1dJRlRfT1BUSU1JWkFUSU9OX0xFVkVMID0gIi1Pbm9uZSI7CgkJCQlTV0lGVF9WRVJTSU9OID0gNS4wOwoJCQl9OwoJCQluYW1lID0gRGVidWc7CgkJfTsKCQkzM0NDMTBGRDIwNDRBM0M2MDAwM0MwNDUgLyogUmVsZWFzZSAqLyA9IHsKCQkJaXNhID0gWENCdWlsZENvbmZpZ3VyYXRpb247CgkJCWJhc2VDb25maWd1cmF0aW9uUmVmZXJlbmNlID0gMzNFNTE5NEYyMzI4Mjg4NjAwMjZFRTREIC8qIEFwcEluZm8ueGNjb25maWcgKi87CgkJCWJ1aWxkU2V0dGluZ3MgPSB7CgkJCQlBU1NFVENBVEFMT0dfQ09NUElMRVJfQVBQSUNPTl9OQU1FID0gQXBwSWNvbjsKCQkJCUNMQU5HX0VOQUJMRV9NT0RVTEVTID0gWUVTOwoJCQkJQ09ERV9TSUdOX0VOVElUTEVNRU5UUyA9IFJ1bm5lci9SZWxlYXNlLmVudGl0bGVtZW50czsKCQkJCUNPREVfU0lHTl9TVFlMRSA9IEF1dG9tYXRpYzsKCQkJCUNPTUJJTkVfSElEUElfSU1BR0VTID0gWUVTOwoJCQkJSU5GT1BMSVNUX0ZJTEUgPSBSdW5uZXIvSW5mby5wbGlzdDsKCQkJCUxEX1JVTlBBVEhfU0VBUkNIX1BBVEhTID0gKAoJCQkJCSIkKGluaGVyaXRlZCkiLAoJCQkJCSJAZXhlY3V0YWJsZV9wYXRoLy4uL0ZyYW1ld29ya3MiLAoJCQkJKTsKCQkJCVBST1ZJU0lPTklOR19QUk9GSUxFX1NQRUNJRklFUiA9ICIiOwoJCQkJU1dJRlRfVkVSU0lPTiA9IDUuMDsKCQkJfTsKCQkJbmFtZSA9IFJlbGVhc2U7CgkJfTsKCQkzM0NDMTExQzIwNDRDNkJBMDAwM0MwNDUgLyogRGVidWcgKi8gPSB7CgkJCWlzYSA9IFhDQnVpbGRDb25maWd1cmF0aW9uOwoJCQlidWlsZFNldHRpbmdzID0gewoJCQkJQ09ERV9TSUdOX1NUWUxFID0gTWFudWFsOwoJCQkJUFJPRFVDVF9OQU1FID0gIiQoVEFSR0VUX05BTUUpIjsKCQkJfTsKCQkJbmFtZSA9IERlYnVnOwoJCX07CgkJMzNDQzExMUQyMDQ0QzZCQTAwMDNDMDQ1IC8qIFJlbGVhc2UgKi8gPSB7CgkJCWlzYSA9IFhDQnVpbGRDb25maWd1cmF0aW9uOwoJCQlidWlsZFNldHRpbmdzID0gewoJCQkJQ09ERV9TSUdOX1NUWUxFID0gQXV0b21hdGljOwoJCQkJUFJPRFVDVF9OQU1FID0gIiQoVEFSR0VUX05BTUUpIjsKCQkJfTsKCQkJbmFtZSA9IFJlbGVhc2U7CgkJfTsKLyogRW5kIFhDQnVpbGRDb25maWd1cmF0aW9uIHNlY3Rpb24gKi8KCi8qIEJlZ2luIFhDQ29uZmlndXJhdGlvbkxpc3Qgc2VjdGlvbiAqLwoJCTMzMUM4MERFMjk0Q0Y3MTAwMDI2M0JFNSAvKiBCdWlsZCBjb25maWd1cmF0aW9uIGxpc3QgZm9yIFBCWE5hdGl2ZVRhcmdldCAiUnVubmVyVGVzdHMiICovID0gewoJCQlpc2EgPSBYQ0NvbmZpZ3VyYXRpb25MaXN0OwoJCQlidWlsZENvbmZpZ3VyYXRpb25zID0gKAoJCQkJMzMxQzgwREIyOTRDRjcxMDAwMjYzQkU1IC8qIERlYnVnICovLAoJCQkJMzMxQzgwREMyOTRDRjcxMDAwMjYzQkU1IC8qIFJlbGVhc2UgKi8sCgkJCQkzMzFDODBERDI5NENGNzEwMDAyNjNCRTUgLyogUHJvZmlsZSAqLywKCQkJKTsKCQkJZGVmYXVsdENvbmZpZ3VyYXRpb25Jc1Zpc2libGUgPSAwOwoJCQlkZWZhdWx0Q29uZmlndXJhdGlvbk5hbWUgPSBSZWxlYXNlOwoJCX07CgkJMzNDQzEwRTgyMDQ0QTNDNjAwMDNDMDQ1IC8qIEJ1aWxkIGNvbmZpZ3VyYXRpb24gbGlzdCBmb3IgUEJYUHJvamVjdCAiUnVubmVyIiAqLyA9IHsKCQkJaXNhID0gWENDb25maWd1cmF0aW9uTGlzdDsKCQkJYnVpbGRDb25maWd1cmF0aW9ucyA9ICgKCQkJCTMzQ0MxMEY5MjA0NEEzQzYwMDAzQzA0NSAvKiBEZWJ1ZyAqLywKCQkJCTMzQ0MxMEZBMjA0NEEzQzYwMDAzQzA0NSAvKiBSZWxlYXNlICovLAoJCQkJMzM4RDBDRTkyMzE0NThCRDAwRkE1Rjc1IC8qIFByb2ZpbGUgKi8sCgkJCSk7CgkJCWRlZmF1bHRDb25maWd1cmF0aW9uSXNWaXNpYmxlID0gMDsKCQkJZGVmYXVsdENvbmZpZ3VyYXRpb25OYW1lID0gUmVsZWFzZTsKCQl9OwoJCTMzQ0MxMEZCMjA0NEEzQzYwMDAzQzA0NSAvKiBCdWlsZCBjb25maWd1cmF0aW9uIGxpc3QgZm9yIFBCWE5hdGl2ZVRhcmdldCAiUnVubmVyIiAqLyA9IHsKCQkJaXNhID0gWENDb25maWd1cmF0aW9uTGlzdDsKCQkJYnVpbGRDb25maWd1cmF0aW9ucyA9ICgKCQkJCTMzQ0MxMEZDMjA0NEEzQzYwMDAzQzA0NSAvKiBEZWJ1ZyAqLywKCQkJCTMzQ0MxMEZEMjA0NEEzQzYwMDAzQzA0NSAvKiBSZWxlYXNlICovLAoJCQkJMzM4RDBDRUEyMzE0NThCRDAwRkE1Rjc1IC8qIFByb2ZpbGUgKi8sCgkJCSk7CgkJCWRlZmF1bHRDb25maWd1cmF0aW9uSXNWaXNpYmxlID0gMDsKCQkJZGVmYXVsdENvbmZpZ3VyYXRpb25OYW1lID0gUmVsZWFzZTsKCQl9OwoJCTMzQ0MxMTFCMjA0NEM2QkEwMDAzQzA0NSAvKiBCdWlsZCBjb25maWd1cmF0aW9uIGxpc3QgZm9yIFBCWEFnZ3JlZ2F0ZVRhcmdldCAiRmx1dHRlciBBc3NlbWJsZSIgKi8gPSB7CgkJCWlzYSA9IFhDQ29uZmlndXJhdGlvbkxpc3Q7CgkJCWJ1aWxkQ29uZmlndXJhdGlvbnMgPSAoCgkJCQkzM0NDMTExQzIwNDRDNkJBMDAwM0MwNDUgLyogRGVidWcgKi8sCgkJCQkzM0NDMTExRDIwNDRDNkJBMDAwM0MwNDUgLyogUmVsZWFzZSAqLywKCQkJCTMzOEQwQ0VCMjMxNDU4QkQwMEZBNUY3NSAvKiBQcm9maWxlICovLAoJCQkpOwoJCQlkZWZhdWx0Q29uZmlndXJhdGlvbklzVmlzaWJsZSA9IDA7CgkJCWRlZmF1bHRDb25maWd1cmF0aW9uTmFtZSA9IFJlbGVhc2U7CgkJfTsKLyogRW5kIFhDQ29uZmlndXJhdGlvbkxpc3Qgc2VjdGlvbiAqLwoJfTsKCXJvb3RPYmplY3QgPSAzM0NDMTBFNTIwNDRBM0M2MDAwM0MwNDUgLyogUHJvamVjdCBvYmplY3QgKi87Cn0K",
+          "Ly8gISQqVVRGOCokIQp7CglhcmNoaXZlVmVyc2lvbiA9IDE7CgljbGFzc2VzID0gewoJfTsKCW9iamVjdFZlcnNpb24gPSA1NDsKCW9iamVjdHMgPSB7CgovKiBCZWdpbiBQQlhBZ2dyZWdhdGVUYXJnZXQgc2VjdGlvbiAqLwoJCTMzQ0MxMTFBMjA0NEM2QkEwMDAzQzA0NSAvKiBGbHV0dGVyIEFzc2VtYmxlICovID0gewoJCQlpc2EgPSBQQlhBZ2dyZWdhdGVUYXJnZXQ7CgkJCWJ1aWxkQ29uZmlndXJhdGlvbkxpc3QgPSAzM0NDMTExQjIwNDRDNkJBMDAwM0MwNDUgLyogQnVpbGQgY29uZmlndXJhdGlvbiBsaXN0IGZvciBQQlhBZ2dyZWdhdGVUYXJnZXQgIkZsdXR0ZXIgQXNzZW1ibGUiICovOwoJCQlidWlsZFBoYXNlcyA9ICgKCQkJCTMzQ0MxMTFFMjA0NEM2QkYwMDAzQzA0NSAvKiBTaGVsbFNjcmlwdCAqLywKCQkJKTsKCQkJZGVwZW5kZW5jaWVzID0gKAoJCQkpOwoJCQluYW1lID0gIkZsdXR0ZXIgQXNzZW1ibGUiOwoJCQlwcm9kdWN0TmFtZSA9IEZMWDsKCQl9OwovKiBFbmQgUEJYQWdncmVnYXRlVGFyZ2V0IHNlY3Rpb24gKi8KCi8qIEJlZ2luIFBCWEJ1aWxkRmlsZSBzZWN0aW9uICovCgkJMzMxQzgwRDgyOTRDRjcxMDAwMjYzQkU1IC8qIFJ1bm5lclRlc3RzLnN3aWZ0IGluIFNvdXJjZXMgKi8gPSB7aXNhID0gUEJYQnVpbGRGaWxlOyBmaWxlUmVmID0gMzMxQzgwRDcyOTRDRjcxMDAwMjYzQkU1IC8qIFJ1bm5lclRlc3RzLnN3aWZ0ICovOyB9OwoJCTMzNUJCRDFCMjJBOUExNUUwMEU5MDcxRCAvKiBHZW5lcmF0ZWRQbHVnaW5SZWdpc3RyYW50LnN3aWZ0IGluIFNvdXJjZXMgKi8gPSB7aXNhID0gUEJYQnVpbGRGaWxlOyBmaWxlUmVmID0gMzM1QkJEMUEyMkE5QTE1RTAwRTkwNzFEIC8qIEdlbmVyYXRlZFBsdWdpblJlZ2lzdHJhbnQuc3dpZnQgKi87IH07CgkJMzNDQzEwRjEyMDQ0QTNDNjAwMDNDMDQ1IC8qIEFwcERlbGVnYXRlLnN3aWZ0IGluIFNvdXJjZXMgKi8gPSB7aXNhID0gUEJYQnVpbGRGaWxlOyBmaWxlUmVmID0gMzNDQzEwRjAyMDQ0QTNDNjAwMDNDMDQ1IC8qIEFwcERlbGVnYXRlLnN3aWZ0ICovOyB9OwoJCTMzQ0MxMEYzMjA0NEEzQzYwMDAzQzA0NSAvKiBBc3NldHMueGNhc3NldHMgaW4gUmVzb3VyY2VzICovID0ge2lzYSA9IFBCWEJ1aWxkRmlsZTsgZmlsZVJlZiA9IDMzQ0MxMEYyMjA0NEEzQzYwMDAzQzA0NSAvKiBBc3NldHMueGNhc3NldHMgKi87IH07CgkJMzNDQzEwRjYyMDQ0QTNDNjAwMDNDMDQ1IC8qIE1haW5NZW51LnhpYiBpbiBSZXNvdXJjZXMgKi8gPSB7aXNhID0gUEJYQnVpbGRGaWxlOyBmaWxlUmVmID0gMzNDQzEwRjQyMDQ0QTNDNjAwMDNDMDQ1IC8qIE1haW5NZW51LnhpYiAqLzsgfTsKCQkzM0NDMTExMzIwNDRCRkEwMDAwM0MwNDUgLyogTWFpbkZsdXR0ZXJXaW5kb3cuc3dpZnQgaW4gU291cmNlcyAqLyA9IHtpc2EgPSBQQlhCdWlsZEZpbGU7IGZpbGVSZWYgPSAzM0NDMTExMjIwNDRCRkEwMDAwM0MwNDUgLyogTWFpbkZsdXR0ZXJXaW5kb3cuc3dpZnQgKi87IH07Ci8qIEVuZCBQQlhCdWlsZEZpbGUgc2VjdGlvbiAqLwoKLyogQmVnaW4gUEJYQ29udGFpbmVySXRlbVByb3h5IHNlY3Rpb24gKi8KCQkzMzFDODBEOTI5NENGNzEwMDAyNjNCRTUgLyogUEJYQ29udGFpbmVySXRlbVByb3h5ICovID0gewoJCQlpc2EgPSBQQlhDb250YWluZXJJdGVtUHJveHk7CgkJCWNvbnRhaW5lclBvcnRhbCA9IDMzQ0MxMEU1MjA0NEEzQzYwMDAzQzA0NSAvKiBQcm9qZWN0IG9iamVjdCAqLzsKCQkJcHJveHlUeXBlID0gMTsKCQkJcmVtb3RlR2xvYmFsSURTdHJpbmcgPSAzM0NDMTBFQzIwNDRBM0M2MDAwM0MwNDU7CgkJCXJlbW90ZUluZm8gPSBSdW5uZXI7CgkJfTsKCQkzM0NDMTExRjIwNDRDNzlGMDAwM0MwNDUgLyogUEJYQ29udGFpbmVySXRlbVByb3h5ICovID0gewoJCQlpc2EgPSBQQlhDb250YWluZXJJdGVtUHJveHk7CgkJCWNvbnRhaW5lclBvcnRhbCA9IDMzQ0MxMEU1MjA0NEEzQzYwMDAzQzA0NSAvKiBQcm9qZWN0IG9iamVjdCAqLzsKCQkJcHJveHlUeXBlID0gMTsKCQkJcmVtb3RlR2xvYmFsSURTdHJpbmcgPSAzM0NDMTExQTIwNDRDNkJBMDAwM0MwNDU7CgkJCXJlbW90ZUluZm8gPSBGTFg7CgkJfTsKLyogRW5kIFBCWENvbnRhaW5lckl0ZW1Qcm94eSBzZWN0aW9uICovCgovKiBCZWdpbiBQQlhDb3B5RmlsZXNCdWlsZFBoYXNlIHNlY3Rpb24gKi8KCQkzM0NDMTEwRTIwNDRBODg0MDAwM0MwNDUgLyogQnVuZGxlIEZyYW1ld29yayAqLyA9IHsKCQkJaXNhID0gUEJYQ29weUZpbGVzQnVpbGRQaGFzZTsKCQkJYnVpbGRBY3Rpb25NYXNrID0gMjE0NzQ4MzY0NzsKCQkJZHN0UGF0aCA9ICIiOwoJCQlkc3RTdWJmb2xkZXJTcGVjID0gMTA7CgkJCWZpbGVzID0gKAoJCQkpOwoJCQluYW1lID0gIkJ1bmRsZSBGcmFtZXdvcmsiOwoJCQlydW5Pbmx5Rm9yRGVwbG95bWVudFBvc3Rwcm9jZXNzaW5nID0gMDsKCQl9OwovKiBFbmQgUEJYQ29weUZpbGVzQnVpbGRQaGFzZSBzZWN0aW9uICovCgovKiBCZWdpbiBQQlhGaWxlUmVmZXJlbmNlIHNlY3Rpb24gKi8KCQkzMzFDODBENTI5NENGNzEwMDAyNjNCRTUgLyogUnVubmVyVGVzdHMueGN0ZXN0ICovID0ge2lzYSA9IFBCWEZpbGVSZWZlcmVuY2U7IGV4cGxpY2l0RmlsZVR5cGUgPSB3cmFwcGVyLmNmYnVuZGxlOyBpbmNsdWRlSW5JbmRleCA9IDA7IHBhdGggPSBSdW5uZXJUZXN0cy54Y3Rlc3Q7IHNvdXJjZVRyZWUgPSBCVUlMVF9QUk9EVUNUU19ESVI7IH07CgkJMzMxQzgwRDcyOTRDRjcxMDAwMjYzQkU1IC8qIFJ1bm5lclRlc3RzLnN3aWZ0ICovID0ge2lzYSA9IFBCWEZpbGVSZWZlcmVuY2U7IGxhc3RLbm93bkZpbGVUeXBlID0gc291cmNlY29kZS5zd2lmdDsgcGF0aCA9IFJ1bm5lclRlc3RzLnN3aWZ0OyBzb3VyY2VUcmVlID0gIjxncm91cD4iOyB9OwoJCTMzMzAwMEVEMjJEM0RFNUQwMDU1NDE2MiAvKiBXYXJuaW5ncy54Y2NvbmZpZyAqLyA9IHtpc2EgPSBQQlhGaWxlUmVmZXJlbmNlOyBsYXN0S25vd25GaWxlVHlwZSA9IHRleHQueGNjb25maWc7IHBhdGggPSBXYXJuaW5ncy54Y2NvbmZpZzsgc291cmNlVHJlZSA9ICI8Z3JvdXA+IjsgfTsKCQkzMzVCQkQxQTIyQTlBMTVFMDBFOTA3MUQgLyogR2VuZXJhdGVkUGx1Z2luUmVnaXN0cmFudC5zd2lmdCAqLyA9IHtpc2EgPSBQQlhGaWxlUmVmZXJlbmNlOyBmaWxlRW5jb2RpbmcgPSA0OyBsYXN0S25vd25GaWxlVHlwZSA9IHNvdXJjZWNvZGUuc3dpZnQ7IHBhdGggPSBHZW5lcmF0ZWRQbHVnaW5SZWdpc3RyYW50LnN3aWZ0OyBzb3VyY2VUcmVlID0gIjxncm91cD4iOyB9OwoJCTMzQ0MxMEVEMjA0NEEzQzYwMDAzQzA0NSAvKiB7e3Byb2plY3RfbmFtZX19LmFwcCAqLyA9IHtpc2EgPSBQQlhGaWxlUmVmZXJlbmNlOyBleHBsaWNpdEZpbGVUeXBlID0gd3JhcHBlci5hcHBsaWNhdGlvbjsgaW5jbHVkZUluSW5kZXggPSAwOyBwYXRoID0gInt7cHJvamVjdF9uYW1lfX0uYXBwIjsgc291cmNlVHJlZSA9IEJVSUxUX1BST0RVQ1RTX0RJUjsgfTsKCQkzM0NDMTBGMDIwNDRBM0M2MDAwM0MwNDUgLyogQXBwRGVsZWdhdGUuc3dpZnQgKi8gPSB7aXNhID0gUEJYRmlsZVJlZmVyZW5jZTsgbGFzdEtub3duRmlsZVR5cGUgPSBzb3VyY2Vjb2RlLnN3aWZ0OyBwYXRoID0gQXBwRGVsZWdhdGUuc3dpZnQ7IHNvdXJjZVRyZWUgPSAiPGdyb3VwPiI7IH07CgkJMzNDQzEwRjIyMDQ0QTNDNjAwMDNDMDQ1IC8qIEFzc2V0cy54Y2Fzc2V0cyAqLyA9IHtpc2EgPSBQQlhGaWxlUmVmZXJlbmNlOyBsYXN0S25vd25GaWxlVHlwZSA9IGZvbGRlci5hc3NldGNhdGFsb2c7IG5hbWUgPSBBc3NldHMueGNhc3NldHM7IHBhdGggPSBSdW5uZXIvQXNzZXRzLnhjYXNzZXRzOyBzb3VyY2VUcmVlID0gIjxncm91cD4iOyB9OwoJCTMzQ0MxMEY1MjA0NEEzQzYwMDAzQzA0NSAvKiBCYXNlICovID0ge2lzYSA9IFBCWEZpbGVSZWZlcmVuY2U7IGxhc3RLbm93bkZpbGVUeXBlID0gZmlsZS54aWI7IG5hbWUgPSBCYXNlOyBwYXRoID0gQmFzZS5scHJvai9NYWluTWVudS54aWI7IHNvdXJjZVRyZWUgPSAiPGdyb3VwPiI7IH07CgkJMzNDQzEwRjcyMDQ0QTNDNjAwMDNDMDQ1IC8qIEluZm8ucGxpc3QgKi8gPSB7aXNhID0gUEJYRmlsZVJlZmVyZW5jZTsgbGFzdEtub3duRmlsZVR5cGUgPSB0ZXh0LnBsaXN0LnhtbDsgbmFtZSA9IEluZm8ucGxpc3Q7IHBhdGggPSBSdW5uZXIvSW5mby5wbGlzdDsgc291cmNlVHJlZSA9ICI8Z3JvdXA+IjsgfTsKCQkzM0NDMTExMjIwNDRCRkEwMDAwM0MwNDUgLyogTWFpbkZsdXR0ZXJXaW5kb3cuc3dpZnQgKi8gPSB7aXNhID0gUEJYRmlsZVJlZmVyZW5jZTsgbGFzdEtub3duRmlsZVR5cGUgPSBzb3VyY2Vjb2RlLnN3aWZ0OyBwYXRoID0gTWFpbkZsdXR0ZXJXaW5kb3cuc3dpZnQ7IHNvdXJjZVRyZWUgPSAiPGdyb3VwPiI7IH07CgkJMzNDRUI0NzIyMkEwNTc3MTAwNEYyQUMwIC8qIEZsdXR0ZXItRGVidWcueGNjb25maWcgKi8gPSB7aXNhID0gUEJYRmlsZVJlZmVyZW5jZTsgbGFzdEtub3duRmlsZVR5cGUgPSB0ZXh0LnhjY29uZmlnOyBwYXRoID0gIkZsdXR0ZXItRGVidWcueGNjb25maWciOyBzb3VyY2VUcmVlID0gIjxncm91cD4iOyB9OwoJCTMzQ0VCNDc0MjJBMDU3NzEwMDRGMkFDMCAvKiBGbHV0dGVyLVJlbGVhc2UueGNjb25maWcgKi8gPSB7aXNhID0gUEJYRmlsZVJlZmVyZW5jZTsgbGFzdEtub3duRmlsZVR5cGUgPSB0ZXh0LnhjY29uZmlnOyBwYXRoID0gIkZsdXR0ZXItUmVsZWFzZS54Y2NvbmZpZyI7IHNvdXJjZVRyZWUgPSAiPGdyb3VwPiI7IH07CgkJMzNDRUI0NzcyMkEwNTc4QTAwNEYyQUMwIC8qIEZsdXR0ZXItR2VuZXJhdGVkLnhjY29uZmlnICovID0ge2lzYSA9IFBCWEZpbGVSZWZlcmVuY2U7IGxhc3RLbm93bkZpbGVUeXBlID0gdGV4dC54Y2NvbmZpZzsgbmFtZSA9ICJGbHV0dGVyLUdlbmVyYXRlZC54Y2NvbmZpZyI7IHBhdGggPSAiZXBoZW1lcmFsL0ZsdXR0ZXItR2VuZXJhdGVkLnhjY29uZmlnIjsgc291cmNlVHJlZSA9ICI8Z3JvdXA+IjsgfTsKCQkzM0U1MTkxMzIzMTc0N0Y0MDAyNkVFNEQgLyogRGVidWdQcm9maWxlLmVudGl0bGVtZW50cyAqLyA9IHtpc2EgPSBQQlhGaWxlUmVmZXJlbmNlOyBsYXN0S25vd25GaWxlVHlwZSA9IHRleHQucGxpc3QuZW50aXRsZW1lbnRzOyBwYXRoID0gRGVidWdQcm9maWxlLmVudGl0bGVtZW50czsgc291cmNlVHJlZSA9ICI8Z3JvdXA+IjsgfTsKCQkzM0U1MTkxNDIzMTc0OTM4MDAyNkVFNEQgLyogUmVsZWFzZS5lbnRpdGxlbWVudHMgKi8gPSB7aXNhID0gUEJYRmlsZVJlZmVyZW5jZTsgZmlsZUVuY29kaW5nID0gNDsgbGFzdEtub3duRmlsZVR5cGUgPSB0ZXh0LnBsaXN0LmVudGl0bGVtZW50czsgcGF0aCA9IFJlbGVhc2UuZW50aXRsZW1lbnRzOyBzb3VyY2VUcmVlID0gIjxncm91cD4iOyB9OwoJCTMzRTUxOTRGMjMyODI4ODYwMDI2RUU0RCAvKiBBcHBJbmZvLnhjY29uZmlnICovID0ge2lzYSA9IFBCWEZpbGVSZWZlcmVuY2U7IGxhc3RLbm93bkZpbGVUeXBlID0gdGV4dC54Y2NvbmZpZzsgcGF0aCA9IEFwcEluZm8ueGNjb25maWc7IHNvdXJjZVRyZWUgPSAiPGdyb3VwPiI7IH07CgkJN0FGQTNDOEUxRDM1MzYwQzAwODMwODJFIC8qIFJlbGVhc2UueGNjb25maWcgKi8gPSB7aXNhID0gUEJYRmlsZVJlZmVyZW5jZTsgbGFzdEtub3duRmlsZVR5cGUgPSB0ZXh0LnhjY29uZmlnOyBwYXRoID0gUmVsZWFzZS54Y2NvbmZpZzsgc291cmNlVHJlZSA9ICI8Z3JvdXA+IjsgfTsKCQk5NzQwRUVCMjFDRjkwMTk1MDA0Mzg0RkMgLyogRGVidWcueGNjb25maWcgKi8gPSB7aXNhID0gUEJYRmlsZVJlZmVyZW5jZTsgZmlsZUVuY29kaW5nID0gNDsgbGFzdEtub3duRmlsZVR5cGUgPSB0ZXh0LnhjY29uZmlnOyBwYXRoID0gRGVidWcueGNjb25maWc7IHNvdXJjZVRyZWUgPSAiPGdyb3VwPiI7IH07Ci8qIEVuZCBQQlhGaWxlUmVmZXJlbmNlIHNlY3Rpb24gKi8KCi8qIEJlZ2luIFBCWEZyYW1ld29ya3NCdWlsZFBoYXNlIHNlY3Rpb24gKi8KCQkzMzFDODBEMjI5NENGNzBGMDAyNjNCRTUgLyogRnJhbWV3b3JrcyAqLyA9IHsKCQkJaXNhID0gUEJYRnJhbWV3b3Jrc0J1aWxkUGhhc2U7CgkJCWJ1aWxkQWN0aW9uTWFzayA9IDIxNDc0ODM2NDc7CgkJCWZpbGVzID0gKAoJCQkpOwoJCQlydW5Pbmx5Rm9yRGVwbG95bWVudFBvc3Rwcm9jZXNzaW5nID0gMDsKCQl9OwoJCTMzQ0MxMEVBMjA0NEEzQzYwMDAzQzA0NSAvKiBGcmFtZXdvcmtzICovID0gewoJCQlpc2EgPSBQQlhGcmFtZXdvcmtzQnVpbGRQaGFzZTsKCQkJYnVpbGRBY3Rpb25NYXNrID0gMjE0NzQ4MzY0NzsKCQkJZmlsZXMgPSAoCgkJCSk7CgkJCXJ1bk9ubHlGb3JEZXBsb3ltZW50UG9zdHByb2Nlc3NpbmcgPSAwOwoJCX07Ci8qIEVuZCBQQlhGcmFtZXdvcmtzQnVpbGRQaGFzZSBzZWN0aW9uICovCgovKiBCZWdpbiBQQlhHcm91cCBzZWN0aW9uICovCgkJMzMxQzgwRDYyOTRDRjcxMDAwMjYzQkU1IC8qIFJ1bm5lclRlc3RzICovID0gewoJCQlpc2EgPSBQQlhHcm91cDsKCQkJY2hpbGRyZW4gPSAoCgkJCQkzMzFDODBENzI5NENGNzEwMDAyNjNCRTUgLyogUnVubmVyVGVzdHMuc3dpZnQgKi8sCgkJCSk7CgkJCXBhdGggPSBSdW5uZXJUZXN0czsKCQkJc291cmNlVHJlZSA9ICI8Z3JvdXA+IjsKCQl9OwoJCTMzQkE4ODZBMjI2RTc4QUYwMDMzMjlENSAvKiBDb25maWdzICovID0gewoJCQlpc2EgPSBQQlhHcm91cDsKCQkJY2hpbGRyZW4gPSAoCgkJCQkzM0U1MTk0RjIzMjgyODg2MDAyNkVFNEQgLyogQXBwSW5mby54Y2NvbmZpZyAqLywKCQkJCTk3NDBFRUIyMUNGOTAxOTUwMDQzODRGQyAvKiBEZWJ1Zy54Y2NvbmZpZyAqLywKCQkJCTdBRkEzQzhFMUQzNTM2MEMwMDgzMDgyRSAvKiBSZWxlYXNlLnhjY29uZmlnICovLAoJCQkJMzMzMDAwRUQyMkQzREU1RDAwNTU0MTYyIC8qIFdhcm5pbmdzLnhjY29uZmlnICovLAoJCQkpOwoJCQlwYXRoID0gQ29uZmlnczsKCQkJc291cmNlVHJlZSA9ICI8Z3JvdXA+IjsKCQl9OwoJCTMzQ0MxMEU0MjA0NEEzQzYwMDAzQzA0NSA9IHsKCQkJaXNhID0gUEJYR3JvdXA7CgkJCWNoaWxkcmVuID0gKAoJCQkJMzNGQUI2NzEyMzI4MzY3NDAwNjVBQzFFIC8qIFJ1bm5lciAqLywKCQkJCTMzQ0VCNDcxMjJBMDU3NzEwMDRGMkFDMCAvKiBGbHV0dGVyICovLAoJCQkJMzMxQzgwRDYyOTRDRjcxMDAwMjYzQkU1IC8qIFJ1bm5lclRlc3RzICovLAoJCQkJMzNDQzEwRUUyMDQ0QTNDNjAwMDNDMDQ1IC8qIFByb2R1Y3RzICovLAoJCQkJRDczOTEyRUMyMkYzN0YzRDAwMEQxM0EwIC8qIEZyYW1ld29ya3MgKi8sCgkJCSk7CgkJCXNvdXJjZVRyZWUgPSAiPGdyb3VwPiI7CgkJfTsKCQkzM0NDMTBFRTIwNDRBM0M2MDAwM0MwNDUgLyogUHJvZHVjdHMgKi8gPSB7CgkJCWlzYSA9IFBCWEdyb3VwOwoJCQljaGlsZHJlbiA9ICgKCQkJCTMzQ0MxMEVEMjA0NEEzQzYwMDAzQzA0NSAvKiB7e3Byb2plY3RfbmFtZX19LmFwcCAqLywKCQkJCTMzMUM4MEQ1Mjk0Q0Y3MTAwMDI2M0JFNSAvKiBSdW5uZXJUZXN0cy54Y3Rlc3QgKi8sCgkJCSk7CgkJCW5hbWUgPSBQcm9kdWN0czsKCQkJc291cmNlVHJlZSA9ICI8Z3JvdXA+IjsKCQl9OwoJCTMzQ0MxMTI0MjA0NEQ2NkUwMDAzQzA0NSAvKiBSZXNvdXJjZXMgKi8gPSB7CgkJCWlzYSA9IFBCWEdyb3VwOwoJCQljaGlsZHJlbiA9ICgKCQkJCTMzQ0MxMEYyMjA0NEEzQzYwMDAzQzA0NSAvKiBBc3NldHMueGNhc3NldHMgKi8sCgkJCQkzM0NDMTBGNDIwNDRBM0M2MDAwM0MwNDUgLyogTWFpbk1lbnUueGliICovLAoJCQkJMzNDQzEwRjcyMDQ0QTNDNjAwMDNDMDQ1IC8qIEluZm8ucGxpc3QgKi8sCgkJCSk7CgkJCW5hbWUgPSBSZXNvdXJjZXM7CgkJCXBhdGggPSAuLjsKCQkJc291cmNlVHJlZSA9ICI8Z3JvdXA+IjsKCQl9OwoJCTMzQ0VCNDcxMjJBMDU3NzEwMDRGMkFDMCAvKiBGbHV0dGVyICovID0gewoJCQlpc2EgPSBQQlhHcm91cDsKCQkJY2hpbGRyZW4gPSAoCgkJCQkzMzVCQkQxQTIyQTlBMTVFMDBFOTA3MUQgLyogR2VuZXJhdGVkUGx1Z2luUmVnaXN0cmFudC5zd2lmdCAqLywKCQkJCTMzQ0VCNDcyMjJBMDU3NzEwMDRGMkFDMCAvKiBGbHV0dGVyLURlYnVnLnhjY29uZmlnICovLAoJCQkJMzNDRUI0NzQyMkEwNTc3MTAwNEYyQUMwIC8qIEZsdXR0ZXItUmVsZWFzZS54Y2NvbmZpZyAqLywKCQkJCTMzQ0VCNDc3MjJBMDU3OEEwMDRGMkFDMCAvKiBGbHV0dGVyLUdlbmVyYXRlZC54Y2NvbmZpZyAqLywKCQkJKTsKCQkJcGF0aCA9IEZsdXR0ZXI7CgkJCXNvdXJjZVRyZWUgPSAiPGdyb3VwPiI7CgkJfTsKCQkzM0ZBQjY3MTIzMjgzNjc0MDA2NUFDMUUgLyogUnVubmVyICovID0gewoJCQlpc2EgPSBQQlhHcm91cDsKCQkJY2hpbGRyZW4gPSAoCgkJCQkzM0NDMTBGMDIwNDRBM0M2MDAwM0MwNDUgLyogQXBwRGVsZWdhdGUuc3dpZnQgKi8sCgkJCQkzM0NDMTExMjIwNDRCRkEwMDAwM0MwNDUgLyogTWFpbkZsdXR0ZXJXaW5kb3cuc3dpZnQgKi8sCgkJCQkzM0U1MTkxMzIzMTc0N0Y0MDAyNkVFNEQgLyogRGVidWdQcm9maWxlLmVudGl0bGVtZW50cyAqLywKCQkJCTMzRTUxOTE0MjMxNzQ5MzgwMDI2RUU0RCAvKiBSZWxlYXNlLmVudGl0bGVtZW50cyAqLywKCQkJCTMzQ0MxMTI0MjA0NEQ2NkUwMDAzQzA0NSAvKiBSZXNvdXJjZXMgKi8sCgkJCQkzM0JBODg2QTIyNkU3OEFGMDAzMzI5RDUgLyogQ29uZmlncyAqLywKCQkJKTsKCQkJcGF0aCA9IFJ1bm5lcjsKCQkJc291cmNlVHJlZSA9ICI8Z3JvdXA+IjsKCQl9OwoJCUQ3MzkxMkVDMjJGMzdGM0QwMDBEMTNBMCAvKiBGcmFtZXdvcmtzICovID0gewoJCQlpc2EgPSBQQlhHcm91cDsKCQkJY2hpbGRyZW4gPSAoCgkJCSk7CgkJCW5hbWUgPSBGcmFtZXdvcmtzOwoJCQlzb3VyY2VUcmVlID0gIjxncm91cD4iOwoJCX07Ci8qIEVuZCBQQlhHcm91cCBzZWN0aW9uICovCgovKiBCZWdpbiBQQlhOYXRpdmVUYXJnZXQgc2VjdGlvbiAqLwoJCTMzMUM4MEQ0Mjk0Q0Y3MEYwMDI2M0JFNSAvKiBSdW5uZXJUZXN0cyAqLyA9IHsKCQkJaXNhID0gUEJYTmF0aXZlVGFyZ2V0OwoJCQlidWlsZENvbmZpZ3VyYXRpb25MaXN0ID0gMzMxQzgwREUyOTRDRjcxMDAwMjYzQkU1IC8qIEJ1aWxkIGNvbmZpZ3VyYXRpb24gbGlzdCBmb3IgUEJYTmF0aXZlVGFyZ2V0ICJSdW5uZXJUZXN0cyIgKi87CgkJCWJ1aWxkUGhhc2VzID0gKAoJCQkJMzMxQzgwRDEyOTRDRjcwRjAwMjYzQkU1IC8qIFNvdXJjZXMgKi8sCgkJCQkzMzFDODBEMjI5NENGNzBGMDAyNjNCRTUgLyogRnJhbWV3b3JrcyAqLywKCQkJCTMzMUM4MEQzMjk0Q0Y3MEYwMDI2M0JFNSAvKiBSZXNvdXJjZXMgKi8sCgkJCSk7CgkJCWJ1aWxkUnVsZXMgPSAoCgkJCSk7CgkJCWRlcGVuZGVuY2llcyA9ICgKCQkJCTMzMUM4MERBMjk0Q0Y3MTAwMDI2M0JFNSAvKiBQQlhUYXJnZXREZXBlbmRlbmN5ICovLAoJCQkpOwoJCQluYW1lID0gUnVubmVyVGVzdHM7CgkJCXByb2R1Y3ROYW1lID0gUnVubmVyVGVzdHM7CgkJCXByb2R1Y3RSZWZlcmVuY2UgPSAzMzFDODBENTI5NENGNzEwMDAyNjNCRTUgLyogUnVubmVyVGVzdHMueGN0ZXN0ICovOwoJCQlwcm9kdWN0VHlwZSA9ICJjb20uYXBwbGUucHJvZHVjdC10eXBlLmJ1bmRsZS51bml0LXRlc3QiOwoJCX07CgkJMzNDQzEwRUMyMDQ0QTNDNjAwMDNDMDQ1IC8qIFJ1bm5lciAqLyA9IHsKCQkJaXNhID0gUEJYTmF0aXZlVGFyZ2V0OwoJCQlidWlsZENvbmZpZ3VyYXRpb25MaXN0ID0gMzNDQzEwRkIyMDQ0QTNDNjAwMDNDMDQ1IC8qIEJ1aWxkIGNvbmZpZ3VyYXRpb24gbGlzdCBmb3IgUEJYTmF0aXZlVGFyZ2V0ICJSdW5uZXIiICovOwoJCQlidWlsZFBoYXNlcyA9ICgKCQkJCTMzQ0MxMEU5MjA0NEEzQzYwMDAzQzA0NSAvKiBTb3VyY2VzICovLAoJCQkJMzNDQzEwRUEyMDQ0QTNDNjAwMDNDMDQ1IC8qIEZyYW1ld29ya3MgKi8sCgkJCQkzM0NDMTBFQjIwNDRBM0M2MDAwM0MwNDUgLyogUmVzb3VyY2VzICovLAoJCQkJMzNDQzExMEUyMDQ0QTg4NDAwMDNDMDQ1IC8qIEJ1bmRsZSBGcmFtZXdvcmsgKi8sCgkJCQkzMzk5RDQ5MDIyOEIyNENGMDA5QTc5QzcgLyogU2hlbGxTY3JpcHQgKi8sCgkJCSk7CgkJCWJ1aWxkUnVsZXMgPSAoCgkJCSk7CgkJCWRlcGVuZGVuY2llcyA9ICgKCQkJCTMzQ0MxMTIwMjA0NEM3OUYwMDAzQzA0NSAvKiBQQlhUYXJnZXREZXBlbmRlbmN5ICovLAoJCQkpOwoJCQluYW1lID0gUnVubmVyOwoJCQlwcm9kdWN0TmFtZSA9IFJ1bm5lcjsKCQkJcHJvZHVjdFJlZmVyZW5jZSA9IDMzQ0MxMEVEMjA0NEEzQzYwMDAzQzA0NSAvKiB7e3Byb2plY3RfbmFtZX19LmFwcCAqLzsKCQkJcHJvZHVjdFR5cGUgPSAiY29tLmFwcGxlLnByb2R1Y3QtdHlwZS5hcHBsaWNhdGlvbiI7CgkJfTsKLyogRW5kIFBCWE5hdGl2ZVRhcmdldCBzZWN0aW9uICovCgovKiBCZWdpbiBQQlhQcm9qZWN0IHNlY3Rpb24gKi8KCQkzM0NDMTBFNTIwNDRBM0M2MDAwM0MwNDUgLyogUHJvamVjdCBvYmplY3QgKi8gPSB7CgkJCWlzYSA9IFBCWFByb2plY3Q7CgkJCWF0dHJpYnV0ZXMgPSB7CgkJCQlMYXN0U3dpZnRVcGRhdGVDaGVjayA9IDA5MjA7CgkJCQlMYXN0VXBncmFkZUNoZWNrID0gMTMwMDsKCQkJCU9SR0FOSVpBVElPTk5BTUUgPSAiIjsKCQkJCVRhcmdldEF0dHJpYnV0ZXMgPSB7CgkJCQkJMzMxQzgwRDQyOTRDRjcwRjAwMjYzQkU1ID0gewoJCQkJCQlDcmVhdGVkT25Ub29sc1ZlcnNpb24gPSAxNC4wOwoJCQkJCQlUZXN0VGFyZ2V0SUQgPSAzM0NDMTBFQzIwNDRBM0M2MDAwM0MwNDU7CgkJCQkJfTsKCQkJCQkzM0NDMTBFQzIwNDRBM0M2MDAwM0MwNDUgPSB7CgkJCQkJCUNyZWF0ZWRPblRvb2xzVmVyc2lvbiA9IDkuMjsKCQkJCQkJTGFzdFN3aWZ0TWlncmF0aW9uID0gMTEwMDsKCQkJCQkJUHJvdmlzaW9uaW5nU3R5bGUgPSBBdXRvbWF0aWM7CgkJCQkJCVN5c3RlbUNhcGFiaWxpdGllcyA9IHsKCQkJCQkJCWNvbS5hcHBsZS5TYW5kYm94ID0gewoJCQkJCQkJCWVuYWJsZWQgPSAxOwoJCQkJCQkJfTsKCQkJCQkJfTsKCQkJCQl9OwoJCQkJCTMzQ0MxMTFBMjA0NEM2QkEwMDAzQzA0NSA9IHsKCQkJCQkJQ3JlYXRlZE9uVG9vbHNWZXJzaW9uID0gOS4yOwoJCQkJCQlQcm92aXNpb25pbmdTdHlsZSA9IE1hbnVhbDsKCQkJCQl9OwoJCQkJfTsKCQkJfTsKCQkJYnVpbGRDb25maWd1cmF0aW9uTGlzdCA9IDMzQ0MxMEU4MjA0NEEzQzYwMDAzQzA0NSAvKiBCdWlsZCBjb25maWd1cmF0aW9uIGxpc3QgZm9yIFBCWFByb2plY3QgIlJ1bm5lciIgKi87CgkJCWNvbXBhdGliaWxpdHlWZXJzaW9uID0gIlhjb2RlIDkuMyI7CgkJCWRldmVsb3BtZW50UmVnaW9uID0gZW47CgkJCWhhc1NjYW5uZWRGb3JFbmNvZGluZ3MgPSAwOwoJCQlrbm93blJlZ2lvbnMgPSAoCgkJCQllbiwKCQkJCUJhc2UsCgkJCSk7CgkJCW1haW5Hcm91cCA9IDMzQ0MxMEU0MjA0NEEzQzYwMDAzQzA0NTsKCQkJcHJvZHVjdFJlZkdyb3VwID0gMzNDQzEwRUUyMDQ0QTNDNjAwMDNDMDQ1IC8qIFByb2R1Y3RzICovOwoJCQlwcm9qZWN0RGlyUGF0aCA9ICIiOwoJCQlwcm9qZWN0Um9vdCA9ICIiOwoJCQl0YXJnZXRzID0gKAoJCQkJMzNDQzEwRUMyMDQ0QTNDNjAwMDNDMDQ1IC8qIFJ1bm5lciAqLywKCQkJCTMzMUM4MEQ0Mjk0Q0Y3MEYwMDI2M0JFNSAvKiBSdW5uZXJUZXN0cyAqLywKCQkJCTMzQ0MxMTFBMjA0NEM2QkEwMDAzQzA0NSAvKiBGbHV0dGVyIEFzc2VtYmxlICovLAoJCQkpOwoJCX07Ci8qIEVuZCBQQlhQcm9qZWN0IHNlY3Rpb24gKi8KCi8qIEJlZ2luIFBCWFJlc291cmNlc0J1aWxkUGhhc2Ugc2VjdGlvbiAqLwoJCTMzMUM4MEQzMjk0Q0Y3MEYwMDI2M0JFNSAvKiBSZXNvdXJjZXMgKi8gPSB7CgkJCWlzYSA9IFBCWFJlc291cmNlc0J1aWxkUGhhc2U7CgkJCWJ1aWxkQWN0aW9uTWFzayA9IDIxNDc0ODM2NDc7CgkJCWZpbGVzID0gKAoJCQkpOwoJCQlydW5Pbmx5Rm9yRGVwbG95bWVudFBvc3Rwcm9jZXNzaW5nID0gMDsKCQl9OwoJCTMzQ0MxMEVCMjA0NEEzQzYwMDAzQzA0NSAvKiBSZXNvdXJjZXMgKi8gPSB7CgkJCWlzYSA9IFBCWFJlc291cmNlc0J1aWxkUGhhc2U7CgkJCWJ1aWxkQWN0aW9uTWFzayA9IDIxNDc0ODM2NDc7CgkJCWZpbGVzID0gKAoJCQkJMzNDQzEwRjMyMDQ0QTNDNjAwMDNDMDQ1IC8qIEFzc2V0cy54Y2Fzc2V0cyBpbiBSZXNvdXJjZXMgKi8sCgkJCQkzM0NDMTBGNjIwNDRBM0M2MDAwM0MwNDUgLyogTWFpbk1lbnUueGliIGluIFJlc291cmNlcyAqLywKCQkJKTsKCQkJcnVuT25seUZvckRlcGxveW1lbnRQb3N0cHJvY2Vzc2luZyA9IDA7CgkJfTsKLyogRW5kIFBCWFJlc291cmNlc0J1aWxkUGhhc2Ugc2VjdGlvbiAqLwoKLyogQmVnaW4gUEJYU2hlbGxTY3JpcHRCdWlsZFBoYXNlIHNlY3Rpb24gKi8KCQkzMzk5RDQ5MDIyOEIyNENGMDA5QTc5QzcgLyogU2hlbGxTY3JpcHQgKi8gPSB7CgkJCWlzYSA9IFBCWFNoZWxsU2NyaXB0QnVpbGRQaGFzZTsKCQkJYWx3YXlzT3V0T2ZEYXRlID0gMTsKCQkJYnVpbGRBY3Rpb25NYXNrID0gMjE0NzQ4MzY0NzsKCQkJZmlsZXMgPSAoCgkJCSk7CgkJCWlucHV0RmlsZUxpc3RQYXRocyA9ICgKCQkJKTsKCQkJaW5wdXRQYXRocyA9ICgKCQkJKTsKCQkJb3V0cHV0RmlsZUxpc3RQYXRocyA9ICgKCQkJKTsKCQkJb3V0cHV0UGF0aHMgPSAoCgkJCSk7CgkJCXJ1bk9ubHlGb3JEZXBsb3ltZW50UG9zdHByb2Nlc3NpbmcgPSAwOwoJCQlzaGVsbFBhdGggPSAvYmluL3NoOwoJCQlzaGVsbFNjcmlwdCA9ICJlY2hvIFwiJFBST0RVQ1RfTkFNRS5hcHBcIiA+IFwiJFBST0pFQ1RfRElSXCIvRmx1dHRlci9lcGhlbWVyYWwvLmFwcF9maWxlbmFtZSAmJiBcIiRGTFVUVEVSX1JPT1RcIi9wYWNrYWdlcy9mbHV0dGVyX3Rvb2xzL2Jpbi9tYWNvc19hc3NlbWJsZS5zaCBlbWJlZFxuIjsKCQl9OwoJCTMzQ0MxMTFFMjA0NEM2QkYwMDAzQzA0NSAvKiBTaGVsbFNjcmlwdCAqLyA9IHsKCQkJaXNhID0gUEJYU2hlbGxTY3JpcHRCdWlsZFBoYXNlOwoJCQlidWlsZEFjdGlvbk1hc2sgPSAyMTQ3NDgzNjQ3OwoJCQlmaWxlcyA9ICgKCQkJKTsKCQkJaW5wdXRGaWxlTGlzdFBhdGhzID0gKAoJCQkJRmx1dHRlci9lcGhlbWVyYWwvRmx1dHRlcklucHV0cy54Y2ZpbGVsaXN0LAoJCQkpOwoJCQlpbnB1dFBhdGhzID0gKAoJCQkJRmx1dHRlci9lcGhlbWVyYWwvdHJpcHdpcmUsCgkJCSk7CgkJCW91dHB1dEZpbGVMaXN0UGF0aHMgPSAoCgkJCQlGbHV0dGVyL2VwaGVtZXJhbC9GbHV0dGVyT3V0cHV0cy54Y2ZpbGVsaXN0LAoJCQkpOwoJCQlvdXRwdXRQYXRocyA9ICgKCQkJKTsKCQkJcnVuT25seUZvckRlcGxveW1lbnRQb3N0cHJvY2Vzc2luZyA9IDA7CgkJCXNoZWxsUGF0aCA9IC9iaW4vc2g7CgkJCXNoZWxsU2NyaXB0ID0gIlwiJEZMVVRURVJfUk9PVFwiL3BhY2thZ2VzL2ZsdXR0ZXJfdG9vbHMvYmluL21hY29zX2Fzc2VtYmxlLnNoICYmIHRvdWNoIEZsdXR0ZXIvZXBoZW1lcmFsL3RyaXB3aXJlIjsKCQl9OwovKiBFbmQgUEJYU2hlbGxTY3JpcHRCdWlsZFBoYXNlIHNlY3Rpb24gKi8KCi8qIEJlZ2luIFBCWFNvdXJjZXNCdWlsZFBoYXNlIHNlY3Rpb24gKi8KCQkzMzFDODBEMTI5NENGNzBGMDAyNjNCRTUgLyogU291cmNlcyAqLyA9IHsKCQkJaXNhID0gUEJYU291cmNlc0J1aWxkUGhhc2U7CgkJCWJ1aWxkQWN0aW9uTWFzayA9IDIxNDc0ODM2NDc7CgkJCWZpbGVzID0gKAoJCQkJMzMxQzgwRDgyOTRDRjcxMDAwMjYzQkU1IC8qIFJ1bm5lclRlc3RzLnN3aWZ0IGluIFNvdXJjZXMgKi8sCgkJCSk7CgkJCXJ1bk9ubHlGb3JEZXBsb3ltZW50UG9zdHByb2Nlc3NpbmcgPSAwOwoJCX07CgkJMzNDQzEwRTkyMDQ0QTNDNjAwMDNDMDQ1IC8qIFNvdXJjZXMgKi8gPSB7CgkJCWlzYSA9IFBCWFNvdXJjZXNCdWlsZFBoYXNlOwoJCQlidWlsZEFjdGlvbk1hc2sgPSAyMTQ3NDgzNjQ3OwoJCQlmaWxlcyA9ICgKCQkJCTMzQ0MxMTEzMjA0NEJGQTAwMDAzQzA0NSAvKiBNYWluRmx1dHRlcldpbmRvdy5zd2lmdCBpbiBTb3VyY2VzICovLAoJCQkJMzNDQzEwRjEyMDQ0QTNDNjAwMDNDMDQ1IC8qIEFwcERlbGVnYXRlLnN3aWZ0IGluIFNvdXJjZXMgKi8sCgkJCQkzMzVCQkQxQjIyQTlBMTVFMDBFOTA3MUQgLyogR2VuZXJhdGVkUGx1Z2luUmVnaXN0cmFudC5zd2lmdCBpbiBTb3VyY2VzICovLAoJCQkpOwoJCQlydW5Pbmx5Rm9yRGVwbG95bWVudFBvc3Rwcm9jZXNzaW5nID0gMDsKCQl9OwovKiBFbmQgUEJYU291cmNlc0J1aWxkUGhhc2Ugc2VjdGlvbiAqLwoKLyogQmVnaW4gUEJYVGFyZ2V0RGVwZW5kZW5jeSBzZWN0aW9uICovCgkJMzMxQzgwREEyOTRDRjcxMDAwMjYzQkU1IC8qIFBCWFRhcmdldERlcGVuZGVuY3kgKi8gPSB7CgkJCWlzYSA9IFBCWFRhcmdldERlcGVuZGVuY3k7CgkJCXRhcmdldCA9IDMzQ0MxMEVDMjA0NEEzQzYwMDAzQzA0NSAvKiBSdW5uZXIgKi87CgkJCXRhcmdldFByb3h5ID0gMzMxQzgwRDkyOTRDRjcxMDAwMjYzQkU1IC8qIFBCWENvbnRhaW5lckl0ZW1Qcm94eSAqLzsKCQl9OwoJCTMzQ0MxMTIwMjA0NEM3OUYwMDAzQzA0NSAvKiBQQlhUYXJnZXREZXBlbmRlbmN5ICovID0gewoJCQlpc2EgPSBQQlhUYXJnZXREZXBlbmRlbmN5OwoJCQl0YXJnZXQgPSAzM0NDMTExQTIwNDRDNkJBMDAwM0MwNDUgLyogRmx1dHRlciBBc3NlbWJsZSAqLzsKCQkJdGFyZ2V0UHJveHkgPSAzM0NDMTExRjIwNDRDNzlGMDAwM0MwNDUgLyogUEJYQ29udGFpbmVySXRlbVByb3h5ICovOwoJCX07Ci8qIEVuZCBQQlhUYXJnZXREZXBlbmRlbmN5IHNlY3Rpb24gKi8KCi8qIEJlZ2luIFBCWFZhcmlhbnRHcm91cCBzZWN0aW9uICovCgkJMzNDQzEwRjQyMDQ0QTNDNjAwMDNDMDQ1IC8qIE1haW5NZW51LnhpYiAqLyA9IHsKCQkJaXNhID0gUEJYVmFyaWFudEdyb3VwOwoJCQljaGlsZHJlbiA9ICgKCQkJCTMzQ0MxMEY1MjA0NEEzQzYwMDAzQzA0NSAvKiBCYXNlICovLAoJCQkpOwoJCQluYW1lID0gTWFpbk1lbnUueGliOwoJCQlwYXRoID0gUnVubmVyOwoJCQlzb3VyY2VUcmVlID0gIjxncm91cD4iOwoJCX07Ci8qIEVuZCBQQlhWYXJpYW50R3JvdXAgc2VjdGlvbiAqLwoKLyogQmVnaW4gWENCdWlsZENvbmZpZ3VyYXRpb24gc2VjdGlvbiAqLwoJCTMzMUM4MERCMjk0Q0Y3MTAwMDI2M0JFNSAvKiBEZWJ1ZyAqLyA9IHsKCQkJaXNhID0gWENCdWlsZENvbmZpZ3VyYXRpb247CgkJCWJ1aWxkU2V0dGluZ3MgPSB7CgkJCQlCVU5ETEVfTE9BREVSID0gIiQoVEVTVF9IT1NUKSI7CgkJCQlDVVJSRU5UX1BST0pFQ1RfVkVSU0lPTiA9IDE7CgkJCQlHRU5FUkFURV9JTkZPUExJU1RfRklMRSA9IFlFUzsKCQkJCU1BUktFVElOR19WRVJTSU9OID0gMS4wOwoJCQkJUFJPRFVDVF9CVU5ETEVfSURFTlRJRklFUiA9IHt7b3JnX25hbWV9fS57e3Byb2plY3RfbmFtZS5jYW1lbENhc2UoKX19LlJ1bm5lclRlc3RzOwoJCQkJUFJPRFVDVF9OQU1FID0gIiQoVEFSR0VUX05BTUUpIjsKCQkJCVNXSUZUX1ZFUlNJT04gPSA1LjA7CgkJCQlURVNUX0hPU1QgPSAiJChCVUlMVF9QUk9EVUNUU19ESVIpL3t7cHJvamVjdF9uYW1lfX0uYXBwLyQoQlVORExFX0VYRUNVVEFCTEVfRk9MREVSX1BBVEgpL3t7cHJvamVjdF9uYW1lfX0iOwoJCQl9OwoJCQluYW1lID0gRGVidWc7CgkJfTsKCQkzMzFDODBEQzI5NENGNzEwMDAyNjNCRTUgLyogUmVsZWFzZSAqLyA9IHsKCQkJaXNhID0gWENCdWlsZENvbmZpZ3VyYXRpb247CgkJCWJ1aWxkU2V0dGluZ3MgPSB7CgkJCQlCVU5ETEVfTE9BREVSID0gIiQoVEVTVF9IT1NUKSI7CgkJCQlDVVJSRU5UX1BST0pFQ1RfVkVSU0lPTiA9IDE7CgkJCQlHRU5FUkFURV9JTkZPUExJU1RfRklMRSA9IFlFUzsKCQkJCU1BUktFVElOR19WRVJTSU9OID0gMS4wOwoJCQkJUFJPRFVDVF9CVU5ETEVfSURFTlRJRklFUiA9IHt7b3JnX25hbWV9fS57e3Byb2plY3RfbmFtZS5jYW1lbENhc2UoKX19LlJ1bm5lclRlc3RzOwoJCQkJUFJPRFVDVF9OQU1FID0gIiQoVEFSR0VUX05BTUUpIjsKCQkJCVNXSUZUX1ZFUlNJT04gPSA1LjA7CgkJCQlURVNUX0hPU1QgPSAiJChCVUlMVF9QUk9EVUNUU19ESVIpL3t7cHJvamVjdF9uYW1lfX0uYXBwLyQoQlVORExFX0VYRUNVVEFCTEVfRk9MREVSX1BBVEgpL3t7cHJvamVjdF9uYW1lfX0iOwoJCQl9OwoJCQluYW1lID0gUmVsZWFzZTsKCQl9OwoJCTMzMUM4MEREMjk0Q0Y3MTAwMDI2M0JFNSAvKiBQcm9maWxlICovID0gewoJCQlpc2EgPSBYQ0J1aWxkQ29uZmlndXJhdGlvbjsKCQkJYnVpbGRTZXR0aW5ncyA9IHsKCQkJCUJVTkRMRV9MT0FERVIgPSAiJChURVNUX0hPU1QpIjsKCQkJCUNVUlJFTlRfUFJPSkVDVF9WRVJTSU9OID0gMTsKCQkJCUdFTkVSQVRFX0lORk9QTElTVF9GSUxFID0gWUVTOwoJCQkJTUFSS0VUSU5HX1ZFUlNJT04gPSAxLjA7CgkJCQlQUk9EVUNUX0JVTkRMRV9JREVOVElGSUVSID0ge3tvcmdfbmFtZX19Lnt7cHJvamVjdF9uYW1lLmNhbWVsQ2FzZSgpfX0uUnVubmVyVGVzdHM7CgkJCQlQUk9EVUNUX05BTUUgPSAiJChUQVJHRVRfTkFNRSkiOwoJCQkJU1dJRlRfVkVSU0lPTiA9IDUuMDsKCQkJCVRFU1RfSE9TVCA9ICIkKEJVSUxUX1BST0RVQ1RTX0RJUikve3twcm9qZWN0X25hbWV9fS5hcHAvJChCVU5ETEVfRVhFQ1VUQUJMRV9GT0xERVJfUEFUSCkve3twcm9qZWN0X25hbWV9fSI7CgkJCX07CgkJCW5hbWUgPSBQcm9maWxlOwoJCX07CgkJMzM4RDBDRTkyMzE0NThCRDAwRkE1Rjc1IC8qIFByb2ZpbGUgKi8gPSB7CgkJCWlzYSA9IFhDQnVpbGRDb25maWd1cmF0aW9uOwoJCQliYXNlQ29uZmlndXJhdGlvblJlZmVyZW5jZSA9IDdBRkEzQzhFMUQzNTM2MEMwMDgzMDgyRSAvKiBSZWxlYXNlLnhjY29uZmlnICovOwoJCQlidWlsZFNldHRpbmdzID0gewoJCQkJQUxXQVlTX1NFQVJDSF9VU0VSX1BBVEhTID0gTk87CgkJCQlDTEFOR19BTkFMWVpFUl9OT05OVUxMID0gWUVTOwoJCQkJQ0xBTkdfQU5BTFlaRVJfTlVNQkVSX09CSkVDVF9DT05WRVJTSU9OID0gWUVTX0FHR1JFU1NJVkU7CgkJCQlDTEFOR19DWFhfTEFOR1VBR0VfU1RBTkRBUkQgPSAiZ251KysxNCI7CgkJCQlDTEFOR19DWFhfTElCUkFSWSA9ICJsaWJjKysiOwoJCQkJQ0xBTkdfRU5BQkxFX01PRFVMRVMgPSBZRVM7CgkJCQlDTEFOR19FTkFCTEVfT0JKQ19BUkMgPSBZRVM7CgkJCQlDTEFOR19XQVJOX0JMT0NLX0NBUFRVUkVfQVVUT1JFTEVBU0lORyA9IFlFUzsKCQkJCUNMQU5HX1dBUk5fQk9PTF9DT05WRVJTSU9OID0gWUVTOwoJCQkJQ0xBTkdfV0FSTl9DT05TVEFOVF9DT05WRVJTSU9OID0gWUVTOwoJCQkJQ0xBTkdfV0FSTl9ERVBSRUNBVEVEX09CSkNfSU1QTEVNRU5UQVRJT05TID0gWUVTOwoJCQkJQ0xBTkdfV0FSTl9ESVJFQ1RfT0JKQ19JU0FfVVNBR0UgPSBZRVNfRVJST1I7CgkJCQlDTEFOR19XQVJOX0RPQ1VNRU5UQVRJT05fQ09NTUVOVFMgPSBZRVM7CgkJCQlDTEFOR19XQVJOX0VNUFRZX0JPRFkgPSBZRVM7CgkJCQlDTEFOR19XQVJOX0VOVU1fQ09OVkVSU0lPTiA9IFlFUzsKCQkJCUNMQU5HX1dBUk5fSU5GSU5JVEVfUkVDVVJTSU9OID0gWUVTOwoJCQkJQ0xBTkdfV0FSTl9JTlRfQ09OVkVSU0lPTiA9IFlFUzsKCQkJCUNMQU5HX1dBUk5fTk9OX0xJVEVSQUxfTlVMTF9DT05WRVJTSU9OID0gWUVTOwoJCQkJQ0xBTkdfV0FSTl9PQkpDX0xJVEVSQUxfQ09OVkVSU0lPTiA9IFlFUzsKCQkJCUNMQU5HX1dBUk5fT0JKQ19ST09UX0NMQVNTID0gWUVTX0VSUk9SOwoJCQkJQ0xBTkdfV0FSTl9SQU5HRV9MT09QX0FOQUxZU0lTID0gWUVTOwoJCQkJQ0xBTkdfV0FSTl9TVVNQSUNJT1VTX01PVkUgPSBZRVM7CgkJCQlDT0RFX1NJR05fSURFTlRJVFkgPSAiLSI7CgkJCQlDT1BZX1BIQVNFX1NUUklQID0gTk87CgkJCQlERUJVR19JTkZPUk1BVElPTl9GT1JNQVQgPSAiZHdhcmYtd2l0aC1kc3ltIjsKCQkJCUVOQUJMRV9OU19BU1NFUlRJT05TID0gTk87CgkJCQlFTkFCTEVfU1RSSUNUX09CSkNfTVNHU0VORCA9IFlFUzsKCQkJCUdDQ19DX0xBTkdVQUdFX1NUQU5EQVJEID0gZ251MTE7CgkJCQlHQ0NfTk9fQ09NTU9OX0JMT0NLUyA9IFlFUzsKCQkJCUdDQ19XQVJOXzY0X1RPXzMyX0JJVF9DT05WRVJTSU9OID0gWUVTOwoJCQkJR0NDX1dBUk5fQUJPVVRfUkVUVVJOX1RZUEUgPSBZRVNfRVJST1I7CgkJCQlHQ0NfV0FSTl9VTklOSVRJQUxJWkVEX0FVVE9TID0gWUVTX0FHR1JFU1NJVkU7CgkJCQlHQ0NfV0FSTl9VTlVTRURfRlVOQ1RJT04gPSBZRVM7CgkJCQlHQ0NfV0FSTl9VTlVTRURfVkFSSUFCTEUgPSBZRVM7CgkJCQlNQUNPU1hfREVQTE9ZTUVOVF9UQVJHRVQgPSAxMC4xNC42OwoJCQkJTVRMX0VOQUJMRV9ERUJVR19JTkZPID0gTk87CgkJCQlTREtST09UID0gbWFjb3N4OwoJCQkJU1dJRlRfQ09NUElMQVRJT05fTU9ERSA9IHdob2xlbW9kdWxlOwoJCQkJU1dJRlRfT1BUSU1JWkFUSU9OX0xFVkVMID0gIi1PIjsKCQkJfTsKCQkJbmFtZSA9IFByb2ZpbGU7CgkJfTsKCQkzMzhEMENFQTIzMTQ1OEJEMDBGQTVGNzUgLyogUHJvZmlsZSAqLyA9IHsKCQkJaXNhID0gWENCdWlsZENvbmZpZ3VyYXRpb247CgkJCWJhc2VDb25maWd1cmF0aW9uUmVmZXJlbmNlID0gMzNFNTE5NEYyMzI4Mjg4NjAwMjZFRTREIC8qIEFwcEluZm8ueGNjb25maWcgKi87CgkJCWJ1aWxkU2V0dGluZ3MgPSB7CgkJCQlBU1NFVENBVEFMT0dfQ09NUElMRVJfQVBQSUNPTl9OQU1FID0gQXBwSWNvbjsKCQkJCUNMQU5HX0VOQUJMRV9NT0RVTEVTID0gWUVTOwoJCQkJQ09ERV9TSUdOX0VOVElUTEVNRU5UUyA9IFJ1bm5lci9EZWJ1Z1Byb2ZpbGUuZW50aXRsZW1lbnRzOwoJCQkJQ09ERV9TSUdOX1NUWUxFID0gQXV0b21hdGljOwoJCQkJQ09NQklORV9ISURQSV9JTUFHRVMgPSBZRVM7CgkJCQlJTkZPUExJU1RfRklMRSA9IFJ1bm5lci9JbmZvLnBsaXN0OwoJCQkJTERfUlVOUEFUSF9TRUFSQ0hfUEFUSFMgPSAoCgkJCQkJIiQoaW5oZXJpdGVkKSIsCgkJCQkJIkBleGVjdXRhYmxlX3BhdGgvLi4vRnJhbWV3b3JrcyIsCgkJCQkpOwoJCQkJUFJPVklTSU9OSU5HX1BST0ZJTEVfU1BFQ0lGSUVSID0gIiI7CgkJCQlTV0lGVF9WRVJTSU9OID0gNS4wOwoJCQl9OwoJCQluYW1lID0gUHJvZmlsZTsKCQl9OwoJCTMzOEQwQ0VCMjMxNDU4QkQwMEZBNUY3NSAvKiBQcm9maWxlICovID0gewoJCQlpc2EgPSBYQ0J1aWxkQ29uZmlndXJhdGlvbjsKCQkJYnVpbGRTZXR0aW5ncyA9IHsKCQkJCUNPREVfU0lHTl9TVFlMRSA9IE1hbnVhbDsKCQkJCVBST0RVQ1RfTkFNRSA9ICIkKFRBUkdFVF9OQU1FKSI7CgkJCX07CgkJCW5hbWUgPSBQcm9maWxlOwoJCX07CgkJMzNDQzEwRjkyMDQ0QTNDNjAwMDNDMDQ1IC8qIERlYnVnICovID0gewoJCQlpc2EgPSBYQ0J1aWxkQ29uZmlndXJhdGlvbjsKCQkJYmFzZUNvbmZpZ3VyYXRpb25SZWZlcmVuY2UgPSA5NzQwRUVCMjFDRjkwMTk1MDA0Mzg0RkMgLyogRGVidWcueGNjb25maWcgKi87CgkJCWJ1aWxkU2V0dGluZ3MgPSB7CgkJCQlBTFdBWVNfU0VBUkNIX1VTRVJfUEFUSFMgPSBOTzsKCQkJCUNMQU5HX0FOQUxZWkVSX05PTk5VTEwgPSBZRVM7CgkJCQlDTEFOR19BTkFMWVpFUl9OVU1CRVJfT0JKRUNUX0NPTlZFUlNJT04gPSBZRVNfQUdHUkVTU0lWRTsKCQkJCUNMQU5HX0NYWF9MQU5HVUFHRV9TVEFOREFSRCA9ICJnbnUrKzE0IjsKCQkJCUNMQU5HX0NYWF9MSUJSQVJZID0gImxpYmMrKyI7CgkJCQlDTEFOR19FTkFCTEVfTU9EVUxFUyA9IFlFUzsKCQkJCUNMQU5HX0VOQUJMRV9PQkpDX0FSQyA9IFlFUzsKCQkJCUNMQU5HX1dBUk5fQkxPQ0tfQ0FQVFVSRV9BVVRPUkVMRUFTSU5HID0gWUVTOwoJCQkJQ0xBTkdfV0FSTl9CT09MX0NPTlZFUlNJT04gPSBZRVM7CgkJCQlDTEFOR19XQVJOX0NPTlNUQU5UX0NPTlZFUlNJT04gPSBZRVM7CgkJCQlDTEFOR19XQVJOX0RFUFJFQ0FURURfT0JKQ19JTVBMRU1FTlRBVElPTlMgPSBZRVM7CgkJCQlDTEFOR19XQVJOX0RJUkVDVF9PQkpDX0lTQV9VU0FHRSA9IFlFU19FUlJPUjsKCQkJCUNMQU5HX1dBUk5fRE9DVU1FTlRBVElPTl9DT01NRU5UUyA9IFlFUzsKCQkJCUNMQU5HX1dBUk5fRU1QVFlfQk9EWSA9IFlFUzsKCQkJCUNMQU5HX1dBUk5fRU5VTV9DT05WRVJTSU9OID0gWUVTOwoJCQkJQ0xBTkdfV0FSTl9JTkZJTklURV9SRUNVUlNJT04gPSBZRVM7CgkJCQlDTEFOR19XQVJOX0lOVF9DT05WRVJTSU9OID0gWUVTOwoJCQkJQ0xBTkdfV0FSTl9OT05fTElURVJBTF9OVUxMX0NPTlZFUlNJT04gPSBZRVM7CgkJCQlDTEFOR19XQVJOX09CSkNfTElURVJBTF9DT05WRVJTSU9OID0gWUVTOwoJCQkJQ0xBTkdfV0FSTl9PQkpDX1JPT1RfQ0xBU1MgPSBZRVNfRVJST1I7CgkJCQlDTEFOR19XQVJOX1JBTkdFX0xPT1BfQU5BTFlTSVMgPSBZRVM7CgkJCQlDTEFOR19XQVJOX1NVU1BJQ0lPVVNfTU9WRSA9IFlFUzsKCQkJCUNPREVfU0lHTl9JREVOVElUWSA9ICItIjsKCQkJCUNPUFlfUEhBU0VfU1RSSVAgPSBOTzsKCQkJCURFQlVHX0lORk9STUFUSU9OX0ZPUk1BVCA9IGR3YXJmOwoJCQkJRU5BQkxFX1NUUklDVF9PQkpDX01TR1NFTkQgPSBZRVM7CgkJCQlFTkFCTEVfVEVTVEFCSUxJVFkgPSBZRVM7CgkJCQlHQ0NfQ19MQU5HVUFHRV9TVEFOREFSRCA9IGdudTExOwoJCQkJR0NDX0RZTkFNSUNfTk9fUElDID0gTk87CgkJCQlHQ0NfTk9fQ09NTU9OX0JMT0NLUyA9IFlFUzsKCQkJCUdDQ19PUFRJTUlaQVRJT05fTEVWRUwgPSAwOwoJCQkJR0NDX1BSRVBST0NFU1NPUl9ERUZJTklUSU9OUyA9ICgKCQkJCQkiREVCVUc9MSIsCgkJCQkJIiQoaW5oZXJpdGVkKSIsCgkJCQkpOwoJCQkJR0NDX1dBUk5fNjRfVE9fMzJfQklUX0NPTlZFUlNJT04gPSBZRVM7CgkJCQlHQ0NfV0FSTl9BQk9VVF9SRVRVUk5fVFlQRSA9IFlFU19FUlJPUjsKCQkJCUdDQ19XQVJOX1VOSU5JVElBTElaRURfQVVUT1MgPSBZRVNfQUdHUkVTU0lWRTsKCQkJCUdDQ19XQVJOX1VOVVNFRF9GVU5DVElPTiA9IFlFUzsKCQkJCUdDQ19XQVJOX1VOVVNFRF9WQVJJQUJMRSA9IFlFUzsKCQkJCU1BQ09TWF9ERVBMT1lNRU5UX1RBUkdFVCA9IDEwLjE0LjY7CgkJCQlNVExfRU5BQkxFX0RFQlVHX0lORk8gPSBZRVM7CgkJCQlPTkxZX0FDVElWRV9BUkNIID0gWUVTOwoJCQkJU0RLUk9PVCA9IG1hY29zeDsKCQkJCVNXSUZUX0FDVElWRV9DT01QSUxBVElPTl9DT05ESVRJT05TID0gREVCVUc7CgkJCQlTV0lGVF9PUFRJTUlaQVRJT05fTEVWRUwgPSAiLU9ub25lIjsKCQkJfTsKCQkJbmFtZSA9IERlYnVnOwoJCX07CgkJMzNDQzEwRkEyMDQ0QTNDNjAwMDNDMDQ1IC8qIFJlbGVhc2UgKi8gPSB7CgkJCWlzYSA9IFhDQnVpbGRDb25maWd1cmF0aW9uOwoJCQliYXNlQ29uZmlndXJhdGlvblJlZmVyZW5jZSA9IDdBRkEzQzhFMUQzNTM2MEMwMDgzMDgyRSAvKiBSZWxlYXNlLnhjY29uZmlnICovOwoJCQlidWlsZFNldHRpbmdzID0gewoJCQkJQUxXQVlTX1NFQVJDSF9VU0VSX1BBVEhTID0gTk87CgkJCQlDTEFOR19BTkFMWVpFUl9OT05OVUxMID0gWUVTOwoJCQkJQ0xBTkdfQU5BTFlaRVJfTlVNQkVSX09CSkVDVF9DT05WRVJTSU9OID0gWUVTX0FHR1JFU1NJVkU7CgkJCQlDTEFOR19DWFhfTEFOR1VBR0VfU1RBTkRBUkQgPSAiZ251KysxNCI7CgkJCQlDTEFOR19DWFhfTElCUkFSWSA9ICJsaWJjKysiOwoJCQkJQ0xBTkdfRU5BQkxFX01PRFVMRVMgPSBZRVM7CgkJCQlDTEFOR19FTkFCTEVfT0JKQ19BUkMgPSBZRVM7CgkJCQlDTEFOR19XQVJOX0JMT0NLX0NBUFRVUkVfQVVUT1JFTEVBU0lORyA9IFlFUzsKCQkJCUNMQU5HX1dBUk5fQk9PTF9DT05WRVJTSU9OID0gWUVTOwoJCQkJQ0xBTkdfV0FSTl9DT05TVEFOVF9DT05WRVJTSU9OID0gWUVTOwoJCQkJQ0xBTkdfV0FSTl9ERVBSRUNBVEVEX09CSkNfSU1QTEVNRU5UQVRJT05TID0gWUVTOwoJCQkJQ0xBTkdfV0FSTl9ESVJFQ1RfT0JKQ19JU0FfVVNBR0UgPSBZRVNfRVJST1I7CgkJCQlDTEFOR19XQVJOX0RPQ1VNRU5UQVRJT05fQ09NTUVOVFMgPSBZRVM7CgkJCQlDTEFOR19XQVJOX0VNUFRZX0JPRFkgPSBZRVM7CgkJCQlDTEFOR19XQVJOX0VOVU1fQ09OVkVSU0lPTiA9IFlFUzsKCQkJCUNMQU5HX1dBUk5fSU5GSU5JVEVfUkVDVVJTSU9OID0gWUVTOwoJCQkJQ0xBTkdfV0FSTl9JTlRfQ09OVkVSU0lPTiA9IFlFUzsKCQkJCUNMQU5HX1dBUk5fTk9OX0xJVEVSQUxfTlVMTF9DT05WRVJTSU9OID0gWUVTOwoJCQkJQ0xBTkdfV0FSTl9PQkpDX0xJVEVSQUxfQ09OVkVSU0lPTiA9IFlFUzsKCQkJCUNMQU5HX1dBUk5fT0JKQ19ST09UX0NMQVNTID0gWUVTX0VSUk9SOwoJCQkJQ0xBTkdfV0FSTl9SQU5HRV9MT09QX0FOQUxZU0lTID0gWUVTOwoJCQkJQ0xBTkdfV0FSTl9TVVNQSUNJT1VTX01PVkUgPSBZRVM7CgkJCQlDT0RFX1NJR05fSURFTlRJVFkgPSAiLSI7CgkJCQlDT1BZX1BIQVNFX1NUUklQID0gTk87CgkJCQlERUJVR19JTkZPUk1BVElPTl9GT1JNQVQgPSAiZHdhcmYtd2l0aC1kc3ltIjsKCQkJCUVOQUJMRV9OU19BU1NFUlRJT05TID0gTk87CgkJCQlFTkFCTEVfU1RSSUNUX09CSkNfTVNHU0VORCA9IFlFUzsKCQkJCUdDQ19DX0xBTkdVQUdFX1NUQU5EQVJEID0gZ251MTE7CgkJCQlHQ0NfTk9fQ09NTU9OX0JMT0NLUyA9IFlFUzsKCQkJCUdDQ19XQVJOXzY0X1RPXzMyX0JJVF9DT05WRVJTSU9OID0gWUVTOwoJCQkJR0NDX1dBUk5fQUJPVVRfUkVUVVJOX1RZUEUgPSBZRVNfRVJST1I7CgkJCQlHQ0NfV0FSTl9VTklOSVRJQUxJWkVEX0FVVE9TID0gWUVTX0FHR1JFU1NJVkU7CgkJCQlHQ0NfV0FSTl9VTlVTRURfRlVOQ1RJT04gPSBZRVM7CgkJCQlHQ0NfV0FSTl9VTlVTRURfVkFSSUFCTEUgPSBZRVM7CgkJCQlNQUNPU1hfREVQTE9ZTUVOVF9UQVJHRVQgPSAxMC4xNC42OwoJCQkJTVRMX0VOQUJMRV9ERUJVR19JTkZPID0gTk87CgkJCQlTREtST09UID0gbWFjb3N4OwoJCQkJU1dJRlRfQ09NUElMQVRJT05fTU9ERSA9IHdob2xlbW9kdWxlOwoJCQkJU1dJRlRfT1BUSU1JWkFUSU9OX0xFVkVMID0gIi1PIjsKCQkJfTsKCQkJbmFtZSA9IFJlbGVhc2U7CgkJfTsKCQkzM0NDMTBGQzIwNDRBM0M2MDAwM0MwNDUgLyogRGVidWcgKi8gPSB7CgkJCWlzYSA9IFhDQnVpbGRDb25maWd1cmF0aW9uOwoJCQliYXNlQ29uZmlndXJhdGlvblJlZmVyZW5jZSA9IDMzRTUxOTRGMjMyODI4ODYwMDI2RUU0RCAvKiBBcHBJbmZvLnhjY29uZmlnICovOwoJCQlidWlsZFNldHRpbmdzID0gewoJCQkJQVNTRVRDQVRBTE9HX0NPTVBJTEVSX0FQUElDT05fTkFNRSA9IEFwcEljb247CgkJCQlDTEFOR19FTkFCTEVfTU9EVUxFUyA9IFlFUzsKCQkJCUNPREVfU0lHTl9FTlRJVExFTUVOVFMgPSBSdW5uZXIvRGVidWdQcm9maWxlLmVudGl0bGVtZW50czsKCQkJCUNPREVfU0lHTl9TVFlMRSA9IEF1dG9tYXRpYzsKCQkJCUNPTUJJTkVfSElEUElfSU1BR0VTID0gWUVTOwoJCQkJSU5GT1BMSVNUX0ZJTEUgPSBSdW5uZXIvSW5mby5wbGlzdDsKCQkJCUxEX1JVTlBBVEhfU0VBUkNIX1BBVEhTID0gKAoJCQkJCSIkKGluaGVyaXRlZCkiLAoJCQkJCSJAZXhlY3V0YWJsZV9wYXRoLy4uL0ZyYW1ld29ya3MiLAoJCQkJKTsKCQkJCVBST1ZJU0lPTklOR19QUk9GSUxFX1NQRUNJRklFUiA9ICIiOwoJCQkJU1dJRlRfT1BUSU1JWkFUSU9OX0xFVkVMID0gIi1Pbm9uZSI7CgkJCQlTV0lGVF9WRVJTSU9OID0gNS4wOwoJCQl9OwoJCQluYW1lID0gRGVidWc7CgkJfTsKCQkzM0NDMTBGRDIwNDRBM0M2MDAwM0MwNDUgLyogUmVsZWFzZSAqLyA9IHsKCQkJaXNhID0gWENCdWlsZENvbmZpZ3VyYXRpb247CgkJCWJhc2VDb25maWd1cmF0aW9uUmVmZXJlbmNlID0gMzNFNTE5NEYyMzI4Mjg4NjAwMjZFRTREIC8qIEFwcEluZm8ueGNjb25maWcgKi87CgkJCWJ1aWxkU2V0dGluZ3MgPSB7CgkJCQlBU1NFVENBVEFMT0dfQ09NUElMRVJfQVBQSUNPTl9OQU1FID0gQXBwSWNvbjsKCQkJCUNMQU5HX0VOQUJMRV9NT0RVTEVTID0gWUVTOwoJCQkJQ09ERV9TSUdOX0VOVElUTEVNRU5UUyA9IFJ1bm5lci9SZWxlYXNlLmVudGl0bGVtZW50czsKCQkJCUNPREVfU0lHTl9TVFlMRSA9IEF1dG9tYXRpYzsKCQkJCUNPTUJJTkVfSElEUElfSU1BR0VTID0gWUVTOwoJCQkJSU5GT1BMSVNUX0ZJTEUgPSBSdW5uZXIvSW5mby5wbGlzdDsKCQkJCUxEX1JVTlBBVEhfU0VBUkNIX1BBVEhTID0gKAoJCQkJCSIkKGluaGVyaXRlZCkiLAoJCQkJCSJAZXhlY3V0YWJsZV9wYXRoLy4uL0ZyYW1ld29ya3MiLAoJCQkJKTsKCQkJCVBST1ZJU0lPTklOR19QUk9GSUxFX1NQRUNJRklFUiA9ICIiOwoJCQkJU1dJRlRfVkVSU0lPTiA9IDUuMDsKCQkJfTsKCQkJbmFtZSA9IFJlbGVhc2U7CgkJfTsKCQkzM0NDMTExQzIwNDRDNkJBMDAwM0MwNDUgLyogRGVidWcgKi8gPSB7CgkJCWlzYSA9IFhDQnVpbGRDb25maWd1cmF0aW9uOwoJCQlidWlsZFNldHRpbmdzID0gewoJCQkJQ09ERV9TSUdOX1NUWUxFID0gTWFudWFsOwoJCQkJUFJPRFVDVF9OQU1FID0gIiQoVEFSR0VUX05BTUUpIjsKCQkJfTsKCQkJbmFtZSA9IERlYnVnOwoJCX07CgkJMzNDQzExMUQyMDQ0QzZCQTAwMDNDMDQ1IC8qIFJlbGVhc2UgKi8gPSB7CgkJCWlzYSA9IFhDQnVpbGRDb25maWd1cmF0aW9uOwoJCQlidWlsZFNldHRpbmdzID0gewoJCQkJQ09ERV9TSUdOX1NUWUxFID0gQXV0b21hdGljOwoJCQkJUFJPRFVDVF9OQU1FID0gIiQoVEFSR0VUX05BTUUpIjsKCQkJfTsKCQkJbmFtZSA9IFJlbGVhc2U7CgkJfTsKLyogRW5kIFhDQnVpbGRDb25maWd1cmF0aW9uIHNlY3Rpb24gKi8KCi8qIEJlZ2luIFhDQ29uZmlndXJhdGlvbkxpc3Qgc2VjdGlvbiAqLwoJCTMzMUM4MERFMjk0Q0Y3MTAwMDI2M0JFNSAvKiBCdWlsZCBjb25maWd1cmF0aW9uIGxpc3QgZm9yIFBCWE5hdGl2ZVRhcmdldCAiUnVubmVyVGVzdHMiICovID0gewoJCQlpc2EgPSBYQ0NvbmZpZ3VyYXRpb25MaXN0OwoJCQlidWlsZENvbmZpZ3VyYXRpb25zID0gKAoJCQkJMzMxQzgwREIyOTRDRjcxMDAwMjYzQkU1IC8qIERlYnVnICovLAoJCQkJMzMxQzgwREMyOTRDRjcxMDAwMjYzQkU1IC8qIFJlbGVhc2UgKi8sCgkJCQkzMzFDODBERDI5NENGNzEwMDAyNjNCRTUgLyogUHJvZmlsZSAqLywKCQkJKTsKCQkJZGVmYXVsdENvbmZpZ3VyYXRpb25Jc1Zpc2libGUgPSAwOwoJCQlkZWZhdWx0Q29uZmlndXJhdGlvbk5hbWUgPSBSZWxlYXNlOwoJCX07CgkJMzNDQzEwRTgyMDQ0QTNDNjAwMDNDMDQ1IC8qIEJ1aWxkIGNvbmZpZ3VyYXRpb24gbGlzdCBmb3IgUEJYUHJvamVjdCAiUnVubmVyIiAqLyA9IHsKCQkJaXNhID0gWENDb25maWd1cmF0aW9uTGlzdDsKCQkJYnVpbGRDb25maWd1cmF0aW9ucyA9ICgKCQkJCTMzQ0MxMEY5MjA0NEEzQzYwMDAzQzA0NSAvKiBEZWJ1ZyAqLywKCQkJCTMzQ0MxMEZBMjA0NEEzQzYwMDAzQzA0NSAvKiBSZWxlYXNlICovLAoJCQkJMzM4RDBDRTkyMzE0NThCRDAwRkE1Rjc1IC8qIFByb2ZpbGUgKi8sCgkJCSk7CgkJCWRlZmF1bHRDb25maWd1cmF0aW9uSXNWaXNpYmxlID0gMDsKCQkJZGVmYXVsdENvbmZpZ3VyYXRpb25OYW1lID0gUmVsZWFzZTsKCQl9OwoJCTMzQ0MxMEZCMjA0NEEzQzYwMDAzQzA0NSAvKiBCdWlsZCBjb25maWd1cmF0aW9uIGxpc3QgZm9yIFBCWE5hdGl2ZVRhcmdldCAiUnVubmVyIiAqLyA9IHsKCQkJaXNhID0gWENDb25maWd1cmF0aW9uTGlzdDsKCQkJYnVpbGRDb25maWd1cmF0aW9ucyA9ICgKCQkJCTMzQ0MxMEZDMjA0NEEzQzYwMDAzQzA0NSAvKiBEZWJ1ZyAqLywKCQkJCTMzQ0MxMEZEMjA0NEEzQzYwMDAzQzA0NSAvKiBSZWxlYXNlICovLAoJCQkJMzM4RDBDRUEyMzE0NThCRDAwRkE1Rjc1IC8qIFByb2ZpbGUgKi8sCgkJCSk7CgkJCWRlZmF1bHRDb25maWd1cmF0aW9uSXNWaXNpYmxlID0gMDsKCQkJZGVmYXVsdENvbmZpZ3VyYXRpb25OYW1lID0gUmVsZWFzZTsKCQl9OwoJCTMzQ0MxMTFCMjA0NEM2QkEwMDAzQzA0NSAvKiBCdWlsZCBjb25maWd1cmF0aW9uIGxpc3QgZm9yIFBCWEFnZ3JlZ2F0ZVRhcmdldCAiRmx1dHRlciBBc3NlbWJsZSIgKi8gPSB7CgkJCWlzYSA9IFhDQ29uZmlndXJhdGlvbkxpc3Q7CgkJCWJ1aWxkQ29uZmlndXJhdGlvbnMgPSAoCgkJCQkzM0NDMTExQzIwNDRDNkJBMDAwM0MwNDUgLyogRGVidWcgKi8sCgkJCQkzM0NDMTExRDIwNDRDNkJBMDAwM0MwNDUgLyogUmVsZWFzZSAqLywKCQkJCTMzOEQwQ0VCMjMxNDU4QkQwMEZBNUY3NSAvKiBQcm9maWxlICovLAoJCQkpOwoJCQlkZWZhdWx0Q29uZmlndXJhdGlvbklzVmlzaWJsZSA9IDA7CgkJCWRlZmF1bHRDb25maWd1cmF0aW9uTmFtZSA9IFJlbGVhc2U7CgkJfTsKLyogRW5kIFhDQ29uZmlndXJhdGlvbkxpc3Qgc2VjdGlvbiAqLwoJfTsKCXJvb3RPYmplY3QgPSAzM0NDMTBFNTIwNDRBM0M2MDAwM0MwNDUgLyogUHJvamVjdCBvYmplY3QgKi87Cn0K",
       "type": "text"
     },
     {
