+++ packages/rapid_cli/lib/src/project/bundles/platform_ui_package_bundle.dart
@@ -66,7 +66,7 @@ final platformUiPackageBundle = MasonBundle.fromJson(<String, dynamic>{
     {
       "path": "pubspec.yaml",
       "data":
-          "e3sjYW5kcm9pZH19bmFtZToge3twcm9qZWN0X25hbWV9fV91aV9hbmRyb2lkCmRlc2NyaXB0aW9uOiBXaWRnZXRzIGFuZCB0aGVtZXMgaW1wbGVtZW50aW5nIHt7cHJvamVjdF9uYW1lLnRpdGxlQ2FzZSgpfX0gZGVzaWduIGxhbmd1YWdlIGZvciBBbmRyb2lkLgp2ZXJzaW9uOiAwLjAuMQpwdWJsaXNoX3RvOiBub25lCgplbnZpcm9ubWVudDoKICBzZGs6ICI+PTMuMC4wIDw0LjAuMCIKICBmbHV0dGVyOiAiPj0zLjEwLjAiCgpkZXBlbmRlbmNpZXM6CiAgZmx1dHRlcjoKICAgIHNkazogZmx1dHRlcgogIGZsdXR0ZXJfbG9jYWxpemF0aW9uczoKICAgIHNkazogZmx1dHRlcgogIHRoZW1lX3RhaWxvcl9hbm5vdGF0aW9uOiBeMi4wLjAKICB7e3Byb2plY3RfbmFtZX19X3VpOgoKZGV2X2RlcGVuZGVuY2llczoKICBmbHV0dGVyX3Rlc3Q6CiAgICBzZGs6IGZsdXR0ZXIKICBhbGNoZW1pc3Q6IF4wLjYuMQogIGJ1aWxkX3J1bm5lcjogXjIuNC41CiAgdGhlbWVfdGFpbG9yOiBeMi4wLjAKICBmbHV0dGVyX2dlbl9ydW5uZXI6IF41LjMuMQogIGZsdXR0ZXJfbGludHM6IF4yLjAuMQoKIyBUaGUgZm9sbG93aW5nIHNlY3Rpb24gaXMgc3BlY2lmaWMgdG8gdGhlIEZsdXR0ZXJHZW4gcGFja2FnZS4KIyBTZWU6IGh0dHBzOi8vcHViLmRldi9wYWNrYWdlcy9mbHV0dGVyX2dlbiNjb25maWd1cmF0aW9uLWZpbGUKZmx1dHRlcl9nZW46CiAgb3V0cHV0OiBsaWIvc3JjLwogICMgVG8gYWRkIHN1cHBvcnQgZm9yIHN2ZyB2aWEgZmx1dHRlcl9zdmcgcGFja2FnZQogICMgaW50ZWdyYXRpb25zOgogICMgICBmbHV0dGVyX3N2ZzogdHJ1ZQoKICBhc3NldHM6CiAgICBvdXRwdXRzOgogICAgICBwYWNrYWdlX3BhcmFtZXRlcl9lbmFibGVkOiB0cnVlCgojIFRoZSBmb2xsb3dpbmcgc2VjdGlvbiBpcyBzcGVjaWZpYyB0byBGbHV0dGVyIHBhY2thZ2VzLgpmbHV0dGVyOgogICMgVGhlIGZvbGxvd2luZyBsaW5lIGVuc3VyZXMgdGhhdCB0aGUgTWF0ZXJpYWwgSWNvbnMgZm9udCBpcwogICMgaW5jbHVkZWQgd2l0aCB5b3VyIGFwcGxpY2F0aW9uLCBzbyB0aGF0IHlvdSBjYW4gdXNlIHRoZSBpY29ucyBpbgogICMgdGhlIG1hdGVyaWFsIEljb25zIGNsYXNzLgogIHVzZXMtbWF0ZXJpYWwtZGVzaWduOiB0cnVlCiAgIyBUbyBhZGQgYXNzZXRzIHRvIHlvdXIgcGFja2FnZSwgYWRkIGFuIGFzc2V0cyBzZWN0aW9uLCBsaWtlIHRoaXM6CiAgIyBhc3NldHM6CiAgIyAgIC0gaW1hZ2VzL2FfZG90X2J1cnIuanBlZwogICMgICAtIGltYWdlcy9hX2RvdF9oYW0uanBlZwogICMKICAjIEZvciBkZXRhaWxzIHJlZ2FyZGluZyBhc3NldHMgaW4gcGFja2FnZXMsIHNlZQogICMgaHR0cHM6Ly9mbHV0dGVyLmRldi9hc3NldHMtYW5kLWltYWdlcy8jZnJvbS1wYWNrYWdlcwogICMKICAjIEFuIGltYWdlIGFzc2V0IGNhbiByZWZlciB0byBvbmUgb3IgbW9yZSByZXNvbHV0aW9uLXNwZWNpZmljICJ2YXJpYW50cyIsIHNlZQogICMgaHR0cHM6Ly9mbHV0dGVyLmRldi9hc3NldHMtYW5kLWltYWdlcy8jcmVzb2x1dGlvbi1hd2FyZQoKICAjIFRvIGFkZCBjdXN0b20gZm9udHMgdG8geW91ciBwYWNrYWdlLCBhZGQgYSBmb250cyBzZWN0aW9uIGhlcmUsCiAgIyBpbiB0aGlzICJmbHV0dGVyIiBzZWN0aW9uLiBFYWNoIGVudHJ5IGluIHRoaXMgbGlzdCBzaG91bGQgaGF2ZSBhCiAgIyAiZmFtaWx5IiBrZXkgd2l0aCB0aGUgZm9udCBmYW1pbHkgbmFtZSwgYW5kIGEgImZvbnRzIiBrZXkgd2l0aCBhCiAgIyBsaXN0IGdpdmluZyB0aGUgYXNzZXQgYW5kIG90aGVyIGRlc2NyaXB0b3JzIGZvciB0aGUgZm9udC4gRm9yCiAgIyBleGFtcGxlOgogICMgZm9udHM6CiAgIyAgIC0gZmFtaWx5OiBTY2h5bGVyCiAgIyAgICAgZm9udHM6CiAgIyAgICAgICAtIGFzc2V0OiBmb250cy9TY2h5bGVyLVJlZ3VsYXIudHRmCiAgIyAgICAgICAtIGFzc2V0OiBmb250cy9TY2h5bGVyLUl0YWxpYy50dGYKICAjICAgICAgICAgc3R5bGU6IGl0YWxpYwogICMgICAtIGZhbWlseTogVHJhamFuIFBybwogICMgICAgIGZvbnRzOgogICMgICAgICAgLSBhc3NldDogZm9udHMvVHJhamFuUHJvLnR0ZgogICMgICAgICAgLSBhc3NldDogZm9udHMvVHJhamFuUHJvX0JvbGQudHRmCiAgIyAgICAgICAgIHdlaWdodDogNzAwCiAgIwogICMgRm9yIGRldGFpbHMgcmVnYXJkaW5nIGZvbnRzIGluIHBhY2thZ2VzLCBzZWUKICAjIGh0dHBzOi8vZmx1dHRlci5kZXYvY3VzdG9tLWZvbnRzLyNmcm9tLXBhY2thZ2VzCnt7L2FuZHJvaWR9fXt7I2lvc319bmFtZToge3twcm9qZWN0X25hbWV9fV91aV9pb3MKZGVzY3JpcHRpb246IFdpZGdldHMgYW5kIHRoZW1lcyBpbXBsZW1lbnRpbmcge3twcm9qZWN0X25hbWUudGl0bGVDYXNlKCl9fSBkZXNpZ24gbGFuZ3VhZ2UgZm9yIGlPUy4KdmVyc2lvbjogMC4wLjEKcHVibGlzaF90bzogbm9uZQoKZW52aXJvbm1lbnQ6CiAgc2RrOiAiPj0zLjAuMCA8NC4wLjAiCiAgZmx1dHRlcjogIj49My4xMC4wIgoKZGVwZW5kZW5jaWVzOgogIGZsdXR0ZXI6CiAgICBzZGs6IGZsdXR0ZXIKICBmbHV0dGVyX2xvY2FsaXphdGlvbnM6CiAgICBzZGs6IGZsdXR0ZXIKICB0aGVtZV90YWlsb3JfYW5ub3RhdGlvbjogXjIuMC4wCiAgY3VwZXJ0aW5vX2ljb25zOiBeMS4wLjUKICB7e3Byb2plY3RfbmFtZX19X3VpOgoKZGV2X2RlcGVuZGVuY2llczoKICBmbHV0dGVyX3Rlc3Q6CiAgICBzZGs6IGZsdXR0ZXIKICBhbGNoZW1pc3Q6IF4wLjYuMQogIGJ1aWxkX3J1bm5lcjogXjIuNC41CiAgdGhlbWVfdGFpbG9yOiBeMi4wLjAKICBmbHV0dGVyX2dlbl9ydW5uZXI6IF41LjMuMQogIGZsdXR0ZXJfbGludHM6IF4yLjAuMQoKIyBUaGUgZm9sbG93aW5nIHNlY3Rpb24gaXMgc3BlY2lmaWMgdG8gdGhlIEZsdXR0ZXJHZW4gcGFja2FnZS4KIyBTZWU6IGh0dHBzOi8vcHViLmRldi9wYWNrYWdlcy9mbHV0dGVyX2dlbiNjb25maWd1cmF0aW9uLWZpbGUKZmx1dHRlcl9nZW46CiAgb3V0cHV0OiBsaWIvc3JjLwogICMgVG8gYWRkIHN1cHBvcnQgZm9yIHN2ZyB2aWEgZmx1dHRlcl9zdmcgcGFja2FnZQogICMgaW50ZWdyYXRpb25zOgogICMgICBmbHV0dGVyX3N2ZzogdHJ1ZQoKICBhc3NldHM6CiAgICBvdXRwdXRzOgogICAgICBwYWNrYWdlX3BhcmFtZXRlcl9lbmFibGVkOiB0cnVlCgojIFRoZSBmb2xsb3dpbmcgc2VjdGlvbiBpcyBzcGVjaWZpYyB0byBGbHV0dGVyIHBhY2thZ2VzLgpmbHV0dGVyOgoKICAjIFRvIGFkZCBhc3NldHMgdG8geW91ciBwYWNrYWdlLCBhZGQgYW4gYXNzZXRzIHNlY3Rpb24sIGxpa2UgdGhpczoKICAjIGFzc2V0czoKICAjICAgLSBpbWFnZXMvYV9kb3RfYnVyci5qcGVnCiAgIyAgIC0gaW1hZ2VzL2FfZG90X2hhbS5qcGVnCiAgIwogICMgRm9yIGRldGFpbHMgcmVnYXJkaW5nIGFzc2V0cyBpbiBwYWNrYWdlcywgc2VlCiAgIyBodHRwczovL2ZsdXR0ZXIuZGV2L2Fzc2V0cy1hbmQtaW1hZ2VzLyNmcm9tLXBhY2thZ2VzCiAgIwogICMgQW4gaW1hZ2UgYXNzZXQgY2FuIHJlZmVyIHRvIG9uZSBvciBtb3JlIHJlc29sdXRpb24tc3BlY2lmaWMgInZhcmlhbnRzIiwgc2VlCiAgIyBodHRwczovL2ZsdXR0ZXIuZGV2L2Fzc2V0cy1hbmQtaW1hZ2VzLyNyZXNvbHV0aW9uLWF3YXJlCgogICMgVG8gYWRkIGN1c3RvbSBmb250cyB0byB5b3VyIHBhY2thZ2UsIGFkZCBhIGZvbnRzIHNlY3Rpb24gaGVyZSwKICAjIGluIHRoaXMgImZsdXR0ZXIiIHNlY3Rpb24uIEVhY2ggZW50cnkgaW4gdGhpcyBsaXN0IHNob3VsZCBoYXZlIGEKICAjICJmYW1pbHkiIGtleSB3aXRoIHRoZSBmb250IGZhbWlseSBuYW1lLCBhbmQgYSAiZm9udHMiIGtleSB3aXRoIGEKICAjIGxpc3QgZ2l2aW5nIHRoZSBhc3NldCBhbmQgb3RoZXIgZGVzY3JpcHRvcnMgZm9yIHRoZSBmb250LiBGb3IKICAjIGV4YW1wbGU6CiAgIyBmb250czoKICAjICAgLSBmYW1pbHk6IFNjaHlsZXIKICAjICAgICBmb250czoKICAjICAgICAgIC0gYXNzZXQ6IGZvbnRzL1NjaHlsZXItUmVndWxhci50dGYKICAjICAgICAgIC0gYXNzZXQ6IGZvbnRzL1NjaHlsZXItSXRhbGljLnR0ZgogICMgICAgICAgICBzdHlsZTogaXRhbGljCiAgIyAgIC0gZmFtaWx5OiBUcmFqYW4gUHJvCiAgIyAgICAgZm9udHM6CiAgIyAgICAgICAtIGFzc2V0OiBmb250cy9UcmFqYW5Qcm8udHRmCiAgIyAgICAgICAtIGFzc2V0OiBmb250cy9UcmFqYW5Qcm9fQm9sZC50dGYKICAjICAgICAgICAgd2VpZ2h0OiA3MDAKICAjCiAgIyBGb3IgZGV0YWlscyByZWdhcmRpbmcgZm9udHMgaW4gcGFja2FnZXMsIHNlZQogICMgaHR0cHM6Ly9mbHV0dGVyLmRldi9jdXN0b20tZm9udHMvI2Zyb20tcGFja2FnZXMKe3svaW9zfX17eyNsaW51eH19bmFtZToge3twcm9qZWN0X25hbWV9fV91aV9saW51eApkZXNjcmlwdGlvbjogV2lkZ2V0cyBhbmQgdGhlbWVzIGltcGxlbWVudGluZyB7e3Byb2plY3RfbmFtZS50aXRsZUNhc2UoKX19IGRlc2lnbiBsYW5ndWFnZSBmb3IgTGludXguCnZlcnNpb246IDAuMC4xCnB1Ymxpc2hfdG86IG5vbmUKCmVudmlyb25tZW50OgogIHNkazogIj49My4wLjAgPDQuMC4wIgogIGZsdXR0ZXI6ICI+PTMuMTAuMCIKCmRlcGVuZGVuY2llczoKICBmbHV0dGVyOgogICAgc2RrOiBmbHV0dGVyCiAgZmx1dHRlcl9sb2NhbGl6YXRpb25zOgogICAgc2RrOiBmbHV0dGVyCiAgdGhlbWVfdGFpbG9yX2Fubm90YXRpb246IF4yLjAuMAogIHlhcnU6IF4wLjguMAogIHlhcnVfaWNvbnM6IF4xLjAuNAogIHt7cHJvamVjdF9uYW1lfX1fdWk6CgpkZXZfZGVwZW5kZW5jaWVzOgogIGZsdXR0ZXJfdGVzdDoKICAgIHNkazogZmx1dHRlcgogIGFsY2hlbWlzdDogXjAuNi4xCiAgYnVpbGRfcnVubmVyOiBeMi40LjUKICB0aGVtZV90YWlsb3I6IF4yLjAuMAogIGZsdXR0ZXJfZ2VuX3J1bm5lcjogXjUuMy4xCiAgZmx1dHRlcl9saW50czogXjIuMC4xCgojIFRoZSBmb2xsb3dpbmcgc2VjdGlvbiBpcyBzcGVjaWZpYyB0byB0aGUgRmx1dHRlckdlbiBwYWNrYWdlLgojIFNlZTogaHR0cHM6Ly9wdWIuZGV2L3BhY2thZ2VzL2ZsdXR0ZXJfZ2VuI2NvbmZpZ3VyYXRpb24tZmlsZQpmbHV0dGVyX2dlbjoKICBvdXRwdXQ6IGxpYi9zcmMvCiAgIyBUbyBhZGQgc3VwcG9ydCBmb3Igc3ZnIHZpYSBmbHV0dGVyX3N2ZyBwYWNrYWdlCiAgIyBpbnRlZ3JhdGlvbnM6CiAgIyAgIGZsdXR0ZXJfc3ZnOiB0cnVlCgogIGFzc2V0czoKICAgIG91dHB1dHM6CiAgICAgIHBhY2thZ2VfcGFyYW1ldGVyX2VuYWJsZWQ6IHRydWUKCiMgVGhlIGZvbGxvd2luZyBzZWN0aW9uIGlzIHNwZWNpZmljIHRvIEZsdXR0ZXIgcGFja2FnZXMuCmZsdXR0ZXI6CiAgIyBUaGUgZm9sbG93aW5nIGxpbmUgZW5zdXJlcyB0aGF0IHRoZSBNYXRlcmlhbCBJY29ucyBmb250IGlzCiAgIyBpbmNsdWRlZCB3aXRoIHlvdXIgYXBwbGljYXRpb24sIHNvIHRoYXQgeW91IGNhbiB1c2UgdGhlIGljb25zIGluCiAgIyB0aGUgbWF0ZXJpYWwgSWNvbnMgY2xhc3MuCiAgdXNlcy1tYXRlcmlhbC1kZXNpZ246IHRydWUKICAjIFRvIGFkZCBhc3NldHMgdG8geW91ciBwYWNrYWdlLCBhZGQgYW4gYXNzZXRzIHNlY3Rpb24sIGxpa2UgdGhpczoKICAjIGFzc2V0czoKICAjICAgLSBpbWFnZXMvYV9kb3RfYnVyci5qcGVnCiAgIyAgIC0gaW1hZ2VzL2FfZG90X2hhbS5qcGVnCiAgIwogICMgRm9yIGRldGFpbHMgcmVnYXJkaW5nIGFzc2V0cyBpbiBwYWNrYWdlcywgc2VlCiAgIyBodHRwczovL2ZsdXR0ZXIuZGV2L2Fzc2V0cy1hbmQtaW1hZ2VzLyNmcm9tLXBhY2thZ2VzCiAgIwogICMgQW4gaW1hZ2UgYXNzZXQgY2FuIHJlZmVyIHRvIG9uZSBvciBtb3JlIHJlc29sdXRpb24tc3BlY2lmaWMgInZhcmlhbnRzIiwgc2VlCiAgIyBodHRwczovL2ZsdXR0ZXIuZGV2L2Fzc2V0cy1hbmQtaW1hZ2VzLyNyZXNvbHV0aW9uLWF3YXJlCgogICMgVG8gYWRkIGN1c3RvbSBmb250cyB0byB5b3VyIHBhY2thZ2UsIGFkZCBhIGZvbnRzIHNlY3Rpb24gaGVyZSwKICAjIGluIHRoaXMgImZsdXR0ZXIiIHNlY3Rpb24uIEVhY2ggZW50cnkgaW4gdGhpcyBsaXN0IHNob3VsZCBoYXZlIGEKICAjICJmYW1pbHkiIGtleSB3aXRoIHRoZSBmb250IGZhbWlseSBuYW1lLCBhbmQgYSAiZm9udHMiIGtleSB3aXRoIGEKICAjIGxpc3QgZ2l2aW5nIHRoZSBhc3NldCBhbmQgb3RoZXIgZGVzY3JpcHRvcnMgZm9yIHRoZSBmb250LiBGb3IKICAjIGV4YW1wbGU6CiAgIyBmb250czoKICAjICAgLSBmYW1pbHk6IFNjaHlsZXIKICAjICAgICBmb250czoKICAjICAgICAgIC0gYXNzZXQ6IGZvbnRzL1NjaHlsZXItUmVndWxhci50dGYKICAjICAgICAgIC0gYXNzZXQ6IGZvbnRzL1NjaHlsZXItSXRhbGljLnR0ZgogICMgICAgICAgICBzdHlsZTogaXRhbGljCiAgIyAgIC0gZmFtaWx5OiBUcmFqYW4gUHJvCiAgIyAgICAgZm9udHM6CiAgIyAgICAgICAtIGFzc2V0OiBmb250cy9UcmFqYW5Qcm8udHRmCiAgIyAgICAgICAtIGFzc2V0OiBmb250cy9UcmFqYW5Qcm9fQm9sZC50dGYKICAjICAgICAgICAgd2VpZ2h0OiA3MDAKICAjCiAgIyBGb3IgZGV0YWlscyByZWdhcmRpbmcgZm9udHMgaW4gcGFja2FnZXMsIHNlZQogICMgaHR0cHM6Ly9mbHV0dGVyLmRldi9jdXN0b20tZm9udHMvI2Zyb20tcGFja2FnZXMKe3svbGludXh9fXt7I21hY29zfX1uYW1lOiB7e3Byb2plY3RfbmFtZX19X3VpX21hY29zCmRlc2NyaXB0aW9uOiBXaWRnZXRzIGFuZCB0aGVtZXMgaW1wbGVtZW50aW5nIHt7cHJvamVjdF9uYW1lLnRpdGxlQ2FzZSgpfX0gZGVzaWduIGxhbmd1YWdlIGZvciBtYWNPUy4KdmVyc2lvbjogMC4wLjEKcHVibGlzaF90bzogbm9uZQoKZW52aXJvbm1lbnQ6CiAgc2RrOiAiPj0zLjAuMCA8NC4wLjAiCiAgZmx1dHRlcjogIj49My4xMC4wIgoKZGVwZW5kZW5jaWVzOgogIGZsdXR0ZXI6CiAgICBzZGs6IGZsdXR0ZXIKICBmbHV0dGVyX2xvY2FsaXphdGlvbnM6CiAgICBzZGs6IGZsdXR0ZXIKICB0aGVtZV90YWlsb3JfYW5ub3RhdGlvbjogXjIuMC4wCiAgY3VwZXJ0aW5vX2ljb25zOiBeMS4wLjUKICBtYWNvc191aToKICAgIGdpdDoKICAgICAgdXJsOiBodHRwczovL2dpdGh1Yi5jb20vbWFjb3N1aS9tYWNvc191aS5naXQKICAgICAgcmVmOiBmMTA1M2ZjMzE5NThkYmFhMzYyYTBhNGM4Mjg4ZThjZjA5MTlhYjhjCiAge3twcm9qZWN0X25hbWV9fV91aToKCmRldl9kZXBlbmRlbmNpZXM6CiAgZmx1dHRlcl90ZXN0OgogICAgc2RrOiBmbHV0dGVyCiAgYWxjaGVtaXN0OiBeMC42LjEKICBidWlsZF9ydW5uZXI6IF4yLjQuNQogIHRoZW1lX3RhaWxvcjogXjIuMC4wCiAgZmx1dHRlcl9nZW5fcnVubmVyOiBeNS4zLjEKICBmbHV0dGVyX2xpbnRzOiBeMi4wLjEKCiMgVGhlIGZvbGxvd2luZyBzZWN0aW9uIGlzIHNwZWNpZmljIHRvIHRoZSBGbHV0dGVyR2VuIHBhY2thZ2UuCiMgU2VlOiBodHRwczovL3B1Yi5kZXYvcGFja2FnZXMvZmx1dHRlcl9nZW4jY29uZmlndXJhdGlvbi1maWxlCmZsdXR0ZXJfZ2VuOgogIG91dHB1dDogbGliL3NyYy8KICAjIFRvIGFkZCBzdXBwb3J0IGZvciBzdmcgdmlhIGZsdXR0ZXJfc3ZnIHBhY2thZ2UKICAjIGludGVncmF0aW9uczoKICAjICAgZmx1dHRlcl9zdmc6IHRydWUKCiAgYXNzZXRzOgogICAgb3V0cHV0czoKICAgICAgcGFja2FnZV9wYXJhbWV0ZXJfZW5hYmxlZDogdHJ1ZQoKIyBUaGUgZm9sbG93aW5nIHNlY3Rpb24gaXMgc3BlY2lmaWMgdG8gRmx1dHRlciBwYWNrYWdlcy4KZmx1dHRlcjoKCiAgIyBUbyBhZGQgYXNzZXRzIHRvIHlvdXIgcGFja2FnZSwgYWRkIGFuIGFzc2V0cyBzZWN0aW9uLCBsaWtlIHRoaXM6CiAgIyBhc3NldHM6CiAgIyAgIC0gaW1hZ2VzL2FfZG90X2J1cnIuanBlZwogICMgICAtIGltYWdlcy9hX2RvdF9oYW0uanBlZwogICMKICAjIEZvciBkZXRhaWxzIHJlZ2FyZGluZyBhc3NldHMgaW4gcGFja2FnZXMsIHNlZQogICMgaHR0cHM6Ly9mbHV0dGVyLmRldi9hc3NldHMtYW5kLWltYWdlcy8jZnJvbS1wYWNrYWdlcwogICMKICAjIEFuIGltYWdlIGFzc2V0IGNhbiByZWZlciB0byBvbmUgb3IgbW9yZSByZXNvbHV0aW9uLXNwZWNpZmljICJ2YXJpYW50cyIsIHNlZQogICMgaHR0cHM6Ly9mbHV0dGVyLmRldi9hc3NldHMtYW5kLWltYWdlcy8jcmVzb2x1dGlvbi1hd2FyZQoKICAjIFRvIGFkZCBjdXN0b20gZm9udHMgdG8geW91ciBwYWNrYWdlLCBhZGQgYSBmb250cyBzZWN0aW9uIGhlcmUsCiAgIyBpbiB0aGlzICJmbHV0dGVyIiBzZWN0aW9uLiBFYWNoIGVudHJ5IGluIHRoaXMgbGlzdCBzaG91bGQgaGF2ZSBhCiAgIyAiZmFtaWx5IiBrZXkgd2l0aCB0aGUgZm9udCBmYW1pbHkgbmFtZSwgYW5kIGEgImZvbnRzIiBrZXkgd2l0aCBhCiAgIyBsaXN0IGdpdmluZyB0aGUgYXNzZXQgYW5kIG90aGVyIGRlc2NyaXB0b3JzIGZvciB0aGUgZm9udC4gRm9yCiAgIyBleGFtcGxlOgogICMgZm9udHM6CiAgIyAgIC0gZmFtaWx5OiBTY2h5bGVyCiAgIyAgICAgZm9udHM6CiAgIyAgICAgICAtIGFzc2V0OiBmb250cy9TY2h5bGVyLVJlZ3VsYXIudHRmCiAgIyAgICAgICAtIGFzc2V0OiBmb250cy9TY2h5bGVyLUl0YWxpYy50dGYKICAjICAgICAgICAgc3R5bGU6IGl0YWxpYwogICMgICAtIGZhbWlseTogVHJhamFuIFBybwogICMgICAgIGZvbnRzOgogICMgICAgICAgLSBhc3NldDogZm9udHMvVHJhamFuUHJvLnR0ZgogICMgICAgICAgLSBhc3NldDogZm9udHMvVHJhamFuUHJvX0JvbGQudHRmCiAgIyAgICAgICAgIHdlaWdodDogNzAwCiAgIwogICMgRm9yIGRldGFpbHMgcmVnYXJkaW5nIGZvbnRzIGluIHBhY2thZ2VzLCBzZWUKICAjIGh0dHBzOi8vZmx1dHRlci5kZXYvY3VzdG9tLWZvbnRzLyNmcm9tLXBhY2thZ2VzCnt7L21hY29zfX17eyN3ZWJ9fW5hbWU6IHt7cHJvamVjdF9uYW1lfX1fdWlfd2ViCmRlc2NyaXB0aW9uOiBXaWRnZXRzIGFuZCB0aGVtZXMgaW1wbGVtZW50aW5nIHt7cHJvamVjdF9uYW1lLnRpdGxlQ2FzZSgpfX0gZGVzaWduIGxhbmd1YWdlIGZvciBXZWIuCnZlcnNpb246IDAuMC4xCnB1Ymxpc2hfdG86IG5vbmUKCmVudmlyb25tZW50OgogIHNkazogIj49My4wLjAgPDQuMC4wIgogIGZsdXR0ZXI6ICI+PTMuMTAuMCIKCmRlcGVuZGVuY2llczoKICBmbHV0dGVyOgogICAgc2RrOiBmbHV0dGVyCiAgZmx1dHRlcl9sb2NhbGl6YXRpb25zOgogICAgc2RrOiBmbHV0dGVyCiAgdGhlbWVfdGFpbG9yX2Fubm90YXRpb246IF4yLjAuMAogIHt7cHJvamVjdF9uYW1lfX1fdWk6CgpkZXZfZGVwZW5kZW5jaWVzOgogIGZsdXR0ZXJfdGVzdDoKICAgIHNkazogZmx1dHRlcgogIGFsY2hlbWlzdDogXjAuNi4xCiAgYnVpbGRfcnVubmVyOiBeMi40LjUKICB0aGVtZV90YWlsb3I6IF4yLjAuMAogIGZsdXR0ZXJfZ2VuX3J1bm5lcjogXjUuMy4xCiAgZmx1dHRlcl9saW50czogXjIuMC4xCgojIFRoZSBmb2xsb3dpbmcgc2VjdGlvbiBpcyBzcGVjaWZpYyB0byB0aGUgRmx1dHRlckdlbiBwYWNrYWdlLgojIFNlZTogaHR0cHM6Ly9wdWIuZGV2L3BhY2thZ2VzL2ZsdXR0ZXJfZ2VuI2NvbmZpZ3VyYXRpb24tZmlsZQpmbHV0dGVyX2dlbjoKICBvdXRwdXQ6IGxpYi9zcmMvCiAgIyBUbyBhZGQgc3VwcG9ydCBmb3Igc3ZnIHZpYSBmbHV0dGVyX3N2ZyBwYWNrYWdlCiAgIyBpbnRlZ3JhdGlvbnM6CiAgIyAgIGZsdXR0ZXJfc3ZnOiB0cnVlCgogIGFzc2V0czoKICAgIG91dHB1dHM6CiAgICAgIHBhY2thZ2VfcGFyYW1ldGVyX2VuYWJsZWQ6IHRydWUKCiMgVGhlIGZvbGxvd2luZyBzZWN0aW9uIGlzIHNwZWNpZmljIHRvIEZsdXR0ZXIgcGFja2FnZXMuCmZsdXR0ZXI6CiAgIyBUaGUgZm9sbG93aW5nIGxpbmUgZW5zdXJlcyB0aGF0IHRoZSBNYXRlcmlhbCBJY29ucyBmb250IGlzCiAgIyBpbmNsdWRlZCB3aXRoIHlvdXIgYXBwbGljYXRpb24sIHNvIHRoYXQgeW91IGNhbiB1c2UgdGhlIGljb25zIGluCiAgIyB0aGUgbWF0ZXJpYWwgSWNvbnMgY2xhc3MuCiAgdXNlcy1tYXRlcmlhbC1kZXNpZ246IHRydWUKICAjIFRvIGFkZCBhc3NldHMgdG8geW91ciBwYWNrYWdlLCBhZGQgYW4gYXNzZXRzIHNlY3Rpb24sIGxpa2UgdGhpczoKICAjIGFzc2V0czoKICAjICAgLSBpbWFnZXMvYV9kb3RfYnVyci5qcGVnCiAgIyAgIC0gaW1hZ2VzL2FfZG90X2hhbS5qcGVnCiAgIwogICMgRm9yIGRldGFpbHMgcmVnYXJkaW5nIGFzc2V0cyBpbiBwYWNrYWdlcywgc2VlCiAgIyBodHRwczovL2ZsdXR0ZXIuZGV2L2Fzc2V0cy1hbmQtaW1hZ2VzLyNmcm9tLXBhY2thZ2VzCiAgIwogICMgQW4gaW1hZ2UgYXNzZXQgY2FuIHJlZmVyIHRvIG9uZSBvciBtb3JlIHJlc29sdXRpb24tc3BlY2lmaWMgInZhcmlhbnRzIiwgc2VlCiAgIyBodHRwczovL2ZsdXR0ZXIuZGV2L2Fzc2V0cy1hbmQtaW1hZ2VzLyNyZXNvbHV0aW9uLWF3YXJlCgogICMgVG8gYWRkIGN1c3RvbSBmb250cyB0byB5b3VyIHBhY2thZ2UsIGFkZCBhIGZvbnRzIHNlY3Rpb24gaGVyZSwKICAjIGluIHRoaXMgImZsdXR0ZXIiIHNlY3Rpb24uIEVhY2ggZW50cnkgaW4gdGhpcyBsaXN0IHNob3VsZCBoYXZlIGEKICAjICJmYW1pbHkiIGtleSB3aXRoIHRoZSBmb250IGZhbWlseSBuYW1lLCBhbmQgYSAiZm9udHMiIGtleSB3aXRoIGEKICAjIGxpc3QgZ2l2aW5nIHRoZSBhc3NldCBhbmQgb3RoZXIgZGVzY3JpcHRvcnMgZm9yIHRoZSBmb250LiBGb3IKICAjIGV4YW1wbGU6CiAgIyBmb250czoKICAjICAgLSBmYW1pbHk6IFNjaHlsZXIKICAjICAgICBmb250czoKICAjICAgICAgIC0gYXNzZXQ6IGZvbnRzL1NjaHlsZXItUmVndWxhci50dGYKICAjICAgICAgIC0gYXNzZXQ6IGZvbnRzL1NjaHlsZXItSXRhbGljLnR0ZgogICMgICAgICAgICBzdHlsZTogaXRhbGljCiAgIyAgIC0gZmFtaWx5OiBUcmFqYW4gUHJvCiAgIyAgICAgZm9udHM6CiAgIyAgICAgICAtIGFzc2V0OiBmb250cy9UcmFqYW5Qcm8udHRmCiAgIyAgICAgICAtIGFzc2V0OiBmb250cy9UcmFqYW5Qcm9fQm9sZC50dGYKICAjICAgICAgICAgd2VpZ2h0OiA3MDAKICAjCiAgIyBGb3IgZGV0YWlscyByZWdhcmRpbmcgZm9udHMgaW4gcGFja2FnZXMsIHNlZQogICMgaHR0cHM6Ly9mbHV0dGVyLmRldi9jdXN0b20tZm9udHMvI2Zyb20tcGFja2FnZXMKe3svd2VifX17eyN3aW5kb3dzfX1uYW1lOiB7e3Byb2plY3RfbmFtZX19X3VpX3dpbmRvd3MKZGVzY3JpcHRpb246IFdpZGdldHMgYW5kIHRoZW1lcyBpbXBsZW1lbnRpbmcge3twcm9qZWN0X25hbWUudGl0bGVDYXNlKCl9fSBkZXNpZ24gbGFuZ3VhZ2UgZm9yIFdpbmRvd3MuCnZlcnNpb246IDAuMC4xCnB1Ymxpc2hfdG86IG5vbmUKCmVudmlyb25tZW50OgogIHNkazogIj49My4wLjAgPDQuMC4wIgogIGZsdXR0ZXI6ICI+PTMuMTAuMCIKCmRlcGVuZGVuY2llczoKICBmbHV0dGVyOgogICAgc2RrOiBmbHV0dGVyCiAgZmx1dHRlcl9sb2NhbGl6YXRpb25zOgogICAgc2RrOiBmbHV0dGVyCiAgdGhlbWVfdGFpbG9yX2Fubm90YXRpb246IF4yLjAuMAogIGZsdWVudF91aTogXjQuNi4yCiAge3twcm9qZWN0X25hbWV9fV91aToKCmRldl9kZXBlbmRlbmNpZXM6CiAgZmx1dHRlcl90ZXN0OgogICAgc2RrOiBmbHV0dGVyCiAgYWxjaGVtaXN0OiBeMC42LjEKICBidWlsZF9ydW5uZXI6IF4yLjQuNQogIHRoZW1lX3RhaWxvcjogXjIuMC4wCiAgZmx1dHRlcl9nZW5fcnVubmVyOiBeNS4zLjEKICBmbHV0dGVyX2xpbnRzOiBeMi4wLjEKCiMgVGhlIGZvbGxvd2luZyBzZWN0aW9uIGlzIHNwZWNpZmljIHRvIHRoZSBGbHV0dGVyR2VuIHBhY2thZ2UuCiMgU2VlOiBodHRwczovL3B1Yi5kZXYvcGFja2FnZXMvZmx1dHRlcl9nZW4jY29uZmlndXJhdGlvbi1maWxlCmZsdXR0ZXJfZ2VuOgogIG91dHB1dDogbGliL3NyYy8KICAjIFRvIGFkZCBzdXBwb3J0IGZvciBzdmcgdmlhIGZsdXR0ZXJfc3ZnIHBhY2thZ2UKICAjIGludGVncmF0aW9uczoKICAjICAgZmx1dHRlcl9zdmc6IHRydWUKCiAgYXNzZXRzOgogICAgb3V0cHV0czoKICAgICAgcGFja2FnZV9wYXJhbWV0ZXJfZW5hYmxlZDogdHJ1ZQoKIyBUaGUgZm9sbG93aW5nIHNlY3Rpb24gaXMgc3BlY2lmaWMgdG8gRmx1dHRlciBwYWNrYWdlcy4KZmx1dHRlcjoKICAjIFRoZSBmb2xsb3dpbmcgbGluZSBlbnN1cmVzIHRoYXQgdGhlIE1hdGVyaWFsIEljb25zIGZvbnQgaXMKICAjIGluY2x1ZGVkIHdpdGggeW91ciBhcHBsaWNhdGlvbiwgc28gdGhhdCB5b3UgY2FuIHVzZSB0aGUgaWNvbnMgaW4KICAjIHRoZSBtYXRlcmlhbCBJY29ucyBjbGFzcy4KICB1c2VzLW1hdGVyaWFsLWRlc2lnbjogdHJ1ZQogICMgVG8gYWRkIGFzc2V0cyB0byB5b3VyIHBhY2thZ2UsIGFkZCBhbiBhc3NldHMgc2VjdGlvbiwgbGlrZSB0aGlzOgogICMgYXNzZXRzOgogICMgICAtIGltYWdlcy9hX2RvdF9idXJyLmpwZWcKICAjICAgLSBpbWFnZXMvYV9kb3RfaGFtLmpwZWcKICAjCiAgIyBGb3IgZGV0YWlscyByZWdhcmRpbmcgYXNzZXRzIGluIHBhY2thZ2VzLCBzZWUKICAjIGh0dHBzOi8vZmx1dHRlci5kZXYvYXNzZXRzLWFuZC1pbWFnZXMvI2Zyb20tcGFja2FnZXMKICAjCiAgIyBBbiBpbWFnZSBhc3NldCBjYW4gcmVmZXIgdG8gb25lIG9yIG1vcmUgcmVzb2x1dGlvbi1zcGVjaWZpYyAidmFyaWFudHMiLCBzZWUKICAjIGh0dHBzOi8vZmx1dHRlci5kZXYvYXNzZXRzLWFuZC1pbWFnZXMvI3Jlc29sdXRpb24tYXdhcmUKCiAgIyBUbyBhZGQgY3VzdG9tIGZvbnRzIHRvIHlvdXIgcGFja2FnZSwgYWRkIGEgZm9udHMgc2VjdGlvbiBoZXJlLAogICMgaW4gdGhpcyAiZmx1dHRlciIgc2VjdGlvbi4gRWFjaCBlbnRyeSBpbiB0aGlzIGxpc3Qgc2hvdWxkIGhhdmUgYQogICMgImZhbWlseSIga2V5IHdpdGggdGhlIGZvbnQgZmFtaWx5IG5hbWUsIGFuZCBhICJmb250cyIga2V5IHdpdGggYQogICMgbGlzdCBnaXZpbmcgdGhlIGFzc2V0IGFuZCBvdGhlciBkZXNjcmlwdG9ycyBmb3IgdGhlIGZvbnQuIEZvcgogICMgZXhhbXBsZToKICAjIGZvbnRzOgogICMgICAtIGZhbWlseTogU2NoeWxlcgogICMgICAgIGZvbnRzOgogICMgICAgICAgLSBhc3NldDogZm9udHMvU2NoeWxlci1SZWd1bGFyLnR0ZgogICMgICAgICAgLSBhc3NldDogZm9udHMvU2NoeWxlci1JdGFsaWMudHRmCiAgIyAgICAgICAgIHN0eWxlOiBpdGFsaWMKICAjICAgLSBmYW1pbHk6IFRyYWphbiBQcm8KICAjICAgICBmb250czoKICAjICAgICAgIC0gYXNzZXQ6IGZvbnRzL1RyYWphblByby50dGYKICAjICAgICAgIC0gYXNzZXQ6IGZvbnRzL1RyYWphblByb19Cb2xkLnR0ZgogICMgICAgICAgICB3ZWlnaHQ6IDcwMAogICMKICAjIEZvciBkZXRhaWxzIHJlZ2FyZGluZyBmb250cyBpbiBwYWNrYWdlcywgc2VlCiAgIyBodHRwczovL2ZsdXR0ZXIuZGV2L2N1c3RvbS1mb250cy8jZnJvbS1wYWNrYWdlcwp7ey93aW5kb3dzfX17eyNtb2JpbGV9fW5hbWU6IHt7cHJvamVjdF9uYW1lfX1fdWlfbW9iaWxlCmRlc2NyaXB0aW9uOiBXaWRnZXRzIGFuZCB0aGVtZXMgaW1wbGVtZW50aW5nIHt7cHJvamVjdF9uYW1lLnRpdGxlQ2FzZSgpfX0gZGVzaWduIGxhbmd1YWdlIGZvciBNb2JpbGUuCnZlcnNpb246IDAuMC4xCnB1Ymxpc2hfdG86IG5vbmUKCmVudmlyb25tZW50OgogIHNkazogIj49My4wLjAgPDQuMC4wIgogIGZsdXR0ZXI6ICI+PTMuMTAuMCIKCmRlcGVuZGVuY2llczoKICBmbHV0dGVyOgogICAgc2RrOiBmbHV0dGVyCiAgZmx1dHRlcl9sb2NhbGl6YXRpb25zOgogICAgc2RrOiBmbHV0dGVyCiAgdGhlbWVfdGFpbG9yX2Fubm90YXRpb246IF4yLjAuMAogIHt7cHJvamVjdF9uYW1lfX1fdWk6CgpkZXZfZGVwZW5kZW5jaWVzOgogIGZsdXR0ZXJfdGVzdDoKICAgIHNkazogZmx1dHRlcgogIGFsY2hlbWlzdDogXjAuNi4xCiAgYnVpbGRfcnVubmVyOiBeMi40LjUKICB0aGVtZV90YWlsb3I6IF4yLjAuMAogIGZsdXR0ZXJfZ2VuX3J1bm5lcjogXjUuMy4xCiAgZmx1dHRlcl9saW50czogXjIuMC4xCgojIFRoZSBmb2xsb3dpbmcgc2VjdGlvbiBpcyBzcGVjaWZpYyB0byB0aGUgRmx1dHRlckdlbiBwYWNrYWdlLgojIFNlZTogaHR0cHM6Ly9wdWIuZGV2L3BhY2thZ2VzL2ZsdXR0ZXJfZ2VuI2NvbmZpZ3VyYXRpb24tZmlsZQpmbHV0dGVyX2dlbjoKICBvdXRwdXQ6IGxpYi9zcmMvCiAgIyBUbyBhZGQgc3VwcG9ydCBmb3Igc3ZnIHZpYSBmbHV0dGVyX3N2ZyBwYWNrYWdlCiAgIyBpbnRlZ3JhdGlvbnM6CiAgIyAgIGZsdXR0ZXJfc3ZnOiB0cnVlCgogIGFzc2V0czoKICAgIG91dHB1dHM6CiAgICAgIHBhY2thZ2VfcGFyYW1ldGVyX2VuYWJsZWQ6IHRydWUKCiMgVGhlIGZvbGxvd2luZyBzZWN0aW9uIGlzIHNwZWNpZmljIHRvIEZsdXR0ZXIgcGFja2FnZXMuCmZsdXR0ZXI6CiAgIyBUaGUgZm9sbG93aW5nIGxpbmUgZW5zdXJlcyB0aGF0IHRoZSBNYXRlcmlhbCBJY29ucyBmb250IGlzCiAgIyBpbmNsdWRlZCB3aXRoIHlvdXIgYXBwbGljYXRpb24sIHNvIHRoYXQgeW91IGNhbiB1c2UgdGhlIGljb25zIGluCiAgIyB0aGUgbWF0ZXJpYWwgSWNvbnMgY2xhc3MuCiAgdXNlcy1tYXRlcmlhbC1kZXNpZ246IHRydWUKICAjIFRvIGFkZCBhc3NldHMgdG8geW91ciBwYWNrYWdlLCBhZGQgYW4gYXNzZXRzIHNlY3Rpb24sIGxpa2UgdGhpczoKICAjIGFzc2V0czoKICAjICAgLSBpbWFnZXMvYV9kb3RfYnVyci5qcGVnCiAgIyAgIC0gaW1hZ2VzL2FfZG90X2hhbS5qcGVnCiAgIwogICMgRm9yIGRldGFpbHMgcmVnYXJkaW5nIGFzc2V0cyBpbiBwYWNrYWdlcywgc2VlCiAgIyBodHRwczovL2ZsdXR0ZXIuZGV2L2Fzc2V0cy1hbmQtaW1hZ2VzLyNmcm9tLXBhY2thZ2VzCiAgIwogICMgQW4gaW1hZ2UgYXNzZXQgY2FuIHJlZmVyIHRvIG9uZSBvciBtb3JlIHJlc29sdXRpb24tc3BlY2lmaWMgInZhcmlhbnRzIiwgc2VlCiAgIyBodHRwczovL2ZsdXR0ZXIuZGV2L2Fzc2V0cy1hbmQtaW1hZ2VzLyNyZXNvbHV0aW9uLWF3YXJlCgogICMgVG8gYWRkIGN1c3RvbSBmb250cyB0byB5b3VyIHBhY2thZ2UsIGFkZCBhIGZvbnRzIHNlY3Rpb24gaGVyZSwKICAjIGluIHRoaXMgImZsdXR0ZXIiIHNlY3Rpb24uIEVhY2ggZW50cnkgaW4gdGhpcyBsaXN0IHNob3VsZCBoYXZlIGEKICAjICJmYW1pbHkiIGtleSB3aXRoIHRoZSBmb250IGZhbWlseSBuYW1lLCBhbmQgYSAiZm9udHMiIGtleSB3aXRoIGEKICAjIGxpc3QgZ2l2aW5nIHRoZSBhc3NldCBhbmQgb3RoZXIgZGVzY3JpcHRvcnMgZm9yIHRoZSBmb250LiBGb3IKICAjIGV4YW1wbGU6CiAgIyBmb250czoKICAjICAgLSBmYW1pbHk6IFNjaHlsZXIKICAjICAgICBmb250czoKICAjICAgICAgIC0gYXNzZXQ6IGZvbnRzL1NjaHlsZXItUmVndWxhci50dGYKICAjICAgICAgIC0gYXNzZXQ6IGZvbnRzL1NjaHlsZXItSXRhbGljLnR0ZgogICMgICAgICAgICBzdHlsZTogaXRhbGljCiAgIyAgIC0gZmFtaWx5OiBUcmFqYW4gUHJvCiAgIyAgICAgZm9udHM6CiAgIyAgICAgICAtIGFzc2V0OiBmb250cy9UcmFqYW5Qcm8udHRmCiAgIyAgICAgICAtIGFzc2V0OiBmb250cy9UcmFqYW5Qcm9fQm9sZC50dGYKICAjICAgICAgICAgd2VpZ2h0OiA3MDAKICAjCiAgIyBGb3IgZGV0YWlscyByZWdhcmRpbmcgZm9udHMgaW4gcGFja2FnZXMsIHNlZQogICMgaHR0cHM6Ly9mbHV0dGVyLmRldi9jdXN0b20tZm9udHMvI2Zyb20tcGFja2FnZXMKe3svbW9iaWxlfX0=",
+          "e3sjYW5kcm9pZH19bmFtZToge3twcm9qZWN0X25hbWV9fV91aV9hbmRyb2lkCmRlc2NyaXB0aW9uOiBXaWRnZXRzIGFuZCB0aGVtZXMgaW1wbGVtZW50aW5nIHt7cHJvamVjdF9uYW1lLnRpdGxlQ2FzZSgpfX0gZGVzaWduIGxhbmd1YWdlIGZvciBBbmRyb2lkLgp2ZXJzaW9uOiAwLjAuMQpwdWJsaXNoX3RvOiBub25lCgplbnZpcm9ubWVudDoKICBzZGs6ICI+PTMuMC4wIDw0LjAuMCIKICBmbHV0dGVyOiAiPj0zLjEwLjAiCgpkZXBlbmRlbmNpZXM6CiAgZmx1dHRlcjoKICAgIHNkazogZmx1dHRlcgogIGZsdXR0ZXJfbG9jYWxpemF0aW9uczoKICAgIHNkazogZmx1dHRlcgogIHRoZW1lX3RhaWxvcl9hbm5vdGF0aW9uOiBeMi4wLjAKICB7e3Byb2plY3RfbmFtZX19X3VpOgoKZGV2X2RlcGVuZGVuY2llczoKICBmbHV0dGVyX3Rlc3Q6CiAgICBzZGs6IGZsdXR0ZXIKICBhbGNoZW1pc3Q6IF4wLjYuMQogIGJ1aWxkX3J1bm5lcjogXjIuNC41CiAgdGhlbWVfdGFpbG9yOiBeMi4wLjAKICBmbHV0dGVyX2dlbl9ydW5uZXI6IF41LjMuMQogIGZsdXR0ZXJfbGludHM6IF4yLjAuMQoKIyBUaGUgZm9sbG93aW5nIHNlY3Rpb24gaXMgc3BlY2lmaWMgdG8gdGhlIEZsdXR0ZXJHZW4gcGFja2FnZS4KIyBTZWU6IGh0dHBzOi8vcHViLmRldi9wYWNrYWdlcy9mbHV0dGVyX2dlbiNjb25maWd1cmF0aW9uLWZpbGUKZmx1dHRlcl9nZW46CiAgb3V0cHV0OiBsaWIvc3JjLwogICMgVG8gYWRkIHN1cHBvcnQgZm9yIHN2ZyB2aWEgZmx1dHRlcl9zdmcgcGFja2FnZQogICMgaW50ZWdyYXRpb25zOgogICMgICBmbHV0dGVyX3N2ZzogdHJ1ZQoKICBhc3NldHM6CiAgICBvdXRwdXRzOgogICAgICBwYWNrYWdlX3BhcmFtZXRlcl9lbmFibGVkOiB0cnVlCgojIFRoZSBmb2xsb3dpbmcgc2VjdGlvbiBpcyBzcGVjaWZpYyB0byBGbHV0dGVyIHBhY2thZ2VzLgpmbHV0dGVyOgogICMgVGhlIGZvbGxvd2luZyBsaW5lIGVuc3VyZXMgdGhhdCB0aGUgTWF0ZXJpYWwgSWNvbnMgZm9udCBpcwogICMgaW5jbHVkZWQgd2l0aCB5b3VyIGFwcGxpY2F0aW9uLCBzbyB0aGF0IHlvdSBjYW4gdXNlIHRoZSBpY29ucyBpbgogICMgdGhlIG1hdGVyaWFsIEljb25zIGNsYXNzLgogIHVzZXMtbWF0ZXJpYWwtZGVzaWduOiB0cnVlCiAgIyBUbyBhZGQgYXNzZXRzIHRvIHlvdXIgcGFja2FnZSwgYWRkIGFuIGFzc2V0cyBzZWN0aW9uLCBsaWtlIHRoaXM6CiAgIyBhc3NldHM6CiAgIyAgIC0gaW1hZ2VzL2FfZG90X2J1cnIuanBlZwogICMgICAtIGltYWdlcy9hX2RvdF9oYW0uanBlZwogICMKICAjIEZvciBkZXRhaWxzIHJlZ2FyZGluZyBhc3NldHMgaW4gcGFja2FnZXMsIHNlZQogICMgaHR0cHM6Ly9mbHV0dGVyLmRldi9hc3NldHMtYW5kLWltYWdlcy8jZnJvbS1wYWNrYWdlcwogICMKICAjIEFuIGltYWdlIGFzc2V0IGNhbiByZWZlciB0byBvbmUgb3IgbW9yZSByZXNvbHV0aW9uLXNwZWNpZmljICJ2YXJpYW50cyIsIHNlZQogICMgaHR0cHM6Ly9mbHV0dGVyLmRldi9hc3NldHMtYW5kLWltYWdlcy8jcmVzb2x1dGlvbi1hd2FyZQoKICAjIFRvIGFkZCBjdXN0b20gZm9udHMgdG8geW91ciBwYWNrYWdlLCBhZGQgYSBmb250cyBzZWN0aW9uIGhlcmUsCiAgIyBpbiB0aGlzICJmbHV0dGVyIiBzZWN0aW9uLiBFYWNoIGVudHJ5IGluIHRoaXMgbGlzdCBzaG91bGQgaGF2ZSBhCiAgIyAiZmFtaWx5IiBrZXkgd2l0aCB0aGUgZm9udCBmYW1pbHkgbmFtZSwgYW5kIGEgImZvbnRzIiBrZXkgd2l0aCBhCiAgIyBsaXN0IGdpdmluZyB0aGUgYXNzZXQgYW5kIG90aGVyIGRlc2NyaXB0b3JzIGZvciB0aGUgZm9udC4gRm9yCiAgIyBleGFtcGxlOgogICMgZm9udHM6CiAgIyAgIC0gZmFtaWx5OiBTY2h5bGVyCiAgIyAgICAgZm9udHM6CiAgIyAgICAgICAtIGFzc2V0OiBmb250cy9TY2h5bGVyLVJlZ3VsYXIudHRmCiAgIyAgICAgICAtIGFzc2V0OiBmb250cy9TY2h5bGVyLUl0YWxpYy50dGYKICAjICAgICAgICAgc3R5bGU6IGl0YWxpYwogICMgICAtIGZhbWlseTogVHJhamFuIFBybwogICMgICAgIGZvbnRzOgogICMgICAgICAgLSBhc3NldDogZm9udHMvVHJhamFuUHJvLnR0ZgogICMgICAgICAgLSBhc3NldDogZm9udHMvVHJhamFuUHJvX0JvbGQudHRmCiAgIyAgICAgICAgIHdlaWdodDogNzAwCiAgIwogICMgRm9yIGRldGFpbHMgcmVnYXJkaW5nIGZvbnRzIGluIHBhY2thZ2VzLCBzZWUKICAjIGh0dHBzOi8vZmx1dHRlci5kZXYvY3VzdG9tLWZvbnRzLyNmcm9tLXBhY2thZ2VzCnt7L2FuZHJvaWR9fXt7I2lvc319bmFtZToge3twcm9qZWN0X25hbWV9fV91aV9pb3MKZGVzY3JpcHRpb246IFdpZGdldHMgYW5kIHRoZW1lcyBpbXBsZW1lbnRpbmcge3twcm9qZWN0X25hbWUudGl0bGVDYXNlKCl9fSBkZXNpZ24gbGFuZ3VhZ2UgZm9yIGlPUy4KdmVyc2lvbjogMC4wLjEKcHVibGlzaF90bzogbm9uZQoKZW52aXJvbm1lbnQ6CiAgc2RrOiAiPj0zLjAuMCA8NC4wLjAiCiAgZmx1dHRlcjogIj49My4xMC4wIgoKZGVwZW5kZW5jaWVzOgogIGZsdXR0ZXI6CiAgICBzZGs6IGZsdXR0ZXIKICBmbHV0dGVyX2xvY2FsaXphdGlvbnM6CiAgICBzZGs6IGZsdXR0ZXIKICB0aGVtZV90YWlsb3JfYW5ub3RhdGlvbjogXjIuMC4wCiAgY3VwZXJ0aW5vX2ljb25zOiBeMS4wLjUKICB7e3Byb2plY3RfbmFtZX19X3VpOgoKZGV2X2RlcGVuZGVuY2llczoKICBmbHV0dGVyX3Rlc3Q6CiAgICBzZGs6IGZsdXR0ZXIKICBhbGNoZW1pc3Q6IF4wLjYuMQogIGJ1aWxkX3J1bm5lcjogXjIuNC41CiAgdGhlbWVfdGFpbG9yOiBeMi4wLjAKICBmbHV0dGVyX2dlbl9ydW5uZXI6IF41LjMuMQogIGZsdXR0ZXJfbGludHM6IF4yLjAuMQoKIyBUaGUgZm9sbG93aW5nIHNlY3Rpb24gaXMgc3BlY2lmaWMgdG8gdGhlIEZsdXR0ZXJHZW4gcGFja2FnZS4KIyBTZWU6IGh0dHBzOi8vcHViLmRldi9wYWNrYWdlcy9mbHV0dGVyX2dlbiNjb25maWd1cmF0aW9uLWZpbGUKZmx1dHRlcl9nZW46CiAgb3V0cHV0OiBsaWIvc3JjLwogICMgVG8gYWRkIHN1cHBvcnQgZm9yIHN2ZyB2aWEgZmx1dHRlcl9zdmcgcGFja2FnZQogICMgaW50ZWdyYXRpb25zOgogICMgICBmbHV0dGVyX3N2ZzogdHJ1ZQoKICBhc3NldHM6CiAgICBvdXRwdXRzOgogICAgICBwYWNrYWdlX3BhcmFtZXRlcl9lbmFibGVkOiB0cnVlCgojIFRoZSBmb2xsb3dpbmcgc2VjdGlvbiBpcyBzcGVjaWZpYyB0byBGbHV0dGVyIHBhY2thZ2VzLgpmbHV0dGVyOgoKICAjIFRvIGFkZCBhc3NldHMgdG8geW91ciBwYWNrYWdlLCBhZGQgYW4gYXNzZXRzIHNlY3Rpb24sIGxpa2UgdGhpczoKICAjIGFzc2V0czoKICAjICAgLSBpbWFnZXMvYV9kb3RfYnVyci5qcGVnCiAgIyAgIC0gaW1hZ2VzL2FfZG90X2hhbS5qcGVnCiAgIwogICMgRm9yIGRldGFpbHMgcmVnYXJkaW5nIGFzc2V0cyBpbiBwYWNrYWdlcywgc2VlCiAgIyBodHRwczovL2ZsdXR0ZXIuZGV2L2Fzc2V0cy1hbmQtaW1hZ2VzLyNmcm9tLXBhY2thZ2VzCiAgIwogICMgQW4gaW1hZ2UgYXNzZXQgY2FuIHJlZmVyIHRvIG9uZSBvciBtb3JlIHJlc29sdXRpb24tc3BlY2lmaWMgInZhcmlhbnRzIiwgc2VlCiAgIyBodHRwczovL2ZsdXR0ZXIuZGV2L2Fzc2V0cy1hbmQtaW1hZ2VzLyNyZXNvbHV0aW9uLWF3YXJlCgogICMgVG8gYWRkIGN1c3RvbSBmb250cyB0byB5b3VyIHBhY2thZ2UsIGFkZCBhIGZvbnRzIHNlY3Rpb24gaGVyZSwKICAjIGluIHRoaXMgImZsdXR0ZXIiIHNlY3Rpb24uIEVhY2ggZW50cnkgaW4gdGhpcyBsaXN0IHNob3VsZCBoYXZlIGEKICAjICJmYW1pbHkiIGtleSB3aXRoIHRoZSBmb250IGZhbWlseSBuYW1lLCBhbmQgYSAiZm9udHMiIGtleSB3aXRoIGEKICAjIGxpc3QgZ2l2aW5nIHRoZSBhc3NldCBhbmQgb3RoZXIgZGVzY3JpcHRvcnMgZm9yIHRoZSBmb250LiBGb3IKICAjIGV4YW1wbGU6CiAgIyBmb250czoKICAjICAgLSBmYW1pbHk6IFNjaHlsZXIKICAjICAgICBmb250czoKICAjICAgICAgIC0gYXNzZXQ6IGZvbnRzL1NjaHlsZXItUmVndWxhci50dGYKICAjICAgICAgIC0gYXNzZXQ6IGZvbnRzL1NjaHlsZXItSXRhbGljLnR0ZgogICMgICAgICAgICBzdHlsZTogaXRhbGljCiAgIyAgIC0gZmFtaWx5OiBUcmFqYW4gUHJvCiAgIyAgICAgZm9udHM6CiAgIyAgICAgICAtIGFzc2V0OiBmb250cy9UcmFqYW5Qcm8udHRmCiAgIyAgICAgICAtIGFzc2V0OiBmb250cy9UcmFqYW5Qcm9fQm9sZC50dGYKICAjICAgICAgICAgd2VpZ2h0OiA3MDAKICAjCiAgIyBGb3IgZGV0YWlscyByZWdhcmRpbmcgZm9udHMgaW4gcGFja2FnZXMsIHNlZQogICMgaHR0cHM6Ly9mbHV0dGVyLmRldi9jdXN0b20tZm9udHMvI2Zyb20tcGFja2FnZXMKe3svaW9zfX17eyNsaW51eH19bmFtZToge3twcm9qZWN0X25hbWV9fV91aV9saW51eApkZXNjcmlwdGlvbjogV2lkZ2V0cyBhbmQgdGhlbWVzIGltcGxlbWVudGluZyB7e3Byb2plY3RfbmFtZS50aXRsZUNhc2UoKX19IGRlc2lnbiBsYW5ndWFnZSBmb3IgTGludXguCnZlcnNpb246IDAuMC4xCnB1Ymxpc2hfdG86IG5vbmUKCmVudmlyb25tZW50OgogIHNkazogIj49My4wLjAgPDQuMC4wIgogIGZsdXR0ZXI6ICI+PTMuMTAuMCIKCmRlcGVuZGVuY2llczoKICBmbHV0dGVyOgogICAgc2RrOiBmbHV0dGVyCiAgZmx1dHRlcl9sb2NhbGl6YXRpb25zOgogICAgc2RrOiBmbHV0dGVyCiAgdGhlbWVfdGFpbG9yX2Fubm90YXRpb246IF4yLjAuMAogIHlhcnU6IF4wLjguMAogIHlhcnVfaWNvbnM6IF4xLjAuNAogIHt7cHJvamVjdF9uYW1lfX1fdWk6CgpkZXZfZGVwZW5kZW5jaWVzOgogIGZsdXR0ZXJfdGVzdDoKICAgIHNkazogZmx1dHRlcgogIGFsY2hlbWlzdDogXjAuNi4xCiAgYnVpbGRfcnVubmVyOiBeMi40LjUKICB0aGVtZV90YWlsb3I6IF4yLjAuMAogIGZsdXR0ZXJfZ2VuX3J1bm5lcjogXjUuMy4xCiAgZmx1dHRlcl9saW50czogXjIuMC4xCgojIFRoZSBmb2xsb3dpbmcgc2VjdGlvbiBpcyBzcGVjaWZpYyB0byB0aGUgRmx1dHRlckdlbiBwYWNrYWdlLgojIFNlZTogaHR0cHM6Ly9wdWIuZGV2L3BhY2thZ2VzL2ZsdXR0ZXJfZ2VuI2NvbmZpZ3VyYXRpb24tZmlsZQpmbHV0dGVyX2dlbjoKICBvdXRwdXQ6IGxpYi9zcmMvCiAgIyBUbyBhZGQgc3VwcG9ydCBmb3Igc3ZnIHZpYSBmbHV0dGVyX3N2ZyBwYWNrYWdlCiAgIyBpbnRlZ3JhdGlvbnM6CiAgIyAgIGZsdXR0ZXJfc3ZnOiB0cnVlCgogIGFzc2V0czoKICAgIG91dHB1dHM6CiAgICAgIHBhY2thZ2VfcGFyYW1ldGVyX2VuYWJsZWQ6IHRydWUKCiMgVGhlIGZvbGxvd2luZyBzZWN0aW9uIGlzIHNwZWNpZmljIHRvIEZsdXR0ZXIgcGFja2FnZXMuCmZsdXR0ZXI6CiAgIyBUaGUgZm9sbG93aW5nIGxpbmUgZW5zdXJlcyB0aGF0IHRoZSBNYXRlcmlhbCBJY29ucyBmb250IGlzCiAgIyBpbmNsdWRlZCB3aXRoIHlvdXIgYXBwbGljYXRpb24sIHNvIHRoYXQgeW91IGNhbiB1c2UgdGhlIGljb25zIGluCiAgIyB0aGUgbWF0ZXJpYWwgSWNvbnMgY2xhc3MuCiAgdXNlcy1tYXRlcmlhbC1kZXNpZ246IHRydWUKICAjIFRvIGFkZCBhc3NldHMgdG8geW91ciBwYWNrYWdlLCBhZGQgYW4gYXNzZXRzIHNlY3Rpb24sIGxpa2UgdGhpczoKICAjIGFzc2V0czoKICAjICAgLSBpbWFnZXMvYV9kb3RfYnVyci5qcGVnCiAgIyAgIC0gaW1hZ2VzL2FfZG90X2hhbS5qcGVnCiAgIwogICMgRm9yIGRldGFpbHMgcmVnYXJkaW5nIGFzc2V0cyBpbiBwYWNrYWdlcywgc2VlCiAgIyBodHRwczovL2ZsdXR0ZXIuZGV2L2Fzc2V0cy1hbmQtaW1hZ2VzLyNmcm9tLXBhY2thZ2VzCiAgIwogICMgQW4gaW1hZ2UgYXNzZXQgY2FuIHJlZmVyIHRvIG9uZSBvciBtb3JlIHJlc29sdXRpb24tc3BlY2lmaWMgInZhcmlhbnRzIiwgc2VlCiAgIyBodHRwczovL2ZsdXR0ZXIuZGV2L2Fzc2V0cy1hbmQtaW1hZ2VzLyNyZXNvbHV0aW9uLWF3YXJlCgogICMgVG8gYWRkIGN1c3RvbSBmb250cyB0byB5b3VyIHBhY2thZ2UsIGFkZCBhIGZvbnRzIHNlY3Rpb24gaGVyZSwKICAjIGluIHRoaXMgImZsdXR0ZXIiIHNlY3Rpb24uIEVhY2ggZW50cnkgaW4gdGhpcyBsaXN0IHNob3VsZCBoYXZlIGEKICAjICJmYW1pbHkiIGtleSB3aXRoIHRoZSBmb250IGZhbWlseSBuYW1lLCBhbmQgYSAiZm9udHMiIGtleSB3aXRoIGEKICAjIGxpc3QgZ2l2aW5nIHRoZSBhc3NldCBhbmQgb3RoZXIgZGVzY3JpcHRvcnMgZm9yIHRoZSBmb250LiBGb3IKICAjIGV4YW1wbGU6CiAgIyBmb250czoKICAjICAgLSBmYW1pbHk6IFNjaHlsZXIKICAjICAgICBmb250czoKICAjICAgICAgIC0gYXNzZXQ6IGZvbnRzL1NjaHlsZXItUmVndWxhci50dGYKICAjICAgICAgIC0gYXNzZXQ6IGZvbnRzL1NjaHlsZXItSXRhbGljLnR0ZgogICMgICAgICAgICBzdHlsZTogaXRhbGljCiAgIyAgIC0gZmFtaWx5OiBUcmFqYW4gUHJvCiAgIyAgICAgZm9udHM6CiAgIyAgICAgICAtIGFzc2V0OiBmb250cy9UcmFqYW5Qcm8udHRmCiAgIyAgICAgICAtIGFzc2V0OiBmb250cy9UcmFqYW5Qcm9fQm9sZC50dGYKICAjICAgICAgICAgd2VpZ2h0OiA3MDAKICAjCiAgIyBGb3IgZGV0YWlscyByZWdhcmRpbmcgZm9udHMgaW4gcGFja2FnZXMsIHNlZQogICMgaHR0cHM6Ly9mbHV0dGVyLmRldi9jdXN0b20tZm9udHMvI2Zyb20tcGFja2FnZXMKe3svbGludXh9fXt7I21hY29zfX1uYW1lOiB7e3Byb2plY3RfbmFtZX19X3VpX21hY29zCmRlc2NyaXB0aW9uOiBXaWRnZXRzIGFuZCB0aGVtZXMgaW1wbGVtZW50aW5nIHt7cHJvamVjdF9uYW1lLnRpdGxlQ2FzZSgpfX0gZGVzaWduIGxhbmd1YWdlIGZvciBtYWNPUy4KdmVyc2lvbjogMC4wLjEKcHVibGlzaF90bzogbm9uZQoKZW52aXJvbm1lbnQ6CiAgc2RrOiAiPj0zLjAuMCA8NC4wLjAiCiAgZmx1dHRlcjogIj49My4xMC4wIgoKZGVwZW5kZW5jaWVzOgogIGZsdXR0ZXI6CiAgICBzZGs6IGZsdXR0ZXIKICBmbHV0dGVyX2xvY2FsaXphdGlvbnM6CiAgICBzZGs6IGZsdXR0ZXIKICB0aGVtZV90YWlsb3JfYW5ub3RhdGlvbjogXjIuMC4wCiAgY3VwZXJ0aW5vX2ljb25zOiBeMS4wLjUKICBtYWNvc191aTogXjIuMC4wCiAge3twcm9qZWN0X25hbWV9fV91aToKCmRldl9kZXBlbmRlbmNpZXM6CiAgZmx1dHRlcl90ZXN0OgogICAgc2RrOiBmbHV0dGVyCiAgYWxjaGVtaXN0OiBeMC42LjEKICBidWlsZF9ydW5uZXI6IF4yLjQuNQogIHRoZW1lX3RhaWxvcjogXjIuMC4wCiAgZmx1dHRlcl9nZW5fcnVubmVyOiBeNS4zLjEKICBmbHV0dGVyX2xpbnRzOiBeMi4wLjEKCiMgVGhlIGZvbGxvd2luZyBzZWN0aW9uIGlzIHNwZWNpZmljIHRvIHRoZSBGbHV0dGVyR2VuIHBhY2thZ2UuCiMgU2VlOiBodHRwczovL3B1Yi5kZXYvcGFja2FnZXMvZmx1dHRlcl9nZW4jY29uZmlndXJhdGlvbi1maWxlCmZsdXR0ZXJfZ2VuOgogIG91dHB1dDogbGliL3NyYy8KICAjIFRvIGFkZCBzdXBwb3J0IGZvciBzdmcgdmlhIGZsdXR0ZXJfc3ZnIHBhY2thZ2UKICAjIGludGVncmF0aW9uczoKICAjICAgZmx1dHRlcl9zdmc6IHRydWUKCiAgYXNzZXRzOgogICAgb3V0cHV0czoKICAgICAgcGFja2FnZV9wYXJhbWV0ZXJfZW5hYmxlZDogdHJ1ZQoKIyBUaGUgZm9sbG93aW5nIHNlY3Rpb24gaXMgc3BlY2lmaWMgdG8gRmx1dHRlciBwYWNrYWdlcy4KZmx1dHRlcjoKCiAgIyBUbyBhZGQgYXNzZXRzIHRvIHlvdXIgcGFja2FnZSwgYWRkIGFuIGFzc2V0cyBzZWN0aW9uLCBsaWtlIHRoaXM6CiAgIyBhc3NldHM6CiAgIyAgIC0gaW1hZ2VzL2FfZG90X2J1cnIuanBlZwogICMgICAtIGltYWdlcy9hX2RvdF9oYW0uanBlZwogICMKICAjIEZvciBkZXRhaWxzIHJlZ2FyZGluZyBhc3NldHMgaW4gcGFja2FnZXMsIHNlZQogICMgaHR0cHM6Ly9mbHV0dGVyLmRldi9hc3NldHMtYW5kLWltYWdlcy8jZnJvbS1wYWNrYWdlcwogICMKICAjIEFuIGltYWdlIGFzc2V0IGNhbiByZWZlciB0byBvbmUgb3IgbW9yZSByZXNvbHV0aW9uLXNwZWNpZmljICJ2YXJpYW50cyIsIHNlZQogICMgaHR0cHM6Ly9mbHV0dGVyLmRldi9hc3NldHMtYW5kLWltYWdlcy8jcmVzb2x1dGlvbi1hd2FyZQoKICAjIFRvIGFkZCBjdXN0b20gZm9udHMgdG8geW91ciBwYWNrYWdlLCBhZGQgYSBmb250cyBzZWN0aW9uIGhlcmUsCiAgIyBpbiB0aGlzICJmbHV0dGVyIiBzZWN0aW9uLiBFYWNoIGVudHJ5IGluIHRoaXMgbGlzdCBzaG91bGQgaGF2ZSBhCiAgIyAiZmFtaWx5IiBrZXkgd2l0aCB0aGUgZm9udCBmYW1pbHkgbmFtZSwgYW5kIGEgImZvbnRzIiBrZXkgd2l0aCBhCiAgIyBsaXN0IGdpdmluZyB0aGUgYXNzZXQgYW5kIG90aGVyIGRlc2NyaXB0b3JzIGZvciB0aGUgZm9udC4gRm9yCiAgIyBleGFtcGxlOgogICMgZm9udHM6CiAgIyAgIC0gZmFtaWx5OiBTY2h5bGVyCiAgIyAgICAgZm9udHM6CiAgIyAgICAgICAtIGFzc2V0OiBmb250cy9TY2h5bGVyLVJlZ3VsYXIudHRmCiAgIyAgICAgICAtIGFzc2V0OiBmb250cy9TY2h5bGVyLUl0YWxpYy50dGYKICAjICAgICAgICAgc3R5bGU6IGl0YWxpYwogICMgICAtIGZhbWlseTogVHJhamFuIFBybwogICMgICAgIGZvbnRzOgogICMgICAgICAgLSBhc3NldDogZm9udHMvVHJhamFuUHJvLnR0ZgogICMgICAgICAgLSBhc3NldDogZm9udHMvVHJhamFuUHJvX0JvbGQudHRmCiAgIyAgICAgICAgIHdlaWdodDogNzAwCiAgIwogICMgRm9yIGRldGFpbHMgcmVnYXJkaW5nIGZvbnRzIGluIHBhY2thZ2VzLCBzZWUKICAjIGh0dHBzOi8vZmx1dHRlci5kZXYvY3VzdG9tLWZvbnRzLyNmcm9tLXBhY2thZ2VzCnt7L21hY29zfX17eyN3ZWJ9fW5hbWU6IHt7cHJvamVjdF9uYW1lfX1fdWlfd2ViCmRlc2NyaXB0aW9uOiBXaWRnZXRzIGFuZCB0aGVtZXMgaW1wbGVtZW50aW5nIHt7cHJvamVjdF9uYW1lLnRpdGxlQ2FzZSgpfX0gZGVzaWduIGxhbmd1YWdlIGZvciBXZWIuCnZlcnNpb246IDAuMC4xCnB1Ymxpc2hfdG86IG5vbmUKCmVudmlyb25tZW50OgogIHNkazogIj49My4wLjAgPDQuMC4wIgogIGZsdXR0ZXI6ICI+PTMuMTAuMCIKCmRlcGVuZGVuY2llczoKICBmbHV0dGVyOgogICAgc2RrOiBmbHV0dGVyCiAgZmx1dHRlcl9sb2NhbGl6YXRpb25zOgogICAgc2RrOiBmbHV0dGVyCiAgdGhlbWVfdGFpbG9yX2Fubm90YXRpb246IF4yLjAuMAogIHt7cHJvamVjdF9uYW1lfX1fdWk6CgpkZXZfZGVwZW5kZW5jaWVzOgogIGZsdXR0ZXJfdGVzdDoKICAgIHNkazogZmx1dHRlcgogIGFsY2hlbWlzdDogXjAuNi4xCiAgYnVpbGRfcnVubmVyOiBeMi40LjUKICB0aGVtZV90YWlsb3I6IF4yLjAuMAogIGZsdXR0ZXJfZ2VuX3J1bm5lcjogXjUuMy4xCiAgZmx1dHRlcl9saW50czogXjIuMC4xCgojIFRoZSBmb2xsb3dpbmcgc2VjdGlvbiBpcyBzcGVjaWZpYyB0byB0aGUgRmx1dHRlckdlbiBwYWNrYWdlLgojIFNlZTogaHR0cHM6Ly9wdWIuZGV2L3BhY2thZ2VzL2ZsdXR0ZXJfZ2VuI2NvbmZpZ3VyYXRpb24tZmlsZQpmbHV0dGVyX2dlbjoKICBvdXRwdXQ6IGxpYi9zcmMvCiAgIyBUbyBhZGQgc3VwcG9ydCBmb3Igc3ZnIHZpYSBmbHV0dGVyX3N2ZyBwYWNrYWdlCiAgIyBpbnRlZ3JhdGlvbnM6CiAgIyAgIGZsdXR0ZXJfc3ZnOiB0cnVlCgogIGFzc2V0czoKICAgIG91dHB1dHM6CiAgICAgIHBhY2thZ2VfcGFyYW1ldGVyX2VuYWJsZWQ6IHRydWUKCiMgVGhlIGZvbGxvd2luZyBzZWN0aW9uIGlzIHNwZWNpZmljIHRvIEZsdXR0ZXIgcGFja2FnZXMuCmZsdXR0ZXI6CiAgIyBUaGUgZm9sbG93aW5nIGxpbmUgZW5zdXJlcyB0aGF0IHRoZSBNYXRlcmlhbCBJY29ucyBmb250IGlzCiAgIyBpbmNsdWRlZCB3aXRoIHlvdXIgYXBwbGljYXRpb24sIHNvIHRoYXQgeW91IGNhbiB1c2UgdGhlIGljb25zIGluCiAgIyB0aGUgbWF0ZXJpYWwgSWNvbnMgY2xhc3MuCiAgdXNlcy1tYXRlcmlhbC1kZXNpZ246IHRydWUKICAjIFRvIGFkZCBhc3NldHMgdG8geW91ciBwYWNrYWdlLCBhZGQgYW4gYXNzZXRzIHNlY3Rpb24sIGxpa2UgdGhpczoKICAjIGFzc2V0czoKICAjICAgLSBpbWFnZXMvYV9kb3RfYnVyci5qcGVnCiAgIyAgIC0gaW1hZ2VzL2FfZG90X2hhbS5qcGVnCiAgIwogICMgRm9yIGRldGFpbHMgcmVnYXJkaW5nIGFzc2V0cyBpbiBwYWNrYWdlcywgc2VlCiAgIyBodHRwczovL2ZsdXR0ZXIuZGV2L2Fzc2V0cy1hbmQtaW1hZ2VzLyNmcm9tLXBhY2thZ2VzCiAgIwogICMgQW4gaW1hZ2UgYXNzZXQgY2FuIHJlZmVyIHRvIG9uZSBvciBtb3JlIHJlc29sdXRpb24tc3BlY2lmaWMgInZhcmlhbnRzIiwgc2VlCiAgIyBodHRwczovL2ZsdXR0ZXIuZGV2L2Fzc2V0cy1hbmQtaW1hZ2VzLyNyZXNvbHV0aW9uLWF3YXJlCgogICMgVG8gYWRkIGN1c3RvbSBmb250cyB0byB5b3VyIHBhY2thZ2UsIGFkZCBhIGZvbnRzIHNlY3Rpb24gaGVyZSwKICAjIGluIHRoaXMgImZsdXR0ZXIiIHNlY3Rpb24uIEVhY2ggZW50cnkgaW4gdGhpcyBsaXN0IHNob3VsZCBoYXZlIGEKICAjICJmYW1pbHkiIGtleSB3aXRoIHRoZSBmb250IGZhbWlseSBuYW1lLCBhbmQgYSAiZm9udHMiIGtleSB3aXRoIGEKICAjIGxpc3QgZ2l2aW5nIHRoZSBhc3NldCBhbmQgb3RoZXIgZGVzY3JpcHRvcnMgZm9yIHRoZSBmb250LiBGb3IKICAjIGV4YW1wbGU6CiAgIyBmb250czoKICAjICAgLSBmYW1pbHk6IFNjaHlsZXIKICAjICAgICBmb250czoKICAjICAgICAgIC0gYXNzZXQ6IGZvbnRzL1NjaHlsZXItUmVndWxhci50dGYKICAjICAgICAgIC0gYXNzZXQ6IGZvbnRzL1NjaHlsZXItSXRhbGljLnR0ZgogICMgICAgICAgICBzdHlsZTogaXRhbGljCiAgIyAgIC0gZmFtaWx5OiBUcmFqYW4gUHJvCiAgIyAgICAgZm9udHM6CiAgIyAgICAgICAtIGFzc2V0OiBmb250cy9UcmFqYW5Qcm8udHRmCiAgIyAgICAgICAtIGFzc2V0OiBmb250cy9UcmFqYW5Qcm9fQm9sZC50dGYKICAjICAgICAgICAgd2VpZ2h0OiA3MDAKICAjCiAgIyBGb3IgZGV0YWlscyByZWdhcmRpbmcgZm9udHMgaW4gcGFja2FnZXMsIHNlZQogICMgaHR0cHM6Ly9mbHV0dGVyLmRldi9jdXN0b20tZm9udHMvI2Zyb20tcGFja2FnZXMKe3svd2VifX17eyN3aW5kb3dzfX1uYW1lOiB7e3Byb2plY3RfbmFtZX19X3VpX3dpbmRvd3MKZGVzY3JpcHRpb246IFdpZGdldHMgYW5kIHRoZW1lcyBpbXBsZW1lbnRpbmcge3twcm9qZWN0X25hbWUudGl0bGVDYXNlKCl9fSBkZXNpZ24gbGFuZ3VhZ2UgZm9yIFdpbmRvd3MuCnZlcnNpb246IDAuMC4xCnB1Ymxpc2hfdG86IG5vbmUKCmVudmlyb25tZW50OgogIHNkazogIj49My4wLjAgPDQuMC4wIgogIGZsdXR0ZXI6ICI+PTMuMTAuMCIKCmRlcGVuZGVuY2llczoKICBmbHV0dGVyOgogICAgc2RrOiBmbHV0dGVyCiAgZmx1dHRlcl9sb2NhbGl6YXRpb25zOgogICAgc2RrOiBmbHV0dGVyCiAgdGhlbWVfdGFpbG9yX2Fubm90YXRpb246IF4yLjAuMAogIGZsdWVudF91aTogXjQuNi4yCiAge3twcm9qZWN0X25hbWV9fV91aToKCmRldl9kZXBlbmRlbmNpZXM6CiAgZmx1dHRlcl90ZXN0OgogICAgc2RrOiBmbHV0dGVyCiAgYWxjaGVtaXN0OiBeMC42LjEKICBidWlsZF9ydW5uZXI6IF4yLjQuNQogIHRoZW1lX3RhaWxvcjogXjIuMC4wCiAgZmx1dHRlcl9nZW5fcnVubmVyOiBeNS4zLjEKICBmbHV0dGVyX2xpbnRzOiBeMi4wLjEKCiMgVGhlIGZvbGxvd2luZyBzZWN0aW9uIGlzIHNwZWNpZmljIHRvIHRoZSBGbHV0dGVyR2VuIHBhY2thZ2UuCiMgU2VlOiBodHRwczovL3B1Yi5kZXYvcGFja2FnZXMvZmx1dHRlcl9nZW4jY29uZmlndXJhdGlvbi1maWxlCmZsdXR0ZXJfZ2VuOgogIG91dHB1dDogbGliL3NyYy8KICAjIFRvIGFkZCBzdXBwb3J0IGZvciBzdmcgdmlhIGZsdXR0ZXJfc3ZnIHBhY2thZ2UKICAjIGludGVncmF0aW9uczoKICAjICAgZmx1dHRlcl9zdmc6IHRydWUKCiAgYXNzZXRzOgogICAgb3V0cHV0czoKICAgICAgcGFja2FnZV9wYXJhbWV0ZXJfZW5hYmxlZDogdHJ1ZQoKIyBUaGUgZm9sbG93aW5nIHNlY3Rpb24gaXMgc3BlY2lmaWMgdG8gRmx1dHRlciBwYWNrYWdlcy4KZmx1dHRlcjoKICAjIFRoZSBmb2xsb3dpbmcgbGluZSBlbnN1cmVzIHRoYXQgdGhlIE1hdGVyaWFsIEljb25zIGZvbnQgaXMKICAjIGluY2x1ZGVkIHdpdGggeW91ciBhcHBsaWNhdGlvbiwgc28gdGhhdCB5b3UgY2FuIHVzZSB0aGUgaWNvbnMgaW4KICAjIHRoZSBtYXRlcmlhbCBJY29ucyBjbGFzcy4KICB1c2VzLW1hdGVyaWFsLWRlc2lnbjogdHJ1ZQogICMgVG8gYWRkIGFzc2V0cyB0byB5b3VyIHBhY2thZ2UsIGFkZCBhbiBhc3NldHMgc2VjdGlvbiwgbGlrZSB0aGlzOgogICMgYXNzZXRzOgogICMgICAtIGltYWdlcy9hX2RvdF9idXJyLmpwZWcKICAjICAgLSBpbWFnZXMvYV9kb3RfaGFtLmpwZWcKICAjCiAgIyBGb3IgZGV0YWlscyByZWdhcmRpbmcgYXNzZXRzIGluIHBhY2thZ2VzLCBzZWUKICAjIGh0dHBzOi8vZmx1dHRlci5kZXYvYXNzZXRzLWFuZC1pbWFnZXMvI2Zyb20tcGFja2FnZXMKICAjCiAgIyBBbiBpbWFnZSBhc3NldCBjYW4gcmVmZXIgdG8gb25lIG9yIG1vcmUgcmVzb2x1dGlvbi1zcGVjaWZpYyAidmFyaWFudHMiLCBzZWUKICAjIGh0dHBzOi8vZmx1dHRlci5kZXYvYXNzZXRzLWFuZC1pbWFnZXMvI3Jlc29sdXRpb24tYXdhcmUKCiAgIyBUbyBhZGQgY3VzdG9tIGZvbnRzIHRvIHlvdXIgcGFja2FnZSwgYWRkIGEgZm9udHMgc2VjdGlvbiBoZXJlLAogICMgaW4gdGhpcyAiZmx1dHRlciIgc2VjdGlvbi4gRWFjaCBlbnRyeSBpbiB0aGlzIGxpc3Qgc2hvdWxkIGhhdmUgYQogICMgImZhbWlseSIga2V5IHdpdGggdGhlIGZvbnQgZmFtaWx5IG5hbWUsIGFuZCBhICJmb250cyIga2V5IHdpdGggYQogICMgbGlzdCBnaXZpbmcgdGhlIGFzc2V0IGFuZCBvdGhlciBkZXNjcmlwdG9ycyBmb3IgdGhlIGZvbnQuIEZvcgogICMgZXhhbXBsZToKICAjIGZvbnRzOgogICMgICAtIGZhbWlseTogU2NoeWxlcgogICMgICAgIGZvbnRzOgogICMgICAgICAgLSBhc3NldDogZm9udHMvU2NoeWxlci1SZWd1bGFyLnR0ZgogICMgICAgICAgLSBhc3NldDogZm9udHMvU2NoeWxlci1JdGFsaWMudHRmCiAgIyAgICAgICAgIHN0eWxlOiBpdGFsaWMKICAjICAgLSBmYW1pbHk6IFRyYWphbiBQcm8KICAjICAgICBmb250czoKICAjICAgICAgIC0gYXNzZXQ6IGZvbnRzL1RyYWphblByby50dGYKICAjICAgICAgIC0gYXNzZXQ6IGZvbnRzL1RyYWphblByb19Cb2xkLnR0ZgogICMgICAgICAgICB3ZWlnaHQ6IDcwMAogICMKICAjIEZvciBkZXRhaWxzIHJlZ2FyZGluZyBmb250cyBpbiBwYWNrYWdlcywgc2VlCiAgIyBodHRwczovL2ZsdXR0ZXIuZGV2L2N1c3RvbS1mb250cy8jZnJvbS1wYWNrYWdlcwp7ey93aW5kb3dzfX17eyNtb2JpbGV9fW5hbWU6IHt7cHJvamVjdF9uYW1lfX1fdWlfbW9iaWxlCmRlc2NyaXB0aW9uOiBXaWRnZXRzIGFuZCB0aGVtZXMgaW1wbGVtZW50aW5nIHt7cHJvamVjdF9uYW1lLnRpdGxlQ2FzZSgpfX0gZGVzaWduIGxhbmd1YWdlIGZvciBNb2JpbGUuCnZlcnNpb246IDAuMC4xCnB1Ymxpc2hfdG86IG5vbmUKCmVudmlyb25tZW50OgogIHNkazogIj49My4wLjAgPDQuMC4wIgogIGZsdXR0ZXI6ICI+PTMuMTAuMCIKCmRlcGVuZGVuY2llczoKICBmbHV0dGVyOgogICAgc2RrOiBmbHV0dGVyCiAgZmx1dHRlcl9sb2NhbGl6YXRpb25zOgogICAgc2RrOiBmbHV0dGVyCiAgdGhlbWVfdGFpbG9yX2Fubm90YXRpb246IF4yLjAuMAogIHt7cHJvamVjdF9uYW1lfX1fdWk6CgpkZXZfZGVwZW5kZW5jaWVzOgogIGZsdXR0ZXJfdGVzdDoKICAgIHNkazogZmx1dHRlcgogIGFsY2hlbWlzdDogXjAuNi4xCiAgYnVpbGRfcnVubmVyOiBeMi40LjUKICB0aGVtZV90YWlsb3I6IF4yLjAuMAogIGZsdXR0ZXJfZ2VuX3J1bm5lcjogXjUuMy4xCiAgZmx1dHRlcl9saW50czogXjIuMC4xCgojIFRoZSBmb2xsb3dpbmcgc2VjdGlvbiBpcyBzcGVjaWZpYyB0byB0aGUgRmx1dHRlckdlbiBwYWNrYWdlLgojIFNlZTogaHR0cHM6Ly9wdWIuZGV2L3BhY2thZ2VzL2ZsdXR0ZXJfZ2VuI2NvbmZpZ3VyYXRpb24tZmlsZQpmbHV0dGVyX2dlbjoKICBvdXRwdXQ6IGxpYi9zcmMvCiAgIyBUbyBhZGQgc3VwcG9ydCBmb3Igc3ZnIHZpYSBmbHV0dGVyX3N2ZyBwYWNrYWdlCiAgIyBpbnRlZ3JhdGlvbnM6CiAgIyAgIGZsdXR0ZXJfc3ZnOiB0cnVlCgogIGFzc2V0czoKICAgIG91dHB1dHM6CiAgICAgIHBhY2thZ2VfcGFyYW1ldGVyX2VuYWJsZWQ6IHRydWUKCiMgVGhlIGZvbGxvd2luZyBzZWN0aW9uIGlzIHNwZWNpZmljIHRvIEZsdXR0ZXIgcGFja2FnZXMuCmZsdXR0ZXI6CiAgIyBUaGUgZm9sbG93aW5nIGxpbmUgZW5zdXJlcyB0aGF0IHRoZSBNYXRlcmlhbCBJY29ucyBmb250IGlzCiAgIyBpbmNsdWRlZCB3aXRoIHlvdXIgYXBwbGljYXRpb24sIHNvIHRoYXQgeW91IGNhbiB1c2UgdGhlIGljb25zIGluCiAgIyB0aGUgbWF0ZXJpYWwgSWNvbnMgY2xhc3MuCiAgdXNlcy1tYXRlcmlhbC1kZXNpZ246IHRydWUKICAjIFRvIGFkZCBhc3NldHMgdG8geW91ciBwYWNrYWdlLCBhZGQgYW4gYXNzZXRzIHNlY3Rpb24sIGxpa2UgdGhpczoKICAjIGFzc2V0czoKICAjICAgLSBpbWFnZXMvYV9kb3RfYnVyci5qcGVnCiAgIyAgIC0gaW1hZ2VzL2FfZG90X2hhbS5qcGVnCiAgIwogICMgRm9yIGRldGFpbHMgcmVnYXJkaW5nIGFzc2V0cyBpbiBwYWNrYWdlcywgc2VlCiAgIyBodHRwczovL2ZsdXR0ZXIuZGV2L2Fzc2V0cy1hbmQtaW1hZ2VzLyNmcm9tLXBhY2thZ2VzCiAgIwogICMgQW4gaW1hZ2UgYXNzZXQgY2FuIHJlZmVyIHRvIG9uZSBvciBtb3JlIHJlc29sdXRpb24tc3BlY2lmaWMgInZhcmlhbnRzIiwgc2VlCiAgIyBodHRwczovL2ZsdXR0ZXIuZGV2L2Fzc2V0cy1hbmQtaW1hZ2VzLyNyZXNvbHV0aW9uLWF3YXJlCgogICMgVG8gYWRkIGN1c3RvbSBmb250cyB0byB5b3VyIHBhY2thZ2UsIGFkZCBhIGZvbnRzIHNlY3Rpb24gaGVyZSwKICAjIGluIHRoaXMgImZsdXR0ZXIiIHNlY3Rpb24uIEVhY2ggZW50cnkgaW4gdGhpcyBsaXN0IHNob3VsZCBoYXZlIGEKICAjICJmYW1pbHkiIGtleSB3aXRoIHRoZSBmb250IGZhbWlseSBuYW1lLCBhbmQgYSAiZm9udHMiIGtleSB3aXRoIGEKICAjIGxpc3QgZ2l2aW5nIHRoZSBhc3NldCBhbmQgb3RoZXIgZGVzY3JpcHRvcnMgZm9yIHRoZSBmb250LiBGb3IKICAjIGV4YW1wbGU6CiAgIyBmb250czoKICAjICAgLSBmYW1pbHk6IFNjaHlsZXIKICAjICAgICBmb250czoKICAjICAgICAgIC0gYXNzZXQ6IGZvbnRzL1NjaHlsZXItUmVndWxhci50dGYKICAjICAgICAgIC0gYXNzZXQ6IGZvbnRzL1NjaHlsZXItSXRhbGljLnR0ZgogICMgICAgICAgICBzdHlsZTogaXRhbGljCiAgIyAgIC0gZmFtaWx5OiBUcmFqYW4gUHJvCiAgIyAgICAgZm9udHM6CiAgIyAgICAgICAtIGFzc2V0OiBmb250cy9UcmFqYW5Qcm8udHRmCiAgIyAgICAgICAtIGFzc2V0OiBmb250cy9UcmFqYW5Qcm9fQm9sZC50dGYKICAjICAgICAgICAgd2VpZ2h0OiA3MDAKICAjCiAgIyBGb3IgZGV0YWlscyByZWdhcmRpbmcgZm9udHMgaW4gcGFja2FnZXMsIHNlZQogICMgaHR0cHM6Ly9mbHV0dGVyLmRldi9jdXN0b20tZm9udHMvI2Zyb20tcGFja2FnZXMKe3svbW9iaWxlfX0=",
       "type": "text"
     },
     {
