+++ packages/rapid_cli/templates/platform_ui_package/__brick__/pubspec.yaml
@@ -236,10 +236,7 @@ dependencies:
     sdk: flutter
   theme_tailor_annotation: ^2.0.0
   cupertino_icons: ^1.0.5
-  macos_ui:
-    git:
-      url: https://github.com/macosui/macos_ui.git
-      ref: f1053fc31958dbaa362a0a4c8288e8cf0919ab8c
+  macos_ui: ^2.0.0
   {{project_name}}_ui:
 
 dev_dependencies:
