+++ packages/rapid_cli/test/src/commands/activate_test.dart
@@ -1,6 +1,5 @@
 import 'package:mocktail/mocktail.dart';
 import 'package:rapid_cli/src/commands/runner.dart';
-import 'package:rapid_cli/src/io.dart';
 import 'package:rapid_cli/src/project/language.dart';
 import 'package:rapid_cli/src/project/platform.dart';
 import 'package:rapid_cli/src/project/project.dart';
@@ -8,7 +7,6 @@ import 'package:test/test.dart';
 
 import '../invocations.dart';
 import '../mock_env.dart';
-import '../mock_fs.dart';
 import '../mocks.dart';
 import '../utils.dart';
 
@@ -462,119 +460,6 @@ void main() {
           ),
         ]);
       },
-    );
-
-    test(
-      'regression for issue 96',
-      withMockFs(() async {
-        final manager = MockProcessManager();
-        final podFile = File('Podfile')
-          ..writeAsStringSync('platform :osx, \'10.14\'');
-        final nativeDirectory = MockMacosNativeDirectory(podFile: podFile);
-        final platformRootPackage =
-            MockMacosRootPackage(nativeDirectory: nativeDirectory);
-
-        final project = MockRapidProject(
-          appModule: MockAppModule(
-            platformDirectory: ({required platform}) => MockPlatformDirectory(
-              rootPackage: platformRootPackage,
-            ),
-          ),
-        );
-        final logger = MockRapidLogger();
-        final rapid = getRapid(
-          project: project,
-          logger: logger,
-        );
-
-        await withMockProcessManager(
-          () async => rapid.activateMacos(
-            orgName: 'test.example',
-            language: Language(languageCode: 'fr'),
-          ),
-          manager: manager,
-        );
-
-        expect(podFile.readAsStringSync(), 'platform :osx, \'10.15.7.7\'');
-      }),
-    );
-
-    test(
-      'regression for issue 96 (not existing podfile)',
-      withMockFs(() async {
-        final manager = MockProcessManager();
-        final podFile = File('Podfile');
-        final nativeDirectory = MockMacosNativeDirectory(podFile: podFile);
-        final platformRootPackage =
-            MockMacosRootPackage(nativeDirectory: nativeDirectory);
-
-        final project = MockRapidProject(
-          appModule: MockAppModule(
-            platformDirectory: ({required platform}) => MockPlatformDirectory(
-              rootPackage: platformRootPackage,
-            ),
-          ),
-        );
-        final logger = MockRapidLogger();
-        final rapid = getRapid(
-          project: project,
-          logger: logger,
-        );
-
-        await withMockProcessManager(
-          () async => rapid.activateMacos(
-            orgName: 'test.example',
-            language: Language(languageCode: 'fr'),
-          ),
-          manager: manager,
-        );
-
-        expect(podFile.readAsStringSync(), '''
-platform :osx, '10.15.7.7'
-
-# CocoaPods analytics sends network stats synchronously affecting flutter build latency.
-ENV['COCOAPODS_DISABLE_STATS'] = 'true'
-
-project 'Runner', {
-  'Debug' => :debug,
-  'Profile' => :release,
-  'Release' => :release,
-}
-
-def flutter_root
-  generated_xcode_build_settings_path = File.expand_path(File.join('..', 'Flutter', 'ephemeral', 'Flutter-Generated.xcconfig'), __FILE__)
-  unless File.exist?(generated_xcode_build_settings_path)
-    raise "#{generated_xcode_build_settings_path} must exist. If you're running pod install manually, make sure \\"flutter pub get\\" is executed first"
-  end
-
-  File.foreach(generated_xcode_build_settings_path) do |line|
-    matches = line.match(/FLUTTER_ROOT\\=(.*)/)
-    return matches[1].strip if matches
-  end
-  raise "FLUTTER_ROOT not found in #{generated_xcode_build_settings_path}. Try deleting Flutter-Generated.xcconfig, then run \\"flutter pub get\\""
-end
-
-require File.expand_path(File.join('packages', 'flutter_tools', 'bin', 'podhelper'), flutter_root)
-
-flutter_macos_podfile_setup
-
-target 'Runner' do
-  use_frameworks!
-  use_modular_headers!
-
-  flutter_install_all_macos_pods File.dirname(File.realpath(__FILE__))
-  target 'RunnerTests' do
-    inherit! :search_paths
-  end
-end
-
-post_install do |installer|
-  installer.pods_project.targets.each do |target|
-    flutter_additional_macos_build_settings(target)
-  end
-end
-''');
-      }),
     );
   });
