+++ packages/rapid_cli/tool/bundles/lib/src/update_platform_native_directory_templates.dart
@@ -141,9 +141,17 @@ void _updatePlatformNativeDirectoryTemplate({
         .where((e) => !e.isBinary)) {
       final content = file.readAsStringSync();
       file.writeAsStringSync(content
-          .replaceAll('10.14', '10.15.7')
-          .replaceAll('10.14.6', '10.15.7'));
+          .replaceAll('10.14.6', '10.14')
+          .replaceAll('10.14', '10.14.6'));
     }
+
+    final podFile = File(p.join(brickDir.path, 'Podfile'));
+    if (podFile.existsSync()) {
+      podFile.deleteSync(recursive: true);
+    }
+    podFile
+      ..createSync(recursive: true)
+      ..writeAsStringSync(_podFileContent);
   }
 }
 
@@ -153,3 +161,49 @@ extension on File {
     return extension == '.ico' || extension == '.png';
   }
 }
+
+const _podFileContent = '''
+platform :osx, '10.14.6'
+
+# CocoaPods analytics sends network stats synchronously affecting flutter build latency.
+ENV['COCOAPODS_DISABLE_STATS'] = 'true'
+
+project 'Runner', {
+  'Debug' => :debug,
+  'Profile' => :release,
+  'Release' => :release,
+}
+
+def flutter_root
+  generated_xcode_build_settings_path = File.expand_path(File.join('..', 'Flutter', 'ephemeral', 'Flutter-Generated.xcconfig'), __FILE__)
+  unless File.exist?(generated_xcode_build_settings_path)
+    raise "#{generated_xcode_build_settings_path} must exist. If you're running pod install manually, make sure \\"flutter pub get\\" is executed first"
+  end
+
+  File.foreach(generated_xcode_build_settings_path) do |line|
+    matches = line.match(/FLUTTER_ROOT\\=(.*)/)
+    return matches[1].strip if matches
+  end
+  raise "FLUTTER_ROOT not found in #{generated_xcode_build_settings_path}. Try deleting Flutter-Generated.xcconfig, then run \\"flutter pub get\\""
+end
+
+require File.expand_path(File.join('packages', 'flutter_tools', 'bin', 'podhelper'), flutter_root)
+
+flutter_macos_podfile_setup
+
+target 'Runner' do
+  use_frameworks!
+  use_modular_headers!
+
+  flutter_install_all_macos_pods File.dirname(File.realpath(__FILE__))
+  target 'RunnerTests' do
+    inherit! :search_paths
+  end
+end
+
+post_install do |installer|
+  installer.pods_project.targets.each do |target|
+    flutter_additional_macos_build_settings(target)
+  end
+end
+''';
